Finite fields

Gapjm.FFields — Module

This module introduces modular arithmetic and finite fields.

The integer x mod. n is constructed by the function Mod(x,n). If n isa Int its type is Mod{UInt64}. If n isa BigInt its type is Mod{BigInt}. Since n is not encoded in the type, the elements 0 and 1 mod. n cannot be constructed from the type, which causes some problems for some Julia functions. For some prime moduli p, the type FFE{p} below does not have such limitations.

Example:

julia> Mod(5,19)
Mod{UInt64}: 5₁₉

julia> a=Mod(5,19)
Mod{UInt64}: 5₁₉

julia> a^2
Mod{UInt64}: 6₁₉

julia> inv(a)
Mod{UInt64}: 4₁₉

julia> a*inv(a)
Mod{UInt64}: 1₁₉

julia> a+2
Mod{UInt64}: 7₁₉

julia> a*2
Mod{UInt64}: -9₁₉

julia> a+1//2
Mod{UInt64}: -4₁₉

julia> Integer(a) # get back an integer from a
5

julia> order(a) # multiplicative order of a
9

The Galois field with p^n elements is obtained as GF(p^n). To work with elements of this field, (as in GAP) the function Z(p^n) returns a generator of the multiplicative group of GF(p^n). Other elements of GF(p^n) are obtained as a power of Z(p^n) or as 0*Z(p^n).

julia> a=Z(64)
FFE{2}: Z₆₄

julia> a^9
FFE{2}: Z₈

julia> a^21
FFE{2}: Z₄

julia> a+1
FFE{2}: Z₆₄⁵⁶

Elements of the prime field can be converted back to integers Mod(,p) or to integers:

julia> a=Z(19)+3
FFE{19}: 5

julia> Mod(a)
Mod{UInt64}: 5₁₉

julia> Int(a)
5

julia> order(a) # order as element of the multiplicative group
9

The field, p, n and p^n can be obtained back as well as which power of Z(p^n) is considered

julia> a=Z(8)^5
FFE{2}: Z₈⁵

julia> F=field(a)
GF(2^3)

julia> char(F)
2

julia> char(a)
2

julia> degree(F)
3

julia> degree(a)
3

julia> length(F)
8

julia> log(a)
5

julia> elements(F)
8-element Vector{FFE{2}}:
   0
   1
  Z₈
 Z₈²
 Z₈³
 Z₈⁴
 Z₈⁵
 Z₈⁶

The type of an element of GF(p^n) is FFE{p}. A p-integral number or a Mod(,p) can be converted to the prime field using this type as constructor.

julia> FFE{19}(2)
FFE{19}: 2

julia> FFE{19}(Mod(2,19))
FFE{19}: 2

Gapjm.FFields.FFE — Type

FFE{p} is the type of the elements of a finite field of characteristic p.

Gapjm.FFields.FFE — Method

FFE{p}(z::Cyc) where z is a p-integral cyclotomic number (that is, z times some number prime to p is a cyclotomic integer), returns the reduction of z mod. p, an element of some extension 𝔽_{pʳ} of the prime field 𝔽ₚ.

julia> FFE{3}(E(7))
Z₇₂₉¹⁰⁴

FFE{p}(i) for i an integer or a fraction with denominator prime to p returns the reduction mod p of i, an element of the prime field 𝔽ₚ.

Gapjm.FFields.Z — Method

Z(p^d) or Z(p,d)

This returns a generator of the multiplicative group of the finite field 𝔽_{pᵈ}, where p must be prime and p^d smaller than 2¹⁵. This multiplicative group is cyclic thus Z(p^d)^a runs over it for a in 0:p^d-1. The zero of the field is 0*Z(p) (the same as 0*Z(p^d); we automatically lower an element to the smallest field which contains it).

The various generators returned by Z for finite fields of characteristic p are compatible. That is, if the field 𝔽_{pⁿ} is a subfield of the field 𝔽_{pᵐ}, that is, n divides m, then Z(p^n)=Z(p^m)^div(p^m-1,p^n-1). This is achieved by choosing Z(p) as the smallest primitive root modulo p and Z(p^n) as a root of the n-th Conway polynomial of characteristic p. Those polynomials where defined by J.H.~Conway and computed by R.A.~Parker.

julia> z=Z(16)
FFE{2}: Z₁₆

julia> z^5
FFE{2}: Z₄