Linear algebra on any field/ring

Gapjm.GLinearAlgebra

Gapjm.Chars.schur_functor

Gapjm.GLinearAlgebra.bigcell_decomposition

Gapjm.GLinearAlgebra.charpoly

Gapjm.GLinearAlgebra.comatrix

Gapjm.GLinearAlgebra.diagconj_elt

Gapjm.GLinearAlgebra.echelon!

Gapjm.GLinearAlgebra.exterior_power

Gapjm.GLinearAlgebra.in_rowspace

Gapjm.GLinearAlgebra.independent_rows

Gapjm.GLinearAlgebra.intersect_rowspace

Gapjm.GLinearAlgebra.lnullspace

Gapjm.GLinearAlgebra.nullspace

Gapjm.GLinearAlgebra.permanent

Gapjm.GLinearAlgebra.rank

Gapjm.GLinearAlgebra.ratio

Gapjm.GLinearAlgebra.rowspace

Gapjm.GLinearAlgebra.solutionmat

Gapjm.GLinearAlgebra.symmetric_power

Gapjm.GLinearAlgebra.transporter

Gapjm.Tools2.traces_words_mats

Gapjm.GLinearAlgebra — Module

GLinearAlgebra: linear algebra over arbitrary fields and rings

The linear algebra package in Julia is not suitable for a general mathematics package: it assumes the field is the Real or Complex numbers and uses floating point to do approximate computations. Here we are interested in functions which work over any field (or sometimes any ring) and assume exact computations.

For more information, look at the helpstrings of echelon!, exterior_power, comatrix, bigcell_decomposition, ratio, charpoly, solutionmat, transporter, permanent, symmetric_power, diagconj_elt, lnullspace, intersect_rowspace, in_rowspace, rowspace, independent_rows.

Gapjm.GLinearAlgebra.echelon! — Function

echelon!(m::AbstractMatrix)

puts m in echelon form in-place and returns:

(m, indices of linearly independent rows of original m)

The echelon form transforms the rows of m into a particular basis of the rowspace. The first non-zero element of each line is 1, and such an element is also the only non-zero in its column. works in any field.

Gapjm.GLinearAlgebra.bigcell_decomposition — Function

bigcell_decomposition(M [, b])

M should be a square matrix, and b specifies a block structure for a matrix of same size as M (it is a Vector of Vectors whose concatenation is 1:size(M,1)). If b is not given, the trivial block structure [[i] for i in axes(M,1)] is assumed.

The function decomposes M as a product P₁ L P where P is upper block-unitriangular (with identity diagonal blocks), P₁ is lower block-unitriangular and L is block-diagonal for the block structure b. If M is symmetric then P₁ is the transposed of P and the result is the pair [P,L]; else the result is the triple [P₁,L,P]. The only condition for this decomposition of M to be possible is that the principal minors according to the block structure be invertible. This routine is used in the Lusztig-Shoji algorithm for computing the Green functions and the example below is extracted from the computation of the Green functions for G₂.

julia> @Pol q
Pol{Int64}: q

julia> M=[q^6 q^0 q^3 q^3 q^5+q q^4+q^2; q^0 q^6 q^3 q^3 q^5+q q^4+q^2; q^3 q^3 q^6 q^0 q^4+q^2 q^5+q; q^3 q^3 q^0 q^6 q^4+q^2 q^5+q; q^5+q q^5+q q^4+q^2 q^4+q^2 q^6+q^4+q^2+1 q^5+2*q^3+q; q^4+q^2 q^4+q^2 q^5+q q^5+q q^5+2*q^3+q q^6+q^4+q^2+1]
6×6 Matrix{Pol{Int64}}:
 q⁶     1      q³     q³     q⁵+q        q⁴+q²
 1      q⁶     q³     q³     q⁵+q        q⁴+q²
 q³     q³     q⁶     1      q⁴+q²       q⁵+q
 q³     q³     1      q⁶     q⁴+q²       q⁵+q
 q⁵+q   q⁵+q   q⁴+q²  q⁴+q²  q⁶+q⁴+q²+1  q⁵+2q³+q
 q⁴+q²  q⁴+q²  q⁵+q   q⁵+q   q⁵+2q³+q    q⁶+q⁴+q²+1

julia> bb=[[2],[4],[6],[3,5],[1]];

julia> (P,L)=bigcell_decomposition(M,bb);

julia> P
6×6 Matrix{Pol{Int64}}:
 1    0  0    0    0        0
 q⁻⁶  1  q⁻³  q⁻³  q⁻¹+q⁻⁵  q⁻²+q⁻⁴
 0    0  1    0    0        0
 q⁻³  0  0    1    q⁻²      q⁻¹
 q⁻¹  0  0    0    1        0
 q⁻²  0  q⁻¹  0    q⁻¹      1

julia> L
6×6 Matrix{Pol{Int64}}:
 q⁶-q⁴-1+q⁻²  0   0            0     0            0
 0            q⁶  0            0     0            0
 0            0   q⁶-q⁴-1+q⁻²  0     0            0
 0            0   0            q⁶-1  0            0
 0            0   0            0     q⁶-q⁴-1+q⁻²  0
 0            0   0            0     0            q⁶-1

julia> M==transpose(P)*L*P
true

Gapjm.GLinearAlgebra.ratio — Function

ratio(v::abstractVector,w::abstractVector)

ratio v/w, nothing if v is not a multiple of w.

Gapjm.GLinearAlgebra.exterior_power — Function

exterior_power(mat,n)

mat should be a square matrix. The function returns the n-th exterior power of mat, in the basis naturally indexed bycombinations(1:r,n) wherer=size(mat,1)

julia> M=[1 2 3 4;2 3 4 1;3 4 1 2;4 1 2 3]
4×4 Matrix{Int64}:
 1  2  3  4
 2  3  4  1
 3  4  1  2
 4  1  2  3

julia> exterior_power(M,2)
6×6 Matrix{Int64}:
  -1   -2   -7   -1  -10  -13
  -2   -8  -10  -10  -12    2
  -7  -10  -13    1    2    1
  -1  -10    1  -13    2    7
 -10  -12    2    2    8   10
 -13    2    1    7   10   -1

Gapjm.GLinearAlgebra.charpoly — Function

charpoly(M::Matrix)

The characteristic polynomial of M (as a Vector of coefficients). This function works over any ring.

Gapjm.GLinearAlgebra.comatrix — Function

comatrix(M::Matrix)

is defined by comatrix(M)*M=det(M)*one(M). This function works over any ring.

Gapjm.GLinearAlgebra.permanent — Function

permanent(m)

returns the permanent of the square matrix m, which is defined by $\sum_{p\in\frak S_n}\prod_{i=1}^n m[i,p(i)]$.

Note the similarity of the definition of the permanent to the definition of the determinant. In fact the only difference is the missing sign of the permutation. However the permanent is quite unlike the determinant, for example it is not multilinear or alternating. It has however important combinatorical properties.

julia> permanent([0 1 1 1;1 0 1 1;1 1 0 1;1 1 1 0])
9 # inefficient way to compute the number of derangements of 1:4

julia> permanent([1 1 0 1 0 0 0; 0 1 1 0 1 0 0;0 0 1 1 0 1 0; 0 0 0 1 1 0 1;1 0 0 0 1 1 0;0 1 0 0 0 1 1;1 0 1 0 0 0 1])
24 # 24 permutations fit the projective plane of order 2 

Gapjm.GLinearAlgebra.symmetric_power — Function

symmetric_power(m,n)

returns the n-th symmetric power of the square matrix m, in the basis naturally indexed by the multisets of 1:n, where n=size(m,1).

julia> m=[1 2;3 4]
2×2 Matrix{Int64}:
 1  2
 3  4

julia> Int.(symmetric_power(m,2))
3×3 Matrix{Int64}:
 1   2   4
 6  10  16
 9  12  16

Gapjm.Chars.schur_functor — Function

schur_functor(mat,l)

mat should be a square matrix and l a partition. The result is the Schur functor of the matrix mat corresponding to partition l; for example, if l==[n] it returns the n-th symmetric power and if l==[1,1,1] it returns the 3rd exterior power. The current algorithm (from Littlewood) is rather inefficient so it is quite slow for partitions of n where n>6.

julia> m=cartan(:A,3)
3×3 Matrix{Int64}:
  2  -1   0
 -1   2  -1
  0  -1   2

julia> schur_functor(m,[2,2])
6×6 Matrix{Rational{Int64}}:
   9//1   -6//1    4//1   3//2   -2//1    1//1
 -12//1   16//1  -16//1  -4//1    8//1   -4//1
   4//1   -8//1   16//1   2//1   -8//1    4//1
  12//1  -16//1   16//1  10//1  -16//1   12//1
  -4//1    8//1  -16//1  -4//1   16//1  -12//1
   1//1   -2//1    4//1   3//2   -6//1    9//1

julia-repl

Gapjm.GLinearAlgebra.transporter — Function

transporter(l1, l2 )

l1 and l2 should be vectors of the same length of square matrices all of the same size. The result is a basis of the vector space of matrices A such that for any i we have A*l1[i]=l2[i]*A –- the basis is returned as a vector of matrices, empty if the vector space is 0. This is useful to find whether two representations are isomorphic.

Gapjm.GLinearAlgebra.diagconj_elt — Function

diagconj_elt(M,N)

M and N must be square matrices of the same size. This function returns a list d such that N==inv(Diagonal(d))*M*Diagonal(d) if such a list exists, and nothing otherwise.

julia> M=[1 2;2 1];N=[1 4;1 1]
2×2 Matrix{Int64}:
 1  4
 1  1

julia> diagconj_elt(M,N)
2-element Vector{Rational{Int64}}:
 1//1
 2//1

Gapjm.Tools2.traces_words_mats — Function

traces_words_mats(mats,words)

given a list mats of matrices and a list words of words returns the list of traces of the corresponding products of the matrices

julia> W=coxgroup(:F,4)
F₄

julia> R=representation(W,17)
4-element Vector{Matrix{Int64}}:
 [-1 -1 0 0; 0 1 0 0; 0 0 1 0; 0 0 0 1]
 [1 0 0 0; -1 -1 -1 0; 0 0 1 0; 0 0 0 1]
 [1 0 0 0; 0 1 0 0; 0 -2 -1 -1; 0 0 0 1]
 [1 0 0 0; 0 1 0 0; 0 0 1 0; 0 0 -1 -1]

julia> traces_words_mats(R,word.(conjugacy_classes(W)))==CharTable(W).irr[17,:]
true

Gapjm.GLinearAlgebra.solutionmat — Function

solutionmat(mat,v)

returns one solution of the equation transpose(mat)*x=v or nothing if no such solution exists. Similar to transpose(mat)\v when mat is invertible.

julia> m=[2 -4 1;0 0 -4;1 -2 -1]
3×3 Matrix{Int64}:
 2  -4   1
 0   0  -4
 1  -2  -1

julia> x=solutionmat(m,[10,-20, -10])
3-element Vector{Rational{Int64}}:
  5//1
 15//4
  0//1

julia> m'*x
3-element Vector{Rational{Int64}}:
  10//1
 -20//1
 -10//1

julia> solutionmat(m,[10, 20, -10])

solutionmat(m,n::AbstractMatrix)

return a matrix x such that x*m==n. This is interesting when m is not invertible.

Gapjm.GLinearAlgebra.rowspace — Function

rowspace(m::AbstractMatrix)

returns a canonical form of the rowspace of m (the vector space generated by the rows of m), the echelon form. This is a particular basis of the rowspace of m: the first non-zero element of each line is 1, and such an element is also the only non-zero element in its column. Integer matrices are converted to Rational before applying this function.

julia> m=[1 2;2 4;5 6]
3×2 Matrix{Int64}:
 1  2
 2  4
 5  6

julia> GLinearAlgebra.rowspace(m)
2×2 Matrix{Rational{Int64}}:
 1//1  0//1
 0//1  1//1

Gapjm.GLinearAlgebra.in_rowspace — Function

in_rowspace(v::AbstractVector,m::AbstractMatrix)

whether v is in the rowspace of m.

Gapjm.GLinearAlgebra.intersect_rowspace — Function

intersect_rowspace(m::AbstractMatrix,n::AbstractMatrix)

The intersection of the rowspaces of m and n.

Gapjm.GLinearAlgebra.lnullspace — Function

lnullspace(m::AbstractMatrix)

The left nullspace of m

Gapjm.GLinearAlgebra.nullspace — Function

GLinearAlgebra.nullspace(m::AbstractMatrix)

computes the right nullspace of m in a type-preserving way. Is not exported to avoid conflict with LinearAlgebra

Gapjm.GLinearAlgebra.rank — Function

GLinearAlgebra.rank(m::AbstractMatrix) computes exactly the rank of m

julia> hilbert(r)=[1//(n+m) for n in 1:r, m in 1:r];

julia> LinearAlgebra.rank(hilbert(11))
10

julia> GLinearAlgebra.rank(hilbert(11)) # correct value
11

Gapjm.GLinearAlgebra.independent_rows — Function

independent_rows(m::AbstractMatrix)

returns a vector of the indices of a set of independent rows of m (the lexicographically first such set).

julia> m=[1 2;2 4;5 6]
3×2 Matrix{Int64}:
 1  2
 2  4
 5  6

julia> GLinearAlgebra.independent_rows(m)
2-element Vector{Int64}:
 1
 3