Hecke algebras

Gapjm.HeckeAlgebras

Gapjm.HeckeAlgebras.hecke

Gapjm.HeckeAlgebras — Module

This module implements Hecke algebras associated to finite complex reflection groups and arbitrary Coxeter groups (these algebras are called Iwahori-Hecke algebras in this last case), and for the algebras of finite groups implements their character tables, Schur elements and representations. For Iwahori-Hecke algebras and G(d,1,1) this module also implements the standard basis (look at the module KLfor Kazhdan-Lusztig bases).

Let (W,S) be a Coxeter system where mₛₜ is the order of st for s,t∈ S. Let R be a commutative ring with 1 and for s∈ S let uₛ₀,uₛ₁∈ R be elements which depend ony on the conjugacy class of s in W (this is the same as requiring that uₛᵢ=uₜᵢ whenever mₛₜ is odd). The Iwahori-Hecke algebra of W over R with parameters uₛᵢ is a deformation of the group algebra of W over R defined as follows: it is the unitary associative R-algebra generated by elements Tₛ, s∈ S subject to the relations:

$(Tₛ-uₛ₀)(Tₛ-uₛ₁)=0$ for all s∈ S (the quadratic relations)

$TₛTₜTₛ…= TₜTₛTₜ…$ with mₛₜ factors on each side (the braid relations)

If uₛ₀=1 and uₛ₁=-1 for all s then the quadratic relations become Tₛ²=1 and the deformation of the group algebra is trivial.

Since the generators Tₛ satisfy the braid relations, the algebra H is in fact a quotient of the group algebra of the braid group associated with W. It follows that, if w=s_1⋯ s_m is a reduced expression of w ∈ W then the product Tₛ_1⋯ Tₛ_m depends only on w. We will therefore denote by T_w. We have T_1=1.

If one of uₛ₀ or uₛ₁ is invertible in R, for example uₛ₁, then by changing the generators to T′ₛ=-Tₛ/uₛ₁, and setting qₛ=-uₛ₀/uₛ₁, the braid relations do no change (since when mₛₜ is odd we have uₛᵢ=uₜᵢ) but the quadratic relations become (T′ₛ-qₛ)(T′ₛ+1)=0. This last form is the most common form considered in the literature. Another common form, considered in the context of Kazhdan-Lusztig theory, is uₛ₀=√qₛ and uₛ₁=-√qₛ⁻¹. The general form of parameters provided is a special case of general cyclotomic Hecke algebras, and can be useful in many contexts.

For some algebras the character table, and in general Kazhdan-Lusztig bases, require a square root of -uₛ₀uₛ₁. We provide a way to specify it with the field .rootpara which can be given when constructing the algebra. If not given a root is automatically extracted when needed (and we know how to compute it) by the function RootParameter. Note however that sometimes an explicit choice of root is necessary which cannot be automatically determined.

There is a universal choice for R and uₛᵢ: Let uₛᵢ:s∈ S,i∈[0,1] be indeterminates such that uₛᵢ=uₜᵢ whenever mₛₜ is odd, and let A=ℤ[uₛᵢ] be the corresponding polynomial ring. Then the Hecke algebra H of W over a with parameters uₛᵢ is called the generic Iwahori-Hecke algebra of with W. Any other algebra with parameters vₛᵢ can be obtained by specialization from H: There is a unique ring homomorphism f:A → R such that f(uₛᵢ)=vₛᵢ for all i. Then we can view R as an A-module via f and we can identify the other algebra to $R⊗ _A H$.

The elements {T_w∣w∈ W} actually form an R-basis of H if one of the uₛᵢ is invertible for all s. The structure constants in that basis is obtained as follows. To multiply T_v by T_w, choose a reduced expression for v, say v=s_1 ⋯ s_k and apply inductively the formula:

$T_sT_w=T_{sw}$ if l(sw)=l(w)+1

$T_sT_w=-uₛ₀uₛ₁T_{sw}+(uₛ₀+uₛ₁)T_w$ if l(sw)=l(w)-1.

If all s we have uₛ₀=q, uₛ₁=-1 then we call the corresponding algebra the one-parameter or Spetsial Iwahori-Hecke algebra associated with W; it can be obtained with the simplified call 'hecke(W,q)'. Certain invariants of the irreducible characters of this algebra play a special role in the representation theory of the underlying finite Coxeter groups, namely the a- and A-invariants. For basic properties of Iwahori-Hecke algebras and their relevance to the representation theory of finite groups of Lie type, see for example Curtis and Reiner 1987, Sections~67 and 68.

In the following example, we compute the multiplication table for the 0-Iwahori–Hecke algebra associated with the Coxeter group of type A_2.

julia> W=coxgroup(:A,2)
A₂

julia> H=hecke(W,0)            # One-parameter algebra with `q=0`
hecke(A₂,0)

julia> T=Tbasis(H);            # Create the `T` basis

julia> el=words(W)
6-element Vector{Vector{Int8}}:
 []       
 [2]      
 [1]      
 [2, 1]   
 [1, 2]   
 [1, 2, 1]

julia> T.(el)*permutedims(T.(el))        # multiplication table
6×6 Matrix{HeckeTElt{Perm{Int16}, Int64, HeckeAlgebra{Int64, FiniteCoxeterGroup{Perm{Int16},Int64}}}}:
 T.    T₂     T₁     T₂₁    T₁₂    T₁₂₁ 
 T₂    -T₂    T₂₁    -T₂₁   T₁₂₁   -T₁₂₁
 T₁    T₁₂    -T₁    T₁₂₁   -T₁₂   -T₁₂₁
 T₂₁   T₁₂₁   -T₂₁   -T₁₂₁  -T₁₂₁  T₁₂₁ 
 T₁₂   -T₁₂   T₁₂₁   -T₁₂₁  -T₁₂₁  T₁₂₁ 
 T₁₂₁  -T₁₂₁  -T₁₂₁  T₁₂₁   T₁₂₁   -T₁₂₁

Thus, we work with algebras with arbitrary parameters. We will see that this also works on the level of characters and representations.

For general complex reflection groups, the picture is similar. The Hecke algebras are deformations of the group algebras, generalizing those for real reflection groups.

The definition is as a quotient of the algebra of the braid group. We assume now that W is a finite reflection group in the complex vector space V. The braid group associated is the fundamental group Π^1 of the space $(V-\bigcup_{H\in\mathcal H} H)/W$, where $\mathcal H$ is the set of reflecting hyperplanes of W. This group is generated by braid reflections, elements which by the natural map from the braid group to the reflection group project to distinguished reflections. The braid reflections which project to a given W-orbit of reflections are conjugate. Let 𝐬 be a representative of such a conjugacy class of braid reflections, let e be the order of the image of 𝐬 in W, and let $u_{𝐬,0},…,u_{𝐬,e-1}$ be indeterminates. The generic Hecke algebra is the $ℤ[u_{𝐬,i}^{± 1}]_{𝐬,i}$-algebra quotient of the braid group algebra by the relations $(𝐬-u_{𝐬,0})…(𝐬-u_{𝐬,e-1})=0$, and in general an Hecke algebra is any algebra obtained from this generic algebra by specializing some of the parameters.

The Hecke algebras can also been defined in terms of presentations. The braid group is presented by homogeneous relations, called braid relations, described in Broué-Malle-Rouquier 1998 and Bessis-Michel 2003 (some were obtained using the VKCURVE package of GAP3, ported also to this Julia package). Further, these relations are such that the reflection group is presented by the same relations, plus relations describing the order of the generating reflections, called the order relations. This allows to define the Hecke algebra by a presentation similar to that of W, with the same braid relations but the order relations replaced by a deformed version. Specifically, for each orbit of reflection hyperplanes of W, let us choose a distinguished reflection s of W, that is a reflection with a non-trivial eigenvalue of minimal argument (i.e., of the form 'E(e)' where e is the order of s; then any reflection around an hyperplane of the same orbit is a conjugate of a power of s). Let then u_{s,0},…,u_{s,e-1} be indeterminates. The generic Hecke algebra is the $ℤ[u_{𝐬,i}^{±1}]_{𝐬,i}$-algebra H with generators T_s in bijection with the generators of W, presented by the braid relations and the deformed order relations $(T_s-u_{𝐬,0})…(T_s-u_{𝐬,e-1})=0$ for each s as above.

Ariki, Koike and Malle have computed character tables for some of these algebras, including those for all 2-dimensional reflection groups, see Broué-Malle 1993 and Malle 1996; the Chevie data contains models of all representation and character tables for real reflection groups; it contains the same for imprimitive groups and for primitive groups of dimension 2 and 3 (these last representations have been computed in Malle-Michel 2010) and contains also models and character tables computed by Michel for G₂₉ and G₃₃; he has computed also partial lists of representations and partial character tables for the remaining groups G₃₁,G₃₂ and G₃₄.

The quotient of the Hecke algebra obtained by specializing $u_{𝐬,i}↦ ζₑⁱ$ is isomorphic to the group algebra of W. It was conjectured for 20 years that over a suitable ring (such as the algebraic closure of the field of fractions $ℚ(u_{𝐬,i})_{𝐬,i }$) the Hecke algebra is itself isomorphic to the group algebra of W over the same ring. This was called the freeness conjecture since the main problem is to show that the Hecke algebra has a basis of cardinality |W|. This has finally been proved in 2019 thanks to the work of many people including Marin, Pfeiffer and Chavli for exceptional groups. Along the way it has been proven that there exists a set {b_w}_{w∈ W} of elements of the Braid group such that b_1=1 and b_w maps to w by the natural quotient map, such that their images T_w form a basis of the Hecke algebra.

It is conjectured that the basis T_w constructed for the freeness conjecture can be chosen such that the linear form t defined by t(T_w)=0 if w≠ 1 and t(1)=1 is a symmetrizing form for the symmetric algebra H. This is well known for all real reflection groups and has been proved in Malle-Mathas 1998 for imprimitive reflection groups and in Malle-Michel 2010 for some primitive groups of dimension 2 and 3. Chlouveraki and Chavli have handled some other 2-dimensional cases. For each irreducible character φ of H we define the Schur element Sᵩ associated to φ by the condition that for any element T of H we have t(T)=∑ᵩ φ(T)/Sᵩ. It can be shown that the Schur elements are Laurent polynomials, and they do not depend on the choice of a basis having the above property. Malle has the computed these Schur elements, assuming the above conjectures; they are in the Chevie data.

finally, benchmarks on julia 1.8

julia> function test_w0(n)
         W=coxgroup(:A,n)
         Tbasis(hecke(W,Pol(:q)))(longest(W))^2
       end
test_w0 (generic function with 1 method)

julia> @btime test_w0(7);
   97.210 ms (1776476 allocations: 127.52 MiB)

in GAP3 the following function takes 0.92s

test_w0:=function(n)local W,T,H;
  W:=CoxeterGroup("A",n);H:=Hecke(W,X(Rationals));T:=Basis(H,"T");
  T(LongestCoxeterWord(W))^2;
end;

Gapjm.HeckeAlgebras.hecke — Method

hecke( W [, parameter][,rootpara=r])

Hecke algebra for the complex reflection group or Coxeter group W. If no parameter is given, 1 is assumed which gives the group algebra of W.

The following forms are accepted for parameter: if parameter is not a vector, it is replaced by fill(parameter,ngens(W)). If it is a vector with one entry, it is replaced with fill(parameter[1],ngens(W)). Otherwise, parameter should be a list of length ngens(W). Entries of parameter corresponding to the same W-orbit of generators should be identical. parameter can be shorter than ngens(W) provided there is at least one entry bound for each orbit of reflections.

An entry in parameter for a reflection of order e can be either a single value or a Vector of length 'e'. If it is a Vector, it is interpreted as the list [u₀,…,u_(e-1)] of parameters for that reflection. If it is a single value q, it is interpreted as the partly specialized list of parameters [q,ζ_e,…,ζ_{e-1}] (thus [q,-1] for Coxeter groups).

Computing characters or representations of Hecke algebra needs sometimes to extract roots of the parameters. These roots are extracted automatically (when possible). For Coxeter groups it is possible to give explicit roots by giving a keyword argument rootpara: it should be a vector containing at the i-th position a square root of -parameter[i][1]*parameter[i][2] (if rootpara is not a Vector it is replaced by fill(rootpara,ngens(W))).

Example

julia> W=coxgroup(:B,2)
B₂

julia> @Pol q
Pol{Int64}: q

julia> H=hecke(W,q)
hecke(B₂,q)

julia> H.para
2-element Vector{Vector{Pol{Int64}}}:
 [q, -1]
 [q, -1]

julia> H=hecke(W,q^2,rootpara=q)
hecke(B₂,q²,rootpara=q)

julia> H.para,rootpara(H)
(Vector{Pol{Int64}}[[q², -1], [q², -1]], Pol{Int64}[q, q])

julia> H=hecke(W,[q^2,q^4],rootpara=[q,q^2])
hecke(B₂,Pol{Int64}[q², q⁴],rootpara=Pol{Int64}[q, q²])

julia> H.para,rootpara(H)
(Vector{Pol{Int64}}[[q², -1], [q⁴, -1]], Pol{Int64}[q, q²])

julia> H=hecke(W,9,rootpara=3)
hecke(B₂,9,rootpara=3)

julia> H.para,rootpara(H)
([[9, -1], [9, -1]], [3, 3])

julia> @Mvp x,y,z,t

julia> H=hecke(W,[[x,y]])
hecke(B₂,Vector{Mvp{Int64, Int64}}[[x, y]])

julia> H.para,rootpara(H)
(Vector{Mvp{Int64, Int64}}[[x, y], [x, y]], Mvp{Cyc{Int64}, Rational{Int64}}[ζ₄x½y½, ζ₄x½y½])

julia> H=hecke(W,[[x,y],[z,t]])
hecke(B₂,Vector{Mvp{Int64, Int64}}[[x, y], [z, t]])

julia> H.para,rootpara(H)
(Vector{Mvp{Int64, Int64}}[[x, y], [z, t]], Mvp{Cyc{Int64}, Rational{Int64}}[ζ₄x½y½, ζ₄t½z½])

julia> hecke(complex_reflection_group(3,1,2),q).para
2-element Vector{Vector{Pol{Cyc{Int64}}}}:
 [q, ζ₃, ζ₃²]
 [q, -1]