Posets

Gapjm.Posets

Gapjm.Posets.CPoset

Gapjm.Posets.Poset

Gapjm.Posets.Poset

Gapjm.Posets.covering_chains

Gapjm.Posets.dual

Gapjm.Posets.hasse

Gapjm.Posets.incidence

Gapjm.Posets.induced

Gapjm.Posets.isjoinlattice

Gapjm.Posets.ismeetlattice

Gapjm.Posets.linear_extension

Gapjm.Posets.maxima

Gapjm.Posets.minima

Gapjm.Posets.moebius

Gapjm.Posets.moebiusmatrix

Gapjm.Posets.partition

Gapjm.Posets.transitive_closure

Gapjm.Posets — Module

This package deals with finite posets.

There are two kinds of posets. A "canonical poset" or CPoset is on the elements 1:n where n=length(P). A Poset is on a given list of elements which may be of any type. A Poset contains internally aCPoset working on the indices of the elements, which is more efficient than working with the elements themselves. Similarly for efficiency many functions work with the CPoset by transformeing their input to indices and their output to elements.

A CPoset has the field:

hasse: a list representing the Hasse diagram of the poset: the i-th entry is the list of elements which cover (are immediate successors of) i, that is the list of j such that i<j and there is no k such that i<k<j.

The following is cached when computed to speed up subsequent computations:

incidence: a boolean matrix such that incidence[i,j]==true iff i<=j. This is sometimes called the ζ-matrix of the poset.

There are several ways of defining a poset. By entering the Hasse diagram:

julia> p=CPoset([[2,3],[4],[4],Int[]])
1<2,3<4

As seen above, p is shown as a list of covering maximal chains; elements which are equivalent for the poset are printed together separated by commas.

julia> length(p) # the number of elements of `p`
4

julia> incidence(p)
4×4 Matrix{Bool}:
 1  1  1  1
 0  1  0  1
 0  0  1  1
 0  0  0  1

julia> linear_extension(p) # a total order compatible with p
4-element Vector{Int64}:
 1
 2
 3
 4

A Poset is constructed from a CPoset and a list of elements

julia> P=Poset(p,[:a,:b,:c,:d])
a<b,c<d

julia> P.C # the CPoset attached to P
1<2,3<4

A convenient constructor for Posets takes a function representing isless for the poset and the list of elements and constructs the poset from the incidence matrix, computed by applying the function to each pair of elements. For isless one can give either a function implementing < or a function implementing ≤ (it is or-ed with = in any case).

julia> l=vec(collect(Iterators.product(1:2,1:2)))
4-element Vector{Tuple{Int64, Int64}}:
 (1, 1)
 (2, 1)
 (1, 2)
 (2, 2)

julia> P=Poset((x,y)->all(map(<=,x,y)),l)
(1, 1)<(2, 1),(1, 2)<(2, 2)

julia> eltype(P) # the type of the elements of P
Tuple{Int64, Int64}

julia> summary(P) # useful for big posets
"Poset{Tuple{Int64, Int64}} of length 4"

A poset can also be constructed from an incidence matrix so the last example could also be entered as

julia> P=Poset(CPoset([all(map(<=,x,y)) for x in l, y in l]),l)
(1, 1)<(2, 1),(1, 2)<(2, 2)

Flexibility on printing a Poset is obtained by setting the function show_element which takes as arguments an IO, the poset, and the index of the element to print:

julia> P.show_element=(io,p,n)->join(io,p.elements[n],".");

julia> P
1.1<2.1,1.2<2.2

julia> delete!(P,:show_element); # back to default

The above fancy printing applies only when printing at the REPL or in pluto or Jupyter. The default printing gives a form which can be input back in Julia

julia> print(P) 
Poset(CPoset([[2, 3], [4], [4], Int64[]]),[(1, 1), (2, 1), (1, 2), (2, 2)])

A poset can be specified by a list of tuples specifing order relations. The transitive closure of these relations is computed, giving an incidence matrix from which the poset is built. The elements of the poset, if not specified separately, will be all the elements which appear in the tuples.

julia> Poset([(:a,:b),(:c,:d)])
a<b
c<d

julia> CPoset([(1,3),(2,5)]) # the CPoset is on 1:maximum(entries)
4
1<3
2<5

To get the order relation ≤ of the poset p between elements i and j just call ≤(p,i,j).

julia> ≤(P,(1,1),(2,1))
true

julia> ≤(P.C,1,2) # the same
true

Intervals in a poset can be computed with strict or not bounds.

julia> interval(P,≤,(1,2)) # elements below (1,2)
2-element Vector{Tuple{Int64, Int64}}:
 (1, 1)
 (1, 2)

julia> interval(P,≥,(1,2)) # elements above (1,2)
2-element Vector{Tuple{Int64, Int64}}:
 (1, 2)
 (2, 2)

julia> interval(P,<,(1,2)) # elements strictly below (1,2)
1-element Vector{Tuple{Int64, Int64}}:
 (1, 1)

julia> interval(P,≥,(2,1),≤,(2,2)) # elements between (2,1) and (2,2)
2-element Vector{Tuple{Int64, Int64}}:
 (2, 1)
 (2, 2)

julia> interval(P,>,(1,1),<,(2,2)) # elements strictly between
2-element Vector{Tuple{Int64, Int64}}:
 (2, 1)
 (1, 2)
julia> interval(P.C,>,1,<,4) # in terms of indices
2-element Vector{Int64}:
 2
 3

A sample of other functions available on posets:

julia> maximal_chains(P)
2-element Vector{Vector{Tuple{Int64, Int64}}}:
 [(1, 1), (2, 1), (2, 2)]
 [(1, 1), (1, 2), (2, 2)]

julia> height(P) # the length of a maximal chain
3

julia> moebiusmatrix(P)
4×4 Matrix{Int64}:
 1  -1  -1   1
 0   1   0  -1
 0   0   1  -1
 0   0   0   1

julia> minima(P)
1-element Vector{Tuple{Int64, Int64}}:
 (1, 1)

julia> maxima(P)
1-element Vector{Tuple{Int64, Int64}}:
 (2, 2)

julia> Q=CPoset(:chain,3)
1<2<3

julia> P1=Poset(Q) # transformed to a Poset with elements 1:3
1<2<3

julia> P⊕ P1 # the ordinal sum
(1, 1)<(2, 1),(1, 2)<(2, 2)<1<2<3

julia> P1*P1
(1, 1)<(2, 1)<(3, 1)<(3, 2)<(3, 3)
(1, 1)<(1, 2)<(2, 2)<(3, 2)
(2, 1)<(2, 2)
(1, 2)<(1, 3)<(2, 3)<(3, 3)
(2, 2)<(2, 3)

julia> P1⊗ P1 # the ordinal product
(1, 1)<(1, 2)<(1, 3)<(2, 1)<(2, 2)<(2, 3)<(3, 1)<(3, 2)<(3, 3)

Finally showpic(p) where p is a CPoset or a Poset gives a graphical display of the poset (on Linux) provided you have the command dot and the command display of imagemagick installed. It may work on MacOs and Windows but I did not test it.

see the on-line help on ⊕, ⊗, chains, chainpoly, covering_chains, coxetermatrix, dual, hasse, height, incidence, induced, interval, isjoinlattice, ismeetlattice, linear_extension, maxima, maximal_chains, minima, moebius, moebiusmatrix, partition, showpic, transitive_closure for more information

Gapjm.Posets.Poset — Type

Poset(p::CPoset,e::AbstractVector)

creates a Poset with order specified by p and elements e.

Poset(W::CoxeterGroup,w=longest(W))

returns as a poset the Bruhat interval [1,w]of W. If w is not given, the whole Bruhat Poset of W is returned (W must then be finite).

julia> W=CoxSym(3)
𝔖 ₃

julia> Poset(W)
.<1,2<21,12<121

The above poset is constructed efficiently by constructing the Hasse diagram, but it could be constructed naively as follows:

julia> p=Poset((x,y)->bruhatless(W,x,y),elements(W))
()<(2,3),(1,2)<(1,2,3),(1,3,2)<(1,3)

The element printing is not so nice, showing permutations instead of words. This can be remedied by giving a function:

julia> p.show_element=(io,x,n)->(e=elements(x,n);print(io,isone(e) ? "." : joindigits(word(W,e))));

julia> p
.<2,1<21,12<121

julia> W=CoxSym(4)
𝔖 ₄

julia> Poset(W,W(1,3))
.<3,1<13

Gapjm.Posets.CPoset — Type

CPoset(m::Matrix{Bool})

Creates a poset from an incidence matrix m, that is m[i,j]==true if and only if i≤j in the poset,

julia> CPoset(Bool[1 1 1 1 1;0 1 0 1 1;0 0 1 1 1;0 0 0 1 0;0 0 0 0 1])
1<2,3<4,5

CPoset(h::Vector{<:Vector{<:Integer}})

Creates a poset from a Hasse diagram given as a Vector whose i-th entry is the list of indices which are immediate successors (covers) of the i-th element, that is h[i] is the list of j such that i<j in the poset and such that there is no k such that i<k<j.

julia> CPoset([[2,3],[4,5],[4,5],Int[],Int[]])
1<2,3<4,5

Gapjm.Posets.hasse — Function

hasse(P::CPoset)

the Hasse diagram of P.

julia> p=CPoset((i,j)->j%i==0,5)
1<3,5
1<2<4

julia> hasse(p)
5-element Vector{Vector{Int64}}:
 [2, 3, 5]
 [4]      
 []       
 []       
 []       

Gapjm.Posets.incidence — Function

incidence(P::CPoset)

returns the incidence matrix (also called the ζ matrix) of P.

julia> p=CPoset([i==6 ? Int[] : [i+1] for i in 1:6])
1<2<3<4<5<6

julia> incidence(p)
6×6 Matrix{Bool}:
 1  1  1  1  1  1
 0  1  1  1  1  1
 0  0  1  1  1  1
 0  0  0  1  1  1
 0  0  0  0  1  1
 0  0  0  0  0  1

Gapjm.Posets.transitive_closure — Function

transitive_closure(M) transitive_closure!(M)

M should be a square boolean matrix representing a relation; transitive_closure returns a boolean matrix representing the transitive closure of this relation; transistive_closure! modifies M in place, doing no allocations. The transitive closure is computed by the Floyd-Warshall algorithm, which is quite fast even for large matrices.

julia> m=[j-i in [0,1] for i in 1:5, j in 1:5]
5×5 Matrix{Bool}:
 1  1  0  0  0
 0  1  1  0  0
 0  0  1  1  0
 0  0  0  1  1
 0  0  0  0  1

julia>transitive_closure(m)
5×5 Matrix{Bool}:
 1  1  1  1  1
 0  1  1  1  1
 0  0  1  1  1
 0  0  0  1  1
 0  0  0  0  1

Gapjm.Posets.linear_extension — Function

linear_extension(P)

returns a linear extension of the poset P, that is a vector l containing a permutation of the integers 1:length(P) such that if i<j in P (that is incidence(P)[i,j] is true), then i is before j in l. This is also called a topological sort of P.

julia> p=CPoset((i,j)->j%i==0,6) # divisibility poset on 1:6
1<5
1<2<4
1<3<6
2<6

julia> linear_extension(p)
6-element Vector{Int64}:
 1
 2
 3
 5
 4
 6

Gapjm.Posets.dual — Function

dual(P::CPoset)

the dual poset to P (the order relation is reversed).

julia> p=CPoset((i,j)->i%4<j%4,8)
4,8<1,5<2,6<3,7

julia> dual(p)
3,7<2,6<1,5<4,8

Gapjm.Posets.partition — Function

partition(P::CPoset)

returns the partition of 1:length(P) induced by the equivalence relation associated to P; that is, i and j are in the same part of the partition if the k such that i<k and j<k are the same as well as the k such that k<i and k<j.

julia> p=CPoset([i==j || i%4<j%4 for i in 1:8, j in 1:8])
4,8<1,5<2,6<3,7

julia> partition(p)
4-element Vector{Vector{Int64}}:
 [4, 8]
 [2, 6]
 [3, 7]
 [1, 5]

Gapjm.Posets.covering_chains — Function

covering_chains(P::CPoset)

A (greedy: the first is longest possible) list of covering chains for P.

Gapjm.Posets.induced — Function

induced(P::CPoset,S)

returns the subposet induced by P on S, a sublist of P.elements. Note that the sublist S does not have to be in the same order as P.elements, so this can be just used to renumber the elements of P.

julia> p=CPoset((i,j)->i%4<j%4,8)
4,8<1,5<2,6<3,7

julia> induced(p,2:6) # indices are renumbered
3<4<1,5<2

julia> induced(Poset(p),2:6) # elements are kept
4<5<2,6<3

Gapjm.Posets.isjoinlattice — Function

isjoinlattice(P::CPoset)

returns true if P is a join semilattice, that is any two elements of P have a unique smallest upper bound; returns false otherwise.

julia> p=CPoset((i,j)->j%i==0,8)
1<5,7
1<2<4<8
1<3<6
2<6

julia> isjoinlattice(p)
false

Gapjm.Posets.ismeetlattice — Function

ismeetlattice(P)

returns true if P is a meet semilattice, that is any two elements of P have a unique highest lower bound; returns false otherwise.

julia> p=CPoset((i,j)->j%i==0,8)
1<5,7
1<2<4<8
1<3<6
2<6

julia> ismeetlattice(p)
true

Gapjm.Posets.moebius — Function

moebius(P,y=maximum(P))

the vector of values μ(x,y) of the Moebius function of P for x varying. Here is an example giving the ususal Moebius function on integers.

julia> p=CPoset((i,j)->i%j==0,1:8)
5,7<1
6<2<1
6<3<1
8<4<2

julia> moebius(p)
8-element Vector{Int64}:
  1
 -1
 -1
  0
 -1
  1
 -1
  0

Gapjm.Posets.moebiusmatrix — Function

moebiusmatrix(P::CPoset) the matrix of the Moebius function μ(x,y)

Gapjm.Posets.maxima — Function

julia> p=CPoset([[3],[3],[4,5],Int[],Int[]])
1,2<3<4,5

julia> maxima(p)
2-element Vector{Int64}:
 4
 5

Gapjm.Posets.minima — Function

julia> p=CPoset([[3],[3],[4,5],Int[],Int[]])
1,2<3<4,5

julia> minima(p)
2-element Vector{Int64}:
 1
 2

Gapjm.Posets.Poset — Type

Poset(W::CoxeterGroup,w=longest(W))

returns as a poset the Bruhat interval [1,w]of W. If w is not given, the whole Bruhat Poset of W is returned (W must then be finite).

julia> W=CoxSym(3)
𝔖 ₃

julia> Poset(W)
.<1,2<21,12<121

The above poset is constructed efficiently by constructing the Hasse diagram, but it could be constructed naively as follows:

julia> p=Poset((x,y)->bruhatless(W,x,y),elements(W))
()<(2,3),(1,2)<(1,2,3),(1,3,2)<(1,3)

The element printing is not so nice, showing permutations instead of words. This can be remedied by giving a function:

julia> p.show_element=(io,x,n)->(e=elements(x,n);print(io,isone(e) ? "." : joindigits(word(W,e))));

julia> p
.<2,1<21,12<121

julia> W=CoxSym(4)
𝔖 ₄

julia> Poset(W,W(1,3))
.<3,1<13