Signed permutations

Gapjm.SPerms

Base.Matrix

Gapjm.SPerms.CoxHyperoctaedral

Gapjm.SPerms.SPerm

PermGroups.Perms.Perm

Gapjm.PermRoot.reflection_subgroup

Gapjm.SPerms.sstab_onmats

PermGroups.Perms.cycles

PermGroups.Perms.cycletype

PermGroups.Perms.orbit

PermGroups.Perms.order

PermGroups.Perms.permute

Gapjm.SPerms.@sperm_str

Gapjm.SPerms — Module

A signed permutation of 1:n is a permutation of the set -n,…,-1,1,…,n which preserves the pairs (-i,i). It is represented internally as the images of 1:n. It is printed as a product of signed cycles.

Examples

julia> SPerm([-2,-1,-3])
SPerm{Int64}: (1,-2)(3,-3)

julia> p=SPerm(-1)
(1,-1)

julia> q=SPerm(1,2)
(1,2)

julia> sort(elements(Group([p,q])))
8-element Vector{SPerm{Int16}}:
 (1,-2)
 (1,-2,-1,2)
 (1,-1)(2,-2)
 (1,-1)
 (2,-2)
 ()
 (1,2,-1,-2)
 (1,2)

The complete type of signed permutations is SPerm{T} where T<:Integer, where Vector{T} is the type of the vector which holds the image of 1:n. This can used to save space or time when possible. If T is not specified we take it to be Int16 since this is a good compromise between speed and compactness.

SPerms have methods copy, hash, ==, isless (total order) so they can be keys in hashes or elements of sets; two SPerms are equal if they move the same points to the same images. For instance,

julia> SPerm([-2,-1,-3])==SPerm([-2,-1,-3,4])
true

SPerms are considered as scalars for broadcasting.

Gapjm.SPerms.SPerm — Type

struct SPerm

An SPerm represents a signed permutation of 1:n, that is a permutation of the set -n,…,-1,1,…,n which preserves the pairs (-i,i). It is implemented by a struct SPerm with one field d, a vector holding the images of 1:n.

PermGroups.Perms.Perm — Method

Perm(p::SPerm) returns the underlying Perm of an SPerm

Gapjm.SPerms.@sperm_str — Macro

@sperm"..."

make a SPerm from a string; allows the style sperm"(1,-2)(5,-6,7)(-4,9)"

PermGroups.Perms.permute — Method

permute(l::AbstractVector,p::SPerm)

returns l permuted by p, a vector r such that r[abs(i^p)]=l[i]*sign(i^p).

Examples

julia> p=SPerm([-2,-1,-3])
SPerm{Int64}: (1,-2)(3,-3)

julia> permute([20,30,40],p)
3-element Vector{Int64}:
 -30
 -20
 -40

PermGroups.Perms.orbit — Method

orbit(a::SPerm,i::Integer) returns the orbit of a on i.

PermGroups.Perms.order — Method

order(a::SPerm) is the order of the signed permutation a.

PermGroups.Perms.cycles — Method

cycles(p::SPerm) the non-trivial cycles of p.

Two cycles which differ only by sign are returned once only.

julia> cycles(SPerm(-1,2)*SPerm(3,-3)*SPerm(4,5,-4,-5))
3-element Vector{Vector{Int16}}:
 [1, -2]
 [3, -3]
 [4, 5, -4, -5]

PermGroups.Perms.cycletype — Method

cycletype(p::SPerm,n=length(p.d)) pair of partitions parameterizing the conjugacy class of p in Bₙ

julia> cycletype(SPerm(1,-1),2)
2-element Vector{Vector{Int64}}:
 [1]
 [1]

Base.Matrix — Method

Matrix(a::SPerm) is the permutation matrix for a

Examples

julia> Matrix(SPerm([-2,-1,-3]))
3×3 Matrix{Int64}:
  0  -1   0
 -1   0   0
  0   0  -1

Gapjm.SPerms.CoxHyperoctaedral — Type

CoxHyperoctaedral(n) The Hyperoctaedral group (the group of all signed permutations of ±1,…,±n) as a Coxeter group of type Bₙ, with generators (1,-1) and (i,i+1)(-i,-i-1).

julia> elements(CoxHyperoctaedral(2))
8-element Vector{SPerm{Int8}}:
 ()
 (1,2)
 (1,-1)
 (1,2,-1,-2)
 (1,-2,-1,2)
 (2,-2)
 (1,-2)
 (1,-1)(2,-2)

Gapjm.PermRoot.reflection_subgroup — Method

reflection_subgroup(W::CoxHyperoctaedral,I) is defined only for I=1:m for m≤ngens(W).

Gapjm.SPerms.sstab_onmats — Function

sstab_onmats([G,]M[,l])

If onmats(m,p)=^(M,p;dims=(1,2)) (simultaneous signed conjugation of rows and columns, or conjugating by the matrix of the signed permutation p), and the argument G is given (which should be an SPermGroup) this is just a fast implementation of centralizer(G,M,onmats). If G is omitted it is taken to be CoxHyperoctaedral(size(M,1)). The program uses sophisticated algorithms, and can handle matrices up to 80×80. If l is given the return group should also centralize l (for the action ^)

julia> uc=UnipotentCharacters(complex_reflection_group(6));

julia> g=sstab_onmats(fourier(uc.families[2]))
Group([(1,18)(3,-6)(8,-21)(10,-16)(11,22)(13,15),(1,-15)(2,-19)(3,-11)(6,22)(7,-12)(13,-18),(2,19)(4,-14)(5,20)(7,12),(1,-11)(2,-19)(3,-15)(5,-20)(6,13)(8,10)(16,21)(17,-17)(18,-22),(1,-22)(2,-19)(3,-13)(5,-20)(6,15)(8,-16)(10,-21)(11,-18)(17,-17),(1,6)(2,-19)(3,-18)(4,14)(8,16)(9,-9)(10,21)(11,-13)(15,-22),(1,13)(3,22)(4,14)(5,-20)(6,-11)(8,21)(9,-9)(10,16)(15,18)(17,-17)])
julia> length(g)
32