Finite Coxeter groups and Weyl groups

Gapjm.Weyl.badprimes

Gapjm.Weyl.coxgroup

Gapjm.Weyl.describe_involution

Gapjm.Weyl.istorus

Gapjm.Weyl.relative_group

Gapjm.Weyl.rootdatum

Gapjm.Weyl.rootlengths

Gapjm.Weyl.standard_parabolic

Gapjm.Weyl.torus

Gapjm.Weyl.two_tree

Gapjm.Weyl.with_inversions

Gapjm.Weyl — Module

Finite Coxeter groups are the finite complex reflection groups which can be defined on a real vector space V.

Weyl groups are the finite complex which reflection groups which can be defined over a rational vector space V (thus over the integers).

These groups being complex reflection groups, all methods defined for complex reflection groups are defined on them. They are in particular implemented as groups of permutations of a root system. The particular root systems for Weyl groups play an important role in mathematics as they classify semi-simple Lie algebras and algebraic groups.

Let us give precise definitions. Let V be a real vector space and Vⱽ its dual. A root system is a finite set of vectors R⊂ V (the roots), together with a map r↦ rⱽ from R to a subset Rⱽ of Vⱽ (the coroots) such that:

For any r∈ R, we have rⱽ(r)=2, so that the formula x↦ x-rⱽ(x)r defines a reflection sᵣ:V→ V with root r and coroot rⱽ.

The reflection sᵣ stabilizes R.

The subgroup W=W(R) of GL(V) generated by the reflections sᵣ is a finite Coxeter group. For Coxeter groups, we use reduced root systems, that is such that the only elements of R colinear with r∈ R are r and -r; for Weyl groups, we also ask that the root system be crystallographic, that is rⱽ(s) is an integer, for any s∈ R,rⱽ∈ Rⱽ.

If we identify V with Vⱽ by choosing a W-invariant bilinear form (.;.), then we have rⱽ=2r/(r;r). A root system R is irreducible if R is not the union of two orthogonal subsets (equivalently the representation of W on the subspace generated by the R is irreducible). If R is reducible then the corresponding Coxeter group is the direct product of the Coxeter groups associated with the irreducible components of R.

Let us now describe how the root system and a presentation of W are encoded in the Dynkin diagrams. Let R be a root system in V. Then we can choose a linear form on V which vanishes on no element of R. According to the sign of the value of this linear form on a root r ∈ R we call r positive or negative. Then there exists a unique subset Π of the positive roots, the simple roots, such that any positive root is a linear combination with non-negative coefficients of roots in Π. Any two sets of simple roots (corresponding to different choices of linear forms as above) can be transformed into each other by a unique element of W(R).

Since the pairing between V and Vⱽ is W-invariant, if Π is a set of simple roots and if we define the Cartan matrix as being the n times n matrix C={rⱽ(s)}ᵣₛ, for r,s∈Π. This matrix is unique up to simultaneous permutation of rows and columns.

For a crystallographic root system the Cartan matrix has integral entries, and in the basis Π (completed by a basis of the orthogonal), sᵣ has an integral matrix. All finite-dimensional (complex) representations of a finite Coxeter group can be realized over the field generated by the entries of the Cartan matrix.

It is precisely the Cartan matrix which is encoded in a Dynkin diagram, as follows. The vertices represent the simple roots (giving generating reflections, called also Coxeter generators or simple reflections); an edge is drawn between s and t if the order mₛₜ of st is greater than 2; the edge is single if mₛₜ=3, double if mₛₜ=4, triple if mₛₜ=6. The arrows indicate the relative root lengths (going from the longer to the shorter root) which may be different when W has more than one orbit on R. The irreducible crystallographic root systems are classified by the following list. The labeling of the nodes is the order of the generators and is shown by the function Diagram.

Aₙ O—O—O—…—O   Bₙ O⇐ O—O—…—O  Cₙ O⇒ O—O—…—O  Dₙ O 2
   1 2 3 … n      1  2 3 … n     1  2 3 … n     ￨
                                              O—O—…—O
                                              1 3 … n

G₂ O⇛ O  F₄ O—O⇒O—O    E₆  O 2    E₇  O 2      E₈  O 2
   1  2     1 2 3 4        ￨          ￨            ￨
                       O—O—O—O—O  O—O—O—O—O—O  O—O—O—O—O—O—O
                       1 3 4 5 6  1 3 4 5 6 7  1 3 4 5 6 7 8

We get the Coxeter Diagram, which describes the underlying Weyl group, if we ignore the arrows: we see that the root systems B_n and C_n correspond to the same Coxeter group (the Coxeter group is defined by the Coxeter matrix {mₛₜ}ₛₜ). Weyl groups can also be characterized as the finite Coxeter groups such that all entries of the Coxeter matrix are in {2,3,4,6}.

Here are the Coxeter diagrams for the finite Coxeter groups which are not crystallographic:

       e        5         5
I₂(e) O—O   H₃ O—O—O  H₄ O—O—O—O
      1 2      1 2 3     1 2 3 4

The Dynkin diagram can be obtained from the Cartan matrix C as follows: the indices for the rows of C label the nodes of the diagram. The edges, for r ≠ s, are given as follows. If Cᵣₛ and Cₛᵣ are integers such that |Cₛᵣ|≥|Cᵣₛ|=1 there is an edge of weight |Cₛᵣ| from r to s with an arrow pointing to s if |Cₛᵣ|>1. For non-integral entries, we put a an edge of weight pᵣₛ≥1 where CᵣₛCₛᵣ=cos^2 (π/pₛᵣ).

The Cartan matrix corresponding to one of the above irreducible root systems (with the specified labeling) is returned by the command cartan which takes as input a Symbol giving the type (that is :A, :B, …, :I) and a positive Int giving the rank (plus an Int giving the bond for type :I). This function returns a matrix with Int entries for Weyl groups, and a matrix of Cyc for the other types. Given two Cartan matrices c1 and c2, their matrix direct sum (corresponding to the orthogonal direct sum of the root systems) can be obtained by cat(c1,c2,dims=[1,2]).

The whole root system can be recovered from the simple roots and the corresponding coroots, since each root is in the orbit of a simple root. The restriction of the simple reflections to the span of R is determined by the Cartan matrix, so R is determined by the Cartan matrix and the set of simple roots.

The function rootdatum takes as input a list of simple roots and a list of the corresponding coroots and produces a CoxeterGroup containing information about the root system R and about W(R). If we label the positive roots by 1:N, and the negative roots by N+1:2N, then each simple reflection is represented by the permutation of 1:2N which it induces on the roots. If only one argument is given, the Cartan matrix of the root system, it is taken as the list of coroots and the list of roots is assumed to be the canonical basis of V.

If one only wants to work with Cartan matrices with a labeling as specified by the above list, the function call can be simplified. Instead of rootdatum(cartan(:D,4)) the following is also possible.

julia> W=coxgroup(:D,4)
D₄

julia> cartan(W)
4×4 Matrix{Int64}:
  2   0  -1   0
  0   2  -1   0
 -1  -1   2  -1
  0   0  -1   2

Also, the Weyl group struct associated to a direct sum of irreducible root systems can be obtained as a product

julia> W=coxgroup(:A,2)*coxgroup(:B,2)
A₂×B₂

julia> cartan(W)
4×4 Matrix{Int64}:
  2  -1   0   0
 -1   2   0   0
  0   0   2  -2
  0   0  -1   2

The same object is constructed by applying coxgroup to the matrix cat(cartan(:A,2), cartan(:B,2),dims=[1,2]).

The elements of a Weyl group are permutations of the roots:

julia> W=coxgroup(:D,4)
D₄

julia> p=W(1,3,2,1,3)
(1,14,13,2)(3,17,8,18)(4,12)(5,20,6,15)(7,10,11,9)(16,24)(19,22,23,21)

julia> word(W,p)
5-element Vector{Int64}:
 1
 3
 1
 2
 3

finally, a benchmark on julia 1.0.2

julia> @btime length(elements(coxgroup(:E,7)))
  531.385 ms (5945569 allocations: 1.08 GiB)

GAP3 for the same computation takes 2.2s

Gapjm.PermRoot.cartan — Method

cartan(type, rank [,bond])

the Cartan matrix for a finite Coxeter group described by type and rank. The recognized types are :A, :B, :Bsym, :C, :D, :E, :F, :Fsym, :G, :Gsym, :I, :H. For type :I a third argument must be given describing the bond between the two generators. The sym types correspond to (non-crystallographic) root systems where all roots have the same length. They afford automorphisms that the crystallographic root system does not afford, which allow to define the "very twisted" Chevalley groups.

julia> cartan(:F,4)
4×4 Matrix{Int64}:
  2  -1   0   0
 -1   2  -1   0
  0  -2   2  -1
  0   0  -1   2

julia> cartan(:I,2,5)
2×2 Matrix{Cyc{Int64}}:
       2  ζ₅²+ζ₅³
 ζ₅²+ζ₅³        2

julia> cartan(:Bsym,2)
2×2 Matrix{Cyc{Int64}}:
   2  -√2
 -√2    2

Gapjm.PermRoot.cartan — Method

cartan(M::AbstractMatrix) Cartan matrix from Coxeter matrix

The argument should be the Coxeter matrix M for a Coxeter group W and the result is the Cartan Matrix C for the standard reflection representation of W. We have C[s,t]=-2cos(π/M[s,t]), where M[s,s]==1 and by convention π/M[s,t]==0 if M[s,t]==∞, which we represent by M[s,t]==0. Since M is symmetric, the resulting C is symmetric, meaning that all roots in the constructed reflection representation have same length.

julia> cartan([1 3;3 1])
2×2 Matrix{Cyc{Int64}}:
  2  -1
 -1   2

Gapjm.PermRoot.roots — Method

roots(C::AbstractMatrix)

returns the set of positive roots defined by the Cartan matrix C. Works for the Cartan matrix of any finite Coxeter group

Gapjm.Weyl.two_tree — Function

two_tree(m)

Given a square matrix m with zeroes (or falses, for a boolean matrix) symmetric with respect to the diagonal, let G be the graph with vertices axes(m)[1] and an edge between i and j iff !iszero(m[i,j]).

If G is a line this function returns it as a Vector{Int}. If G is a tree with one vertex c of valence 3 the function returns (c,b1,b2,b3) where b1,b2,b3 are the branches from this vertex sorted by increasing length. Otherwise the function returns nothing

julia> Weyl.two_tree(cartan(:A,4))
4-element Vector{Int64}:
 1
 2
 3
 4

julia> Weyl.two_tree(cartan(:E,8))
(4, [2], [3, 1], [5, 6, 7, 8])

Gapjm.PermRoot.reflection_subgroup — Method

reflection_subgroup(W::FiniteCoxeterGroup,I)

The subgroup of W generated by refls(W,I)

A theorem discovered independently by Deodhar1989 and Dyer1990 is that a subgroup H of a Coxeter system (W,S) generated by reflections has a canonical Coxeter generating set, formed of the t ∈ Ref(H) such l(tt')>l(t) for any t'∈ Ref(H) different from t. This is used by reflection_subgroup to determine the Coxeter system of H.

julia> W=coxgroup(:G,2)
G₂

julia> Diagram(W)
O⇛ O G₂
1  2

julia> H=reflection_subgroup(W,[2,6])
G₂₍₂₆₎=Ã₁×A₁

julia> Diagram(H)
O Ã₁
1
O A₁
2

The notation G₂₍₂₆₎ means that W.G.roots[2:6] form a system of simple roots for H.

A reflection subgroup has specific properties the most important of which is inclusion which gives the positions of the roots of H in the roots of W. The inverse (partial) map is restriction.

julia> inclusion(H)
4-element Vector{Int64}:
  2
  6
  8
 12

julia> restriction(H)
12-element Vector{Int64}:
 0
 1
 0
 0
 0
 2
 0
 3
 0
 0
 0
 4

If H is a standard parabolic subgroup of a Coxeter group W then the length function on H (with respect to its set of generators) is the restriction of the length function on W. This need not no longer be true for arbitrary reflection subgroups of W:

julia> word(W,H(2))
5-element Vector{Int64}:
 1
 2
 1
 2
 1

In this package, finite reflection groups are represented as permutation groups on a set of roots. Consequently, a reflection subgroup H⊆ W is a permutation subgroup, thus its elements are represented as permutations of the roots of the parent group.

julia> elH=word.(Ref(H),elements(H))
4-element Vector{Vector{Int64}}:
 []
 [2]
 [1]
 [1, 2]

julia> elW=word.(Ref(W),elements(H))
4-element Vector{Vector{Int64}}:
 []
 [1, 2, 1, 2, 1]
 [2]
 [1, 2, 1, 2, 1, 2]

julia> map(w->H(w...),elH)==map(w->W(w...),elW)
true

Another basic result about reflection subgroups of Coxeter groups is that each coset of H in W contains a unique element of minimal length, see reduced.

Gapjm.Weyl.coxgroup — Function

coxgroup(type,rank[,bond])

This is equivalent to rootdatum(cartan(type,rank[,bond])).

The resulting object W, that we will call a Coxeter datum, has an additional entry compared to a PermRootGroup.

W.rootdec: the root vectors, given as linear combinations of simple roots. The first nref(W) roots are positive, the next nref(W) are the corresponding negative roots. Moreover, the first semisimplerank(W) roots are the simple roots. The positive roots are ordered by increasing height.

see the following functions for how to get various information on the root system and the Coxeter group

nref, coroots, rootlengths, simple_reps, simple_conjugating, reflrep, simpleroots, simplecoroots, PermX, cartan, inclusion, restriction, action, rank, semisimplerank

gens(W): the generators, as permutations of the root vectors. They are in the same order as the simple roots.

julia> W=coxgroup(:A,3)
A₃

julia> cartan(W)
3×3 Matrix{Int64}:
  2  -1   0
 -1   2  -1
  0  -1   2

julia> W.rootdec
12-element Vector{Vector{Int64}}:
 [1, 0, 0]
 [0, 1, 0]
 [0, 0, 1]
 [1, 1, 0]
 [0, 1, 1]
 [1, 1, 1]
 [-1, 0, 0]
 [0, -1, 0]
 [0, 0, -1]
 [-1, -1, 0]
 [0, -1, -1]
 [-1, -1, -1]

 julia> reflrep(W)
3-element Vector{Matrix{Int64}}:
 [-1 0 0; 1 1 0; 0 0 1]
 [1 1 0; 0 -1 0; 0 1 1]
 [1 0 0; 0 1 1; 0 0 -1]

Gapjm.Weyl.rootlengths — Function

rootlengths(W::FiniteCoxeterGroup) the vector of the (squared) length of the roots of W. The shortest roots in an irreducible subsystem are given the length 1. In a Weyl group the others then have length 2 (or 3 in type G₂). The matrix of the W-invariant bilinear form is given by map(i->rootlengths(W)[i]*W.cartan[i,:],1:semisimplerank(W)).

Gapjm.Weyl.rootdatum — Function

rootdatum(string or symbol,...) root datum from type

rootdatum(C::AbstractMatrix) adjoint root datum from Cartan matrix C

rootdatum(R::AbstractMatrix,CR::AbstractMatrix)

root datum from R whose rows are the simple roots on a basis of X(T) and CR whose rows are the simple coroots on a basis of Y(T)

Gapjm.Weyl.describe_involution — Function

describe_involution(W,w)

Given an involution w of a Coxeter group W, by a theorem of Richardson1982 there is a unique parabolic subgroup P of W such that that w is the longest element of P, and is central in P. The function returns I such that P=reflection_subgroup(W,I), so that w=longest(reflection_subgroup(W,I)).

julia> W=coxgroup(:A,2)
A₂

julia> w=longest(W)
(1,5)(2,4)(3,6)

julia> describe_involution(W,w)
1-element Vector{Int64}:
 3

julia> w==longest(reflection_subgroup(W,[3]))
true

For now does not work for abscox groups.

Gapjm.Weyl.badprimes — Function

badprimes(W)

Let W be a Weyl group. A prime is bad for W if it divides a coefficient of some root on the simple roots. The function badprimes returns the list of primes which are bad for W.

julia> W=coxgroup(:E,8)
E₈

julia> badprimes(W)
3-element Vector{Int64}:
 5
 2
 3

Gapjm.Weyl.standard_parabolic — Function

standard_parabolic(W::PermRootGroup, H)

Let H be a reflection subgroup of W. Returns an element w∈ W such that H^w is a standard parabolic subgroup of W (or nothing if H is not parabolic or not conjugate to a standard parabolic).

standard_parabolic(W,H)

Given a reflection subgroup H or the indices of its simple roots returns nothing if H is not parabolic, otherwise returns w such that H^w is a standard parabolic subgroup of W.

julia> W=coxgroup(:E,6)
E₆

julia> R=reflection_subgroup(W,[20,30,19,22])
E₆₍₁₉‚₁‚₉‚₂₀₎=A₄₍₃₁₂₄₎Φ₁²

julia> p=standard_parabolic(W,R)
(1,4,49,12,10)(2,54,62,3,19)(5,17,43,60,9)(6,21,34,36,20)(7,24,45,41,53)(8,65,50,15,22)(11,32,31,27,28)(13,48,46,37,40)(14,51,58,44,29)(16,23,35,33,30)(18,26,39,55,38)(42,57,70,72,56)(47,68,67,63,64)(52,59,71,69,66)

julia> reflection_subgroup(W,[20,30,19,22].^p)
E₆₍₂₄₅₆₎=A₄Φ₁²

julia> R=reflection_subgroup(W,[1,2,3,5,6,35])
E₆₍₁‚₃‚₂‚₃₅‚₅‚₆₎=A₂₍₁₃₎×A₂₍₂₆₎×A₂₍₄₅₎

julia> standard_parabolic(W,R)

Gapjm.CoxGroups.inversions — Function

inversions(W::FiniteCoxeterGroup, w::AbstractVector{<:Integer})

Given a word w=s₁…sₙ (a vector of integers) representing an element of W, returns the inversions of w in the order of the reflections: W(s₁), W(s₁,s₂,s₁), …, W(s₁,s₂,…,sₙ,sₙ₋₁,…,s₁).

julia> W=coxgroup(:A,3)
A₃

julia> inversions(W,W(1,2,1))
3-element Vector{Int64}:
 1
 2
 4

julia> inversions(W,[2,1,2])
3-element Vector{Int16}:
 2
 4
 1

inversions(W,w)

Returns the inversions of the element w of the finite Coxeter group W, that is, the list of the indices of reflections r of W such that l(rw)<l(w) where l is the Coxeter length.

julia> W=coxgroup(:A,3)
A₃

julia> inversions(W,W(1,2,1))
3-element Vector{Int64}:
 1
 2
 4

Gapjm.Weyl.with_inversions — Function

with_inversions(W,N)

W should be a finite Coxeter group and N a subset of 1:nref(W). Returns the element w of W such that N==inversions(W,w). Returns nothing if no such element exists.

julia> W=coxgroup(:A,2)
A₂

julia> map(N->with_inversions(W,N),combinations(1:nref(W)))
8-element Vector{Union{Nothing, Perm{Int16}}}:
 ()
 (1,4)(2,3)(5,6)
 (1,3)(2,5)(4,6)
 nothing
 nothing
 (1,6,2)(3,5,4)
 (1,2,6)(3,4,5)
 (1,5)(2,4)(3,6)

Gapjm.Weyl.torus — Function

torus(m::Matrix)

m should be an integral matrix of finite order. The function returns the coset T of the trivial Coxeter group such that T.F==m. This corresponds to an algebraic torus 𝐓 of rank size(m,1), with an isogeny which acts by m on X(𝐓).

julia> torus([0 -1;1 -1])
Φ₃

torus(W,i)

This returns the torus twisted by a representative of the i-th conjugacy class of W. This is the same as twistings(W,Int[])[i].

julia> W=coxgroup(:A,3)
A₃

julia> twistings(W,Int[])
5-element Vector{Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}}:
 A₃₍₎=Φ₁³
 A₃₍₎=Φ₁²Φ₂
 A₃₍₎=Φ₁Φ₂²
 A₃₍₎=Φ₁Φ₃
 A₃₍₎=Φ₂Φ₄

julia> torus(W,2)
A₃₍₎=Φ₁²Φ₂

julia> WF=spets(W,Perm(1,3))
²A₃

julia> twistings(WF,Int[])
5-element Vector{Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}}:
 ²A₃₍₎=Φ₂³
 ²A₃₍₎=Φ₁Φ₂²
 ²A₃₍₎=Φ₁²Φ₂
 ²A₃₍₎=Φ₂Φ₆
 ²A₃₍₎=Φ₁Φ₄

julia> torus(WF,2)
²A₃₍₎=Φ₁Φ₂²

torus(rank)

This function returns the object corresponding to the notion of a torus of dimension rank, a Coxeter group of semisimple rank 0 and given rank. This corresponds to a split torus; the extension to Coxeter cosets is more useful.

julia> torus(3)
Φ₁³

Gapjm.Weyl.istorus — Function

istorus(W) whether W is a torus

Gapjm.Semisimple.SubTorus — Type

SubTorus(W,Y::Matrix{<:Integer})

The function returns the subtorus 𝐒 of the maximal torus 𝐓 of the reductive group represented by the Weyl group W such that Y(𝐒) is the (pure) sublattice of Y(𝐓) generated by the (integral) vectors Y. A basis of Y(𝐒) adapted to Y(𝐓) is computed and stored in the field S.gens of the returned SubTorus struct. Here, adapted means that there is a set of integral vectors, stored in S.complement, such that M=vcat(S.gens,S.complement) is a basis of Y(𝐓) (equivalently M∈GL(Z^{rank(W)}). An error is raised if Y does not define a pure sublattice.

julia> W=coxgroup(:A,4)
A₄

julia> SubTorus(W,[1 2 3 4;2 3 4 1;3 4 1 1])
SubTorus(A₄,[1 0 3 -13; 0 1 2 7; 0 0 4 -3])

julia> SubTorus(W,[1 2 3 4;2 3 4 1;3 4 1 2])
ERROR: not a pure sublattice
Stacktrace:
 [1] error(::String) at ./error.jl:33
 [2] Gapjm.Weyl.SubTorus(::FiniteCoxeterGroup{Perm{Int16},Int64}, ::Matrix{Int64}) at /home/jmichel/julia/Gapjm.jl/src/Weyl.jl:1082
 [3] top-level scope at REPL[25]:1

Gapjm.Weyl.relative_group — Function

relative_group(W::FiniteCoxeterGroup,J)

J should be a if distinguished subset of S==eachindex(gens(W)), that is if for s∈ S-J we set v(s,J)=w₀^{J∪ s}w₀ᴶ then J is stable by all v(s,J). Then $R=N_W(W_J)/W_J$ is a Coxeter group with Coxeter system the v(s,J). The program returns R in its reflection representation on $X(ZL_J/ZG)$. (according to Lusztig1976, the images of the roots of W in $X(ZL_J/ZG)$ form a root system).

R.prop has the fields: :relativeIndices=setdiff(S,J) :parentMap= the list of v(s,J) :MappingFromNormalizer maps J-reduced elements of $N_W(W_J)$ to elements of R