d-Harish-Chandra series

Chevie.dSeries

Chevie.dSeries.Series

Chevie.Symbols.ennola

Chevie.Uch.cuspidal_data

Chevie.dSeries Рђћ Module

d-Harish-Chandra series describe unipotent l-blocks of a finite reductive group $­Юљє(­Юћй_q)$ for $l|╬д_d(q)$ (at least, when l is not too small which means mostly not a bad prime for ­Юљє). Some of the facts stated below are still partly conjectural, we do not try to distinguish precisely what has been established and what is still conjectural.

If (­ЮљІ,╬╗) is a d-cuspidal pair then the constituents of the Lusztig induced $R_­ЮљІ^­Юљє(╬╗)$ are called a d-Harish-Chandra series; they form the unipotent part of an l-block of $­Юљє^F$ (Brou├Е et al., 1993; 5.24). It is conjectured (Brou├Е and Michel, 1997; 5.7) (and proven in some cases) that the $­Юљє^F$-endomorphism algebra of the l-adic cohomology of the variety ­ЮљЌ which defines the Lusztig induction is a d-cyclotomic Hecke algebra $H_­Юљє(­ЮљІ,╬╗)$ for the group $W_­Юљє(­ЮљІ,╬╗):=N_­Юљє(­ЮљІ,╬╗)/­ЮљІ$, which is a complex reflection group РђЊ- here d-cyclotomic means that the parameters of $H_­Юљє(­ЮљІ,╬╗)$ are monomials in q and that $H_­Юљє(­ЮљІ,╬╗)$ specializes to the algebra of $W_­Юљє(­ЮљІ,╬╗)$ for $qРєд╬Х_d$.

It follows that the decomposition of the Lusztig induction is of the form $R_­ЮљІ^­Юљє(╬╗)=РѕЉ_{¤ЋРѕѕIrr(W_­Юљє(­ЮљІ,╬╗))}(-1)^{nрхЕ} ¤Ћ(1)╬│рхЕ,$ where ╬│рхЕ is a unipotent character of ­Юљє^F attached to ¤Ћ and where nрхЕ is the degree $H^{nрхЕ}_c(­ЮљЌ)$ where ╬│рхЕ occurss; and further for any ¤Ћ we have $R_­ЮљІ^­Юљє(╬╗)(1)= (-1)^{nрхЕ} ╬│рхЕ(1)SрхЕ$ where SрхЕ is the Schur element of the character of $H_­Юљє(­ЮљІ,╬╗)$ which deforms to ¤Ћ. The function Series allows to explore a d-Harish-Chandra series.

julia> W=rootdatum("3D4")
┬│DРѓё

julia> l=cuspidal_data(W,3)
2-element Vector{@NamedTuple{levi::Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}, cuspidal::Int64, d::Root1}}:
 (levi = ┬│DРѓё, cuspidal = 8, d = ╬ХРѓЃ)
 (levi = ┬│DРѓёРѓЇРѓј=╬дРѓЃ┬▓, cuspidal = 1, d = ╬ХРѓЃ)

julia> Series(W,l[2]...)
╬ХРѓЃ-series R^┬│DРѓё_{┬│DРѓёРѓЇРѓј=╬дРѓЃ┬▓}(╬╗==Id)  H_G(L,╬╗)==hecke(GРѓё,Vector{Mvp{Cyc{Int64}, Int64}}[[╬ХРѓЃq┬▓, ╬ХРѓЃ, ╬ХРѓЃq]])
РћїРћђРћгРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ
Рћѓ Рћѓ    ╬│рхЕ    ¤є  ╬х family nРЂ░Рћѓ
РћюРћђРћ╝РћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћц
Рћѓ1Рћѓ  ¤єРѓЂРђџРѓђ ¤єРѓЂРђџРѓђ  1         1Рћѓ
Рћѓ2Рћѓ  ¤єРѓЂРђџРѓє ¤єРѓЂРђџРѓё  1         2Рћѓ
Рћѓ3Рћѓ  ¤єРѓѓРђџРѓѓ ¤єРѓЂРђџРѓѕ -1         5Рћѓ
Рћѓ6Рћѓ ¤єРђ│РѓЂРђџРѓЃ ¤єРѓѓРђџРѓЁ  1         4Рћѓ
Рћѓ5Рћѓ ¤єРђ▓РѓЂРђџРѓЃ ¤єРѓѓРђџРѓЃ -1         3Рћѓ
Рћѓ7Рћѓ  ¤єРѓѓРђџРѓЂ ¤єРѓѓРђџРѓЂ -1         5Рћѓ
Рћѓ4Рћѓ┬│DРѓё[1] ¤єРѓЃРђџРѓѓ  1         5Рћѓ
РћћРћђРћ┤РћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў

Above we explore the 3-series corresponding to $R_­ЮљЊ^­Юљє(Id)$ where ­Юљє is the triality group and ­ЮљЊ is the torus of type (q┬▓+q+1)┬▓. The group $W_­Юљє(­ЮљЊ)$ is the complex reflection group GРѓё. The displays shows in the column '╬│рхЕ' the name of the unipotent characters constituents of $R_­ЮљЊ^­Юљє(Id)$, and in the first column the number of these characters in the list of unipotent characters. In the column '¤є' the name of the character of $W_­Юљє(­ЮљЊ)$ corresponding to the unipotent character ╬│рхЕ is shown; in the column '╬х' we show the sign $(-1)^{nрхЕ}$. Finally in the last column we show in which family of unipotent characters is ╬│рхЕ.

The theory of d-Harish-Chandra series can be generalized to spetsial complex reflection groups using some axioms. We show below an example.

julia> W=complex_reflection_group(4)
GРѓё

julia> l=cuspidal_data(W,3)
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = GРѓё, cuspidal = 3, d = ╬ХРѓЃ)
 (levi = GРѓё, cuspidal = 6, d = ╬ХРѓЃ)
 (levi = GРѓё, cuspidal = 7, d = ╬ХРѓЃ)
 (levi = GРѓё, cuspidal = 10, d = ╬ХРѓЃ)
 (levi = GРѓёРѓЇРѓј=╬дРѓЂ╬дРђ▓РѓЃ, cuspidal = 1, d = ╬ХРѓЃ)

julia> Series(W,l[5]...)
╬ХРѓЃ-series R^GРѓё_{GРѓёРѓЇРѓј=╬дРѓЂ╬дРђ▓РѓЃ}(╬╗==Id)  W_G(L,╬╗)==ZРѓє
РћїРћђРћгРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ
Рћѓ Рћѓ   ╬│рхЕ ¤є(mod 3)  ╬х parameter family nРЂ░Рћѓ
РћюРћђРћ╝РћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћц
Рћѓ1Рћѓ ¤єРѓЂРђџРѓђ        1  1      ╬ХРѓЃq┬▓         1Рћѓ
Рћѓ5Рћѓ ¤єРѓѓРђџРѓЃ       ╬ХРѓє  1      -╬ХРѓЃq         2Рћѓ
Рћѓ2Рћѓ ¤єРѓЂРђџРѓё       ╬ХРѓЃ -1        ╬ХРѓЃ         4Рћѓ
Рћѓ8Рћѓ ZРѓЃ:2       -1 -1     -╬ХРѓЃ┬▓q         2Рћѓ
Рћѓ9РћѓZРѓЃ:11      ╬ХРѓЃ┬▓ -1       ╬ХРѓЃ┬▓         4Рћѓ
Рћѓ4Рћѓ ¤єРѓѓРђџРѓЁ      ╬ХРѓєРЂх -1       -╬ХРѓЃ         4Рћѓ
РћћРћђРћ┤РћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў

Above we explore the 3-series corresponding to the trivial character of the torus of type (q-1)(q-╬ХРѓЃ). For cyclic groups $W_­Юљє(­ЮљІ,╬╗)$ we display the parameters in the table since they are associated to characters of $W_­Юљє(­ЮљІ,╬╗)$. Finally the mention (mod 3) which appears in the ¤є column means that in this case the axioms leave an ambiguity in the correspondence between unipotent characters ╬│рхЕ and characters ¤Ћ (as well as with parameters): the correspondence is known only up to a translation by 3 (in this case, the same as a global multiplication of all ¤Ћ by -1).

Finally, we should note that if the reflection group or coset W is not defined over the integers, what counts is not cyclotomic polynomials but factors of them over the field of definition of W. In this case, one should not give as argument an integer d representing $╬Х_d$ but specify a root of unity. For instance, in the above case we get a different answer with:

julia> cuspidal_data(W,E(3,2))
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = GРѓё, cuspidal = 2, d = ╬ХРѓЃ┬▓)
 (levi = GРѓё, cuspidal = 5, d = ╬ХРѓЃ┬▓)
 (levi = GРѓё, cuspidal = 7, d = ╬ХРѓЃ┬▓)
 (levi = GРѓё, cuspidal = 10, d = ╬ХРѓЃ┬▓)
 (levi = GРѓёРѓЇРѓј=╬дРѓЂ╬дРђ│РѓЃ, cuspidal = 1, d = ╬ХРѓЃ┬▓)

Chevie.Uch.cuspidal_data Рђћ Function

cuspidal_data(W[,d[,ad]];proper=false,all=false)

returns named tuples (levi=LF,cuspidal=╬╗,d=d) where LF is a d-split Levi (with d-center of dimension ad if ad is given) and ╬╗ is a d-cuspidal character of LF. If d=1 this returns ordinary cuspidal characters. The character ╬╗ is given as its index in the list of unipotent characters. If d was given as an integer, it is returned as a Root1 representing E(d).

If the keyword proper=true is given, only the data where LF!=W (or equivalently ad>0) are returned.

If d is omitted, data for all d orders of eigenvalues of elements of W is returned. If in addition the keyword argument all=true is given, data for all eigenvalues of elements of W is returned.

julia> cuspidal_data(coxgroup(:F,4),1)
9-element Vector{@NamedTuple{levi::Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}, cuspidal::Int64, d::Root1}}:
 (levi = FРѓё, cuspidal = 31, d = 1)
 (levi = FРѓё, cuspidal = 32, d = 1)
 (levi = FРѓё, cuspidal = 33, d = 1)
 (levi = FРѓё, cuspidal = 34, d = 1)
 (levi = FРѓё, cuspidal = 35, d = 1)
 (levi = FРѓё, cuspidal = 36, d = 1)
 (levi = FРѓё, cuspidal = 37, d = 1)
 (levi = FРѓёРѓЇРѓѓРѓЃРѓј=BРѓѓРѓЇРѓѓРѓЂРѓј╬дРѓЂ┬▓, cuspidal = 6, d = 1)
 (levi = FРѓёРѓЇРѓј=╬дРѓЂРЂ┤, cuspidal = 1, d = 1)

julia> cuspidal_data(complex_reflection_group(4),3)
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = GРѓё, cuspidal = 3, d = ╬ХРѓЃ)
 (levi = GРѓё, cuspidal = 6, d = ╬ХРѓЃ)
 (levi = GРѓё, cuspidal = 7, d = ╬ХРѓЃ)
 (levi = GРѓё, cuspidal = 10, d = ╬ХРѓЃ)
 (levi = GРѓёРѓЇРѓј=╬дРѓЂ╬дРђ▓РѓЃ, cuspidal = 1, d = ╬ХРѓЃ)

Chevie.dSeries.Series Рђћ Type

Series(W, L, cuspidal, d)

If the reflection coset or group W corresponds to the algebraic group ­Юљє and cuspidal is the index of a d-cuspidal unipotent character ╬╗ of the d-split Levi ­ЮљІ, constructs the d-series $R_­ЮљІ^­Юљє(╬╗)$. The result s has the following properties and functions:

s.spets: the reflection group or coset W.

s.levi: the subcoset L.

s.cuspidal: the index of ╬╗ in UnipotentCharacters(L).

s.d: the value of d (a Root1).

relative_group(s): the group $W_­Юљє(­ЮљІ,╬╗)$.

s.e: the order of $W_­Юљє(­ЮљІ,╬╗)$.

dSeries.RLG(s): the UnipotentCharacter given by $R_­ЮљІ^­Юљє(╬╗)$.

signs(s): for each character ¤є of relative_group(s) the sign $(-1)^{n_¤є}$ in RLG(s) of the corresponding constituent ╬│рхЕ of RLG(s).

degree(s): the generic degree of RLG(s), as a CycPol.

charnumbers(s): the indices in UnipotentCharacters(W) of the

constituents of RLG(s).

hecke(s): the hecke algebra $H_­Юљє(­ЮљІ,╬╗)$.

The function Series has another form:

Series(W [,d [,ad]];k...)

where it returns a vector of Series corresponding to the cuspidal_data described by the arguments and the keywords.

julia> W=complex_reflection_group(4)
GРѓё

julia> Series(W,3;proper=true)
1-element Vector{Series}:
 ╬ХРѓЃ-series R^GРѓё_{GРѓёРѓЇРѓј=╬дРѓЂ╬дРђ▓РѓЃ}(╬╗==Id)  W_G(L,╬╗)==ZРѓє

julia> s=Series(W,3,1)[1]
╬ХРѓЃ-series R^GРѓё_{GРѓёРѓЇРѓј=╬дРѓЂ╬дРђ▓РѓЃ}(╬╗==Id)  W_G(L,╬╗)==ZРѓє
РћїРћђРћгРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћљ
Рћѓ Рћѓ   ╬│рхЕ ¤є(mod 3)  ╬х parameter family nРЂ░Рћѓ
РћюРћђРћ╝РћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћц
Рћѓ1Рћѓ ¤єРѓЂРђџРѓђ        1  1      ╬ХРѓЃq┬▓         1Рћѓ
Рћѓ5Рћѓ ¤єРѓѓРђџРѓЃ       ╬ХРѓє  1      -╬ХРѓЃq         2Рћѓ
Рћѓ2Рћѓ ¤єРѓЂРђџРѓё       ╬ХРѓЃ -1        ╬ХРѓЃ         4Рћѓ
Рћѓ8Рћѓ ZРѓЃ:2       -1 -1     -╬ХРѓЃ┬▓q         2Рћѓ
Рћѓ9РћѓZРѓЃ:11      ╬ХРѓЃ┬▓ -1       ╬ХРѓЃ┬▓         4Рћѓ
Рћѓ4Рћѓ ¤єРѓѓРђџРѓЁ      ╬ХРѓєРЂх -1       -╬ХРѓЃ         4Рћѓ
РћћРћђРћ┤РћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћђРћў

julia> s.spets
GРѓё

julia> s.levi
GРѓёРѓЇРѓј=╬дРѓЂ╬дРђ▓РѓЃ

julia> s.cuspidal
1

julia> s.d
Root1: ╬ХРѓЃ

julia> hecke(s)
hecke(GРѓєРђџРѓЂРђџРѓЂ,Vector{Mvp{Cyc{Int64}, Int64}}[[╬ХРѓЃq┬▓, -╬ХРѓЃq, ╬ХРѓЃ, -╬ХРѓЃ┬▓q, ╬ХРѓЃ┬▓, -╬ХРѓЃ]])

julia> degree(s)
╬ХРѓЃ╬дРѓЂ╬дРѓѓ┬▓╬дРђ│РѓЃ╬дРѓё╬дРѓє

julia> dSeries.RLG(s)
[GРѓё]:<¤єРѓЂРђџРѓђ>-<¤єРѓЂРђџРѓё>-<¤єРѓѓРђџРѓЁ>+<¤єРѓѓРђџРѓЃ>-<ZРѓЃ:2>-<ZРѓЃ:11>

julia> charnumbers(s)
6-element Vector{Int64}:
 1
 5
 2
 8
 9
 4

julia> signs(s)
6-element Vector{Int64}:
  1
  1
 -1
 -1
 -1
 -1

julia> relative_group(s)
GРѓєРђџРѓЂРђџРѓЂ

Chevie.Symbols.ennola Рђћ Function

ennola(S::CharSymbol)

Ennola of e-symbol S (of content 1 or 0) The order of Ennola (order of center of reflection group) is computed automatically: it is e for content 1 and gcd(e,rank(S)) for content 0.

ennola(W[,z=E(gcd(degrees(W)))])

Let W be an irreducible spetsial reflection group or coset, and z the generator of the center of W, viewed as the root of unity E(gcd(degrees(W)))]). Let ­ЮћЙ be the spets attached to W. A property checked case-by case is that, for a unipotent character ╬│ of ­ЮћЙ with polynomial generic degree deg ╬│(q) , deg ╬│(zq) is equal to ┬▒deg ╬│'(q) for another unipotent character ╬│'; ┬▒╬│' is called the Ennola transform of ╬│. For W a Weyl group, the spets ­ЮћЙ is a finite reductive group, in which case z=-1 if -1 is in W and z=1 otherwise. The function returns the signed permutation e done by ennola on the unipotent degrees (as an SPerm of 1:length(UnipotentCharacters(W))).

The SPerm e is not uniquely determined by the degrees since two degrees may be equal, but is uniquely determined by some additional axioms that we do not recall here (they include a description of the Ennola-permutation in terms of the Z-based rings attached to each Lusztig family).

If a second argument z is given, it should be a power of the default z and the corresponding power of e is returned.

julia> ennola(rootdatum("3D4"))
SPerm{Int64}: (3,-4)(5,-5)(6,-6)(7,-8)

julia> ennola(complex_reflection_group(14))
SPerm{Int64}: (2,43,-14,16,41,34)(3,35,40,18,-11,42)(4,-37,25,-17,-26,-36)(5,-6,-79)(7,-7)(8,-74)(9,-73)(10,-52,13,31,-50,29)(12,53,15,32,-51,-30)(19,71,70,21,67,68,20,69,72)(22,-39,27,-33,-28,-38)(23,24,-66,-23,-24,66)(44,46,49,-44,-46,-49)(45,48,47,-45,-48,-47)(54,-63,-55,-57,62,-56)(58,-65,-59,-61,64,-60)(75,-77)(76,-76)(78,-78)

The above example shows that it may happen that the order of z-Ennola (here 18) is greater than the order of z (here 6); this is related to the presence of irrationalities qРЁЊ in the character table of the spetsial Hecke algebra of W.

For a non-irreducible group, z-ennola is defined if z can be considered an element of the the centre of each irreducible component.