Eigenspaces

Chevie.Eigenspaces

Chevie.Eigenspaces.eigenspace_projector

Chevie.Eigenspaces.position_regular_class

Chevie.Eigenspaces.regular_eigenvalues

Chevie.Eigenspaces.relative_degrees

Chevie.Eigenspaces.relative_group

Chevie.Eigenspaces.split_levis

Chevie.Uch.cuspidal

Chevie.Eigenspaces — Module

Eigenspaces and d-Harish-Chandra series

Let Wϕ be a reflection coset on a vector space V; that is Φ∈GL(V) normalizes the reflection group W. Let Lwϕ be a reflection subcoset; that is L is a parabolic subgroup of W (the fixator of a subspace of V) and w∈ W is such that wΦ normalizes L. There are several interesting cases where the relative_group $N_W(Lwϕ)/L$, or a subgroup of it normalizing some further data attached to L, is itself a reflection group.

A first example is the case where ϕ=1 and w=1, W is the Weyl group of a finite reductive group $𝐆^F$ and the Levi subgroup $𝐋^F$ corresponding to L has a cuspidal unipotent character. Then $N_W(L)/L$ is a Coxeter group acting on the space X(Z𝐋)⊗ℝ. A combinatorial characterization of such parabolic subgroups of Coxeter groups is that they are normalized by the longest element of larger parabolic subgroups, see (Lusztig, 1976; 5.7.1).

A second example is when L is trivial and wϕ is a ζ-regular element, that is the ζ-eigenspace $V_ζ$ of wϕ contains a vector outside all the reflecting hyperplanes of W. Then $N_W(Lwϕ)/L=C_W(wϕ)$ is a reflection group in its action on $V_ζ$.

A similar but more general example is when $V_ζ$ is the ζ-eigenspace of some element of the reflection coset Wϕ, and is of maximal dimension among such ζ-eigenspaces. Then the set of elements of Wϕ which act by ζ on $V_ζ$ is a certain subcoset Lwϕ, and $N_W(Lwϕ)/L$ is a reflection group in its action on $V_ζ$, see (Lehrer and Springer, 1999; 2.5).

Finally, a still more general example, but which only occurs for Weyl groups or Spetsial reflection groups, is when 𝐋 is a ζ-split Levi subgroup (which means that the corresponding subcoset Lwϕ is formed of all the elements which act by ζ on some subspace V_ζ of V), and λ is a d-cuspidal unipotent character of 𝐋 (which means that the multiplicity of ζ as a root of the degree of λ is the same as the multiplicity of ζ as a root of the generic order of the semi-simple part of 𝐆); then $N_W(Lwϕ,λ)/L$ is a complex reflection group in its action on V_ζ.

Further, in the above cases the relative group describes the decomposition of a Lusztig induction.

When $𝐆^F$ is a finite reductive group, and λ a cuspidal unipotent character of the Levi subgroup $𝐋^F$, then the $𝐆^F$-endomorphism algebra of the Harish-Chandra induced representation $R_𝐋^𝐆(λ)$ is a Hecke algebra attached to the group $N_W(L)/L$, thus the dimension of the characters of this group describe the multiplicities in the Harish-Chandra induced.

Similarly, when 𝐋 is a ζ-split Levi subgroup, and λ is a d-cuspidal unipotent character of 𝐋 then (conjecturally) the $𝐆^F$-endomorphism algebra of the Lusztig induced $R_𝐋^𝐆(λ)$ is a cyclotomic Hecke algebra for to the group $N_W(Lwϕ,λ)/L$. The constituents of $R_𝐋^𝐆(λ)$ are called a ζ-Harish-Chandra series. In the case of rational groups or cosets, corresponding to finite reductive groups, the conjugacy class of Lwϕ depends only on the order d of ζ, so one also talks of d-Harish-Chandra series. These series correspond to ℓ-blocks where l is a prime divisor of Φ_d(q) which does not divide any other cyclotomic factor of the order of $𝐆^F$.

The functions relative_degrees, regular_eigenvalues, eigenspace_projector, position_regular_class, split_levis, cuspidal in this module and the functions in the module dSeries allow to explore these situations.

Chevie.Eigenspaces.relative_degrees — Function

relative_degrees(WF,ζ::Root1=1)

Let WF be a reflection group or a reflection coset and ζ be a root of unity. Then if $V_ζ$ is the ζ-eigenspace of some element of WF, and is of maximal dimension among such ζ-eigenspaces (and if WF is a coset W is the group of WF) then $N_W(V_ζ)/C_W(V_ζ)$ is a reflection group in its action on $V_ζ$. The function relative_degrees returns the reflection degrees of this complex reflection group, which are a subset of those of W. These degrees are computed by an invariant-theoretic formula: if (d₁,ε₁),…,(dₙ,εₙ) are the generalized degrees of WF (see degrees) they are the dᵢ such that ζ^{dᵢ}=εᵢ.

The eigenvalue ζ can also be specified by giving an integer d (which then specifies ζ=E(d)) or a fraction a//b which then specifies ζ=E(b,a). If omitted, ζ is taken to be 1.

julia> W=coxgroup(:E,8)
E₈

julia> relative_degrees(W,4) # the degrees of G₃₂
4-element Vector{Int64}:
  8
 12
 20
 24

Chevie.Eigenspaces.regular_eigenvalues — Function

regular_eigenvalues(W)

Let W be a reflection group or a reflection coset. A root of unity ζ is a regular eigenvalue for W if some element of W has a ζ-eigenvector which lies outside of the reflecting hyperplanes. The function returns the list of regular eigenvalues for W.

julia> regular_eigenvalues(coxgroup(:G,2))
6-element Vector{Root1}:
   1
  -1
  ζ₃
 ζ₃²
  ζ₆
 ζ₆⁵

julia> W=complex_reflection_group(6)
G₆

julia> L=twistings(W,[2])[4]
G₆₍₂₎=G₃‚₁‚₁[ζ₄]Φ′₄

julia> regular_eigenvalues(L)
3-element Vector{Root1}:
    ζ₄
  ζ₁₂⁷
 ζ₁₂¹¹

Chevie.Eigenspaces.eigenspace_projector — Function

eigenspace_projector(WF,w,ζ::Root1=1)

Let WF be a reflection group or a reflection coset, let w be an element of WF and let ζ be a root of unity. The function returns the unique w-invariant projector on the ζ-eigenspace of w.

The eigenvalue ζ can also be specified by giving an integer d (which then specifies ζ=E(d)) or a fraction a//b which then specifies ζ=E(b,a). If omitted, ζ is taken to be 1.

julia> W=coxgroup(:A,3)
A₃

julia> w=W(1:3...)
(1,12,3,2)(4,11,10,5)(6,9,8,7)

julia> p=eigenspace_projector(W,w,1//4)
3×3 Matrix{Cyc{Rational{Int64}}}:
  (1+ζ₄)/4   ζ₄/2  (-1+ζ₄)/4
  (1-ζ₄)/4   1//2   (1+ζ₄)/4
 (-1-ζ₄)/4  -ζ₄/2   (1-ζ₄)/4

julia> GenLinearAlgebra.rank(p)
1

Chevie.Eigenspaces.position_regular_class — Function

position_regular_class(WF,ζ::Root1=1)

Let WF be a reflection group or a reflection coset and ζ be a root of unity such that some element of WF has a non-trivial ζ-eigenspace. The function returns the index of a conjugacy class of WF whose ζ-eigenspace is maximal (amongst all ζ-eigenspaces of elements of W). If no element of WF has a non-trivial ζ-eigenspace the function returns nothing.

The eigenvalue ζ can also be specified by giving an integer d (which then specifies ζ=E(d)) or a fraction a//b which then specifies ζ=E(b,a). If omitted, ζ is taken to be 1.

julia> W=coxgroup(:E,8)
E₈

julia> position_regular_class(W,30)
65

julia> W=complex_reflection_group(6)
G₆

julia> L=twistings(W,[2])[4]
G₆₍₂₎=G₃‚₁‚₁[ζ₄]Φ′₄

julia> position_regular_class(L,7//12)
2

Chevie.Eigenspaces.split_levis — Function

split_levis(WF,ζ::Root1=1[,ad])

Let WF be a reflection group or a reflection coset. If W is a reflection group it is treated as the trivial coset 'Spets(W)'. A Levi is a subcoset of the form W₁F₁ where W₁ is a parabolic subgroup of W, that is the centralizer of some subspace of V, and F₁∈ WF normalizes W₁.

Let ζ be a root of unity. split_levis returns a list of representatives of conjugacy classes of ζ-split Levis of W. A ζ-split Levi is a subcoset of WF formed of all the elements which act by ζ on a given subspace V_ζ. If the additional argument ad is given, it returns only those subcosets such that the common ζ-eigenspace of their elements is of dimension ad. These notions make sense and thus are implemented for any complex reflection group.

In terms of algebraic groups, an F-stable Levi subgroup of the reductive group 𝐆 is ζ-split if and only if it is the centralizer of the Φ-part of its center, where Φ is a cyclotomic polynomial with ζ as a root. When ζ=1, we get the notion of a split Levi, which is the same as a Levi sugroup of an F-stable parabolic subgroup of 𝐆.

The eigenvalue ζ can also be specified by giving an integer d (which then specifies ζ=E(d)) or a fraction a//b which then specifies ζ=E(b,a). If omitted, ζ is taken to be 1.

julia> W=coxgroup(:A,3)
A₃

julia> split_levis(W,4)
2-element Vector{Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}}:
 A₃
 A₃₍₎=Φ₂Φ₄

julia> W=spets(coxgroup(:D,4),Perm(1,2,4))
³D₄

julia> split_levis(W,3)
3-element Vector{Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}}:
 ³D₄
 ³D₄₍₁₃₎=A₂Φ₃
 ³D₄₍₎=Φ₃²

julia> W=coxgroup(:E,8)
E₈

julia> split_levis(W,4,2)
3-element Vector{Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}}:
 E₈₍₃₄₂₅₎=D₄₍₁₃₂₄₎Φ₄²
 E₈₍₅₇₂₃₎=(A₁A₁)×(A₁A₁)Φ₄²
 E₈₍₃₅₆₁₎=²(A₂A₂)₍₁₄₂₃₎Φ₄²

julia> split_levis(complex_reflection_group(5))
4-element Vector{Spets{PRSG{Cyc{Rational{Int64}}, Int16}}}:
 G₅
 G₅₍₁₎=G₃‚₁‚₁Φ₁
 G₅₍₂₎=G₃‚₁‚₁Φ₁
 G₅₍₎=Φ₁²

Chevie.Uch.cuspidal — Function

cuspidal(uc::UnipotentCharacters[,e])

A unipotent character γ of a finite reductive group 𝐆 is e-cuspidal if its Lusztig restriction to any proper e-split Levi is zero. When e==1 (the default when e is omitted) we recover the usual notion of cuspidal character. Equivalently the Φₑ-part of the generic degree of γ is equal to the Φₑ-part of the generic order of the adjoint group of 𝐆. This makes sense for any Spetsial complex reflection group and is implemented for them.

The function returns the list of indices of unipotent characters which are e-cuspidal.

julia> W=coxgroup(:D,4)
D₄

julia> cuspidal(UnipotentCharacters(W))
1-element Vector{Int64}:
 14

julia> cuspidal(UnipotentCharacters(W),6)
8-element Vector{Int64}:
  1
  2
  6
  7
  8
  9
 10
 12

julia> cuspidal(UnipotentCharacters(complex_reflection_group(4)),3)
4-element Vector{Int64}:
  3
  6
  7
 10

Chevie.Eigenspaces.relative_group — Method

relative_group(W::FiniteCoxeterGroup,J)

J should be a if distinguished subset of S==eachindex(gens(W)), that is if for s∈ S-J we set v(s,J)=longest(W,J)*longest(W,vcat(s,J)) then J is stable by all v(s,J). Then we have an isomorphism $N_W(W_J)=W_J⋊ N$ where N is the group generated by the v(s,J), which form a Coxeter system for N. The v(s,J) are the J-reduced elements of $N_W(W_J)$, and N is isomorphic to the quotient $R=N_W(W_J)/W_J$, which has a natural reflection representation on $X(ZL_J/ZG)$, using that by (Lusztig, 1976), the images of the roots of W in $X(ZL_J)$ form a root system.

relative_group returns R as a reflection group on X(ZL_J/ZG), with some extra attributes reflecting its origin

R.relative_indices=setdiff(S,J) in an order corresponding to that of the generators of R.

R.toparent= the list of v(s,J) for .relative_indices; defines an isomorphism R→ N.

R.fromparent(w) is a function mapping w∈ N to its image in R by the isomorphism inverse to .toparent.

julia> W=coxgroup(:E,8)
E₈

julia> relative_group(W,2:5)
F₄₍₄₃₂₁₎