Decomposition Matrices

These functions interface with Thomas Breuer's package GenericDecMats.

Chevie.InducedDecompositionMatrix

Chevie.Chars.decomposition_matrix

GenericDecMats.generic_decomposition_matrix

Chevie.Chars.decomposition_matrix — Function

decomposition_matrix(W,p)

This provides an interface to some decomposition matrices for Weyl groups available in the Chevie library: those for W(E₆), W(E₇), W(E₈) for p dividing |W|.

GenericDecMats.generic_decomposition_matrix — Function

generic_decomposition_matrix(W,d)

This function obtains the $Φ_d$-decomposition matrix for the reductive group specified by the Coxeter group or coset W from the package GenericDecMats.

julia> W=rootdatum(:psu,5)
psu₅

julia> generic_decomposition_matrix(W,13)
!!! Φ-decomposition matrices available for ²A₄: Φ₁₀ Φ₂ Φ₄ Φ₆

julia> generic_decomposition_matrix(W,10)
Φ₁₀-decomposition matrix for psu₅
┌──────┬─────────────────────────┐
│      │ps 21 ps ps ps 2111 11111│
├──────┼─────────────────────────┤
│2.    │ 1  .  .  .  .    .     .│
│²A₂:2 │ .  1  .  .  .    .     .│
│11.   │ .  .  1  .  .    .     .│
│1.1   │ 1  .  .  1  .    .     .│
│.2    │ .  .  .  .  1    .     .│
│²A₂:11│ .  1  .  .  .    1     .│
│.11   │ .  .  .  1  .    .     1│
└──────┴─────────────────────────┘

The matrix itself is stored in the field .scalar of the returned struct.

Chevie.InducedDecompositionMatrix — Type

InducedDecompositionMatrix(R,W,d)

returns the induced from the Levi L to the reductive group W of the generic $Φ_d$-decomposition matrix of L.

julia> W=rootdatum(:psu,6)
psu₆

julia> L=reflection_subgroup(W,[1,2,4,5])
psu₆₍₁₂₄₅₎=(A₂A₂)₍₁₂₄₃₎Φ₁

julia> InducedDecompositionMatrix(L,W,6)
Induced Φ₆-decomposition matrix from psu₆₍₁₂₄₅₎=(A₂A₂)₍₁₂₄₃₎Φ₁ to psu₆

┌────┬────────┐
│    │ps ps A₂│
├────┼────────┤
│²A₅ │ .  .  .│
│.3  │ 1  .  .│
│3.  │ 1  .  .│
│.21 │ 1  1  .│
│1.2 │ 2  1  .│
│21. │ 1  1  .│
│2.1 │ 2  1  .│
│.111│ .  1  1│
│111.│ .  1  1│
│1.11│ 1  2  1│
│11.1│ 1  2  1│
└────┴────────┘

The matrix itself is stored in the field .scalar of the returned struct.