Non-connected reductive groups

Chevie.Sscoset

Chevie.Semisimple.isisolated

Chevie.Semisimple.quasi_isolated_reps

PermGroups.Groups.centralizer

Chevie.Sscoset — Module

Quasi-Semisimple elements of non-connected reductive groups

We also use Coxeter cosets to represented non-connected reductive groups of the form 𝐆 ⋊ σ where 𝐆 is a connected reductive group and σ is an algebraic automorphism of 𝐆; more specifically to represent the coset 𝐆 .σ. We may always choose σ∈𝐆⋅σ quasi-semisimple, which means that σ preserves a pair 𝐓 ⊂ 𝐁 of a maximal torus and a Borel subgroup of 𝐆, and further quasi-central, which means that the Weyl group of $C_𝐆(σ)$ is $W^σ$. Then σ defines an automorphism F₀ of the root datum (X(𝐓), Φ, Y(𝐓), Φᵛ), thus a Coxeter coset. We refer to (Digne and Michel, 2018) for details.

We have extended the functions for semi-simple elements to work with quasi-semisimple elements tσ∈ 𝐓 ⋅σ. Here, as in (Digne and Michel, 2018), σ is a quasi-central automorphism uniquely defined by a diagram automorphism of (W,S), taking σ symplectic in type A₂ₙ.

Here are some examples:

julia> WF=rootdatum(:u,6)
u₆

We can see WF as the coset GL₆⋅σ where σ is the composed of transpose, inverse and the longest element of W.

julia> l=quasi_isolated_reps(WF)
4-element Vector{SemisimpleElement{Root1}}:
 <1,1,1,1,1,1>
 <ζ₄,ζ₄,ζ₄,ζ₄³,ζ₄³,ζ₄³>
 <ζ₄,ζ₄,1,1,ζ₄³,ζ₄³>
 <ζ₄,1,1,1,1,ζ₄³>

we define an element tσ∈ 𝐓 ⋅σ to be quasi-isolated if the Weyl group of C_𝐆 (tσ) is not in any proper parabolic subgroup of $W^σ$. This generalizes the definition for connected groups. The above shows the elements t where tσ runs over representatives of quasi-isolated quasi-semisimple classes of 𝐆 ⋅σ. The given representatives have been chosen σ-stable.

julia> centralizer.(Ref(WF),l)
4-element Vector{ExtendedCox{Perm{Int16}, FiniteCoxeterGroup{Perm{Int16},Rational{Int64}}}}:
 C₃₍₃₂₁₎
 ²A₃₍₃₁₂₎
 (A₁A₁)₍₁₃₎×A₁₍₂₎
 B₂Φ₁

in the above example, the groups C_𝐆(tσ) are computed and displayed as extended Coxeter groups (following the same convention as for centralisers in connected reductive groups).

We define an element tσ∈ 𝐓⋅σ to be isolated if the Weyl group of C_𝐆 (tσ)⁰ is not in any proper parabolic subgroup of $W^σ$. This generalizes the definition for connected groups.

julia> isisolated.(Ref(WF),l)
4-element BitVector:
 1
 1
 1
 0

PermGroups.Groups.centralizer — Method

centralizer(WF::Spets,t::SemisimpleElement{Root1})

WF should be a Coxeter coset representing an algebraic coset 𝐆 ⋅σ, where 𝐆 is a connected reductive group (represented by W:=Group(WF)), and σ is a quasi-central automorphism of 𝐆 defined by WF. The element t should be a semisimple element of 𝐆. The function returns an extended reflection group describing $C_𝐆(tσ)$, with the reflection group part representing $C_𝐆(tσ)⁰$, and the diagram automorphism part being those induced by $C_𝐆(tσ)/C_𝐆(tσ)⁰$ on $C_𝐆(tσ)⁰$.

julia> WF=rootdatum(:u,6)
u₆

julia> s=ss(Group(WF),[1//4,0,0,0,0,3//4])
SemisimpleElement{Root1}: <ζ₄,1,1,1,1,ζ₄³>

julia> centralizer(WF,s)
B₂Φ₁

julia> centralizer(WF,one(s))
C₃₍₃₂₁₎

Chevie.Semisimple.quasi_isolated_reps — Method

quasi_isolated_reps(WF::Spets,p=0)

WF should be a Coxeter coset representing an algebraic coset 𝐆 ⋅σ, where 𝐆 is a connected reductive group (represented by W=Group(WF)), and σ is a quasi-central automorphism of 𝐆 defined by WF. The function returns a list of semisimple elements of 𝐆 such that tσ, when t runs over this list, are representatives of the conjugacy classes of quasi-isolated quasisemisimple elements of 𝐆 ⋅σ (an element tσ∈ 𝐓 ⋅σ is quasi-isolated if the Weyl group of C_𝐆 (tσ) is not in any proper parabolic subgroup of W^σ). If a second argument p is given, it lists only those representatives which exist in characteristic p.

julia> WF=rootdatum("2E6sc")
²E₆sc

julia> quasi_isolated_reps(WF)
5-element Vector{SemisimpleElement{Root1}}:
 <1,1,1,1,1,1>
 <1,-1,ζ₄,1,ζ₄,1>
 <1,1,1,-1,1,1>
 <1,ζ₃²,1,ζ₃,1,1>
 <1,ζ₄³,1,-1,1,1>

julia> quasi_isolated_reps(WF,2)
2-element Vector{SemisimpleElement{Root1}}:
 <1,1,1,1,1,1>
 <1,ζ₃²,1,ζ₃,1,1>

julia> quasi_isolated_reps(WF,3)
4-element Vector{SemisimpleElement{Root1}}:
 <1,1,1,1,1,1>
 <1,-1,ζ₄,1,ζ₄,1>
 <1,1,1,-1,1,1>
 <1,ζ₄³,1,-1,1,1>

Chevie.Semisimple.isisolated — Method

isisolated(WF::Spets,t::SemisimpleElement{Root1})

WF should be a Coxeter coset representing an algebraic coset 𝐆 ⋅σ, where 𝐆 is a connected reductive group (represented by W=Group(WF)), and σ is a quasi-central automorphism of 𝐆 defined by WF. The element t should be a semisimple element of 𝐆. The function returns a boolean describing whether tσ is isolated, that is whether the Weyl group of C_𝐆 (tσ)⁰ is not in any proper parabolic subgroup of W^σ.

julia> WF=rootdatum(:u,6)
u₆

julia> l=quasi_isolated_reps(WF)
4-element Vector{SemisimpleElement{Root1}}:
 <1,1,1,1,1,1>
 <ζ₄,ζ₄,ζ₄,ζ₄³,ζ₄³,ζ₄³>
 <ζ₄,ζ₄,1,1,ζ₄³,ζ₄³>
 <ζ₄,1,1,1,1,ζ₄³>

julia> isisolated.(Ref(WF),l)
4-element BitVector:
 1
 1
 1
 0