Unipotent classes of reductive groups

Chevie.Ucl

Chevie.Ucl.GreenTable

Chevie.Ucl.ICCTable

Chevie.Ucl.UnipotentClass

Chevie.Ucl.UnipotentClasses

Chevie.Ucl.XTable

Chevie.Ucl.UnipotentValues

Chevie.Ucl.distinguished_parabolics

Chevie.Ucl.induced_linear_form

Chevie.Ucl.special_pieces

Chevie.Ucl — Module

This module gives information about the unipotent conjugacy classes of a connected reductive group 𝐆 over an algebraically closed field k of characteristic p, and various associated invariants, such as the generalised Springer correspondence. The unipotent classes depend on p; their classification differs when p is not good (that is, when p divides one of the coefficients of the highest root of one of the irreducible components). In good characteristic, the unipotent classes are in bijection with nilpotent orbits on the Lie algebra 𝔤 of 𝐆.

For each unipotent conjugacy class, we give the following information for a representative u of that class:

the centralizer $C_𝐆 (u)$, that we describe by the reductive part of $C_𝐆 (u)^0$, by the group of components $A(u):=C_𝐆 (u)/C_𝐆 (u)^0$, and by the dimension of its unipotent radical.

in good characteristic, the Dynkin-Richardson diagram.

the generalised Springer correspondence, attaching a character of the Weyl group or a relative Weyl group to each character of A(u).

The Dynkin-Richarson diagram is attached to a nilpotent element $e$ of the Lie algebra $𝔤$. By the Jacobson-Morozov theorem there exists an $𝔰𝔩₂$ subalgebra of $𝔤$ containing $e$ as the element $\begin{pmatrix}0&1\\0&0 \end{pmatrix}$. Let $𝐒$ be the torus $\begin{pmatrix}h&0\\0&h^{-1} \end{pmatrix}$ of $SL₂$ and let $𝐓$ be a maximal torus containing $𝐒$, so that $𝐒$ is the image of a one-parameter subgroup $σ∈ Y(𝐓)$. Consider the root decomposition $𝔤=∑_{α∈Φ}𝔤_α$ given by $𝐓$ and the root system Φ; then $α↦⟨σ,α⟩$ defines a linear form on $Φ$, determined by its value on simple roots Π. It is possible to choose a system of simple roots such that $⟨σ,α⟩≥ 0$ for $α∈Π$, and then $⟨σ,α⟩∈{0,1,2}$ for any $α∈Π$. The Dynkin diagram of $Π$ decorated by these values $0,1,2$ is called the Dynkin-Richardson diagram of $e$, and in good characteristic is a complete invariant of the $𝔤$-orbit of $e$. The Dynkin-Richardson diagrams thus classify unipotent classes of 𝐆 in good characteristic.

Another classification of unipotent classes was given by (Bala and Carter, 1976). A standard parabolic subgroup 𝐏 of 𝐆 associated with the subset I⊂Π of the simple roots is distinguished if the linear form σ taking the value 2 on α∈ I and 0 on other simple roots satisfies 2n₀+semisimplerank(𝐆)=n₂, where nᵢ is the number of roots in Φ where σ takes the value i. Given a distinguished parabolic 𝐏, there is a unique unipotent class which is dense in the unipotent radical of 𝐏. For this class, the linear form described by the Dynkin-Richardson diagram is equal to σ. Such unipotent classes are called distinguished. The Bala-Carter theorem says that every unipotent class is distinguished in the smallest Levi subgroup 𝐋 which contains it, and that such pairs of 𝐋 and the distinguished parabolic 𝐏 of 𝐋 taken up to 𝐆-conjugacy are in bijection with unipotent classes of 𝐆. The function induced_linear_form allows to find the Dynkin-Richardson diagram in 𝐆 of a unipotent class of 𝐋 and thus recover these diagrams from the Bala-Carter classification.

Let $ℬ$ be the variety of all Borel subgroups and let $ℬᵤ$ be the subvariety of Borel subgroups containing the unipotent element u. Then $dim C_𝐆(u)=rank 𝐆 + 2 dim ℬ_u$ and in good characteristic this dimension can be computed from the linear form σ defined by the Dynkin-Richardson diagram: the dimension of the class of u is the number of roots α such that $⟨σ,α⟩∉{0,1}$.

We now describe the Springer correspondence. Indecomposable locally constant $𝐆$-equivariant sheaves on a unipotent class $C$, called local systems, are parameterised by irreducible characters of $A(u)$ for u∈ C. The ordinary Springer correspondence is a bijection between irreducible characters of the Weyl group and a large subset of the local systems, containing all trivial local systems (those parameterised by the trivial character of $A(u)$ for each $u$). More generally, the generalised Springer correspondence associates to each local system a (unique up to $𝐆$-conjugacy) cuspidal datum, a Levi subgroup $𝐋$ of $𝐆$ and a cuspidal local system on an unipotent class of $𝐋$, and a character of the relative Weyl group $W_𝐆 (𝐋):=N_𝐆 (𝐋)/𝐋$. There are few cuspidal local systems (at most one in each dimension for classical groups). The ordinary Springer correspondence is the special case where $𝐋$ is a maximal torus, the cuspidal local system is the trivial system on the identity element, and $W_𝐆 (𝐋)=W$, the Weyl group.

The Springer correspondence gives information on the character values of a finite reductive group as follows: assume that $k$ is the algebraic closure of a finite field $𝔽_q$ and that $F$ is the Frobenius attached to an $𝔽_q$-structure of $𝐆$. Let $C$ be an $F$-stable unipotent class and let $u∈ C^F$; we call $C$ the geometric class of $u$ and the $𝐆^F$-classes within $C^F$ are parameterised by the $F$-conjugacy classes of $A(u)$, denoted by $H¹(F,A(u))$ (most of the time we can find $u∈ C$ such that $F$ acts trivially on $A(u)$ and $H¹(F,A(u))$ parametrises the conjugacy classes). To an $F$-stable character $φ$ of $A(u)$ we associate the characteristic function of the corresponding local system (actually associated with an extension $φ̃$ of $φ$ to $A(u).F$); it is a class function $Y_{u,φ}$ on $𝐆^F$ which can be normalised so that: $Y_{u,φ}(u₁)=φ̃(cF)$ if $u₁$ is geometrically conjugate to $u$ and its $𝐆^F$-class is parameterised by the $F$-conjugacy class $cF$ of $A(u)$, otherwise $Y_{u,φ}(u₁)=0$. If the pair $u,φ$ corresponds via the Springer correspondence to the character $χ$ of $W_𝐆(𝐋)$, then $Y_{u,φ}$ is also denoted $Yᵪ$. There is another important class of functions indexed by local systems: to a local system on class $C$ is attached an intersection cohomology complex, which is a complex of sheaves supported on the closure $C̄$. Such a complex of sheaves is associated with its characteristic function, a class function of $𝐆^F$ obtained by taking the alternating trace of the Frobenius acting on the stalks of the cohomology sheaves. If $Y_ψ$ is the characteristic function of a local system, the characteristic function of the corresponding intersection cohomology complex is denoted by $X_ψ$. This function is supported on $C̄$, and Lusztig has shown that $X_ψ=∑_χ P_{ψ,χ} Yᵪ$ where $P_{ψ,χ}$ are integer polynomials in $q$ and $Yᵪ$ are attached to local systems on classes lying in $C̄$.

Lusztig and Shoji have given an algorithm to compute the matrix $P_{ψ,χ}$, which is implemented in Chevie. The relation to the characters of $𝐆(𝔽_q)$, considering for simplicity the ordinary Springer correspondence, is that the restriction to the unipotent elements of the almost character $R_χ$ is equal to $q^{bᵪ} Xᵪ$, where $bᵪ$ is $dim ℬᵤ$ for an element u of the class C such that the support of χ is $C̄$. The restrictions of the Deligne-Lusztig characters $R_w$ for w∈ W on the unipotents are called the Green functions and can also be computed by Chevie. The values of all unipotent characters on unipotent elements can also be computed in principle by applying Lusztig's Fourier transform matrix (see the section on the Fourier matrix) but there is a difficulty in that the $Xᵪ$ must first be multiplied by some roots of unity which are not known in all cases (and when known may depend on the congruence class of $q$ modulo some small prime).

Finally, we describe how unipotent classes of 𝐆 are parameterised in various quasisimple groups. In classical types, the classes are parametrised by partitions corresponding to the Jordan form in the standard representation. Thus,

for Aₙ we have partitions of n+1.

for B_n we have partitions of 2n+1 where even parts occur an even number of times. In characteristic 2, types B and C are isogenous so have the same classification; thus see the next paragraph.

for C_n we have partitions of 2n where odd parts occur an even number of times. In characteristic 2, there are 2ᵏ classes attached to a partition where k is the number of even parts which occur an even number of times.

for D_n we have partitions of 2n where even parts occur an even number of times, excepted there are two classes when all parts are even. In characteristic 2, we have partitions of 2n where odd parts occur an even number of times, excepted there are 2ᵏ+δ classes attached to a partition where k is the number of even parts which occur an even number of times, and δ is 2 when all parts are even and 0 otherwise.

In exceptional groups, the names of the classes are derived from the Bala-Carter classification. The name of a class parametrised by (𝐋,𝐏) is of the form l(p) where l is the name of 𝐋 and (p) is present if there is more than one distinguished parabolic in 𝐋 and describes which one it is. Before the Bala-Carter classification was universally adopted, Shoji and Mizuno used a different scheme where a class was sometimes parametrised by a reductive subgroup of maximal rank which was not a Levi. These older labels can be obtained instead by giving the IO property :shoji=>true or :mizuno=>true. In a bad characteristic p, there are extra classes. Each of them is associated with a class c in good characteristic and is called (c)ₚ.

We illustrate the above descriptions on some examples:

julia> UnipotentClasses(rootdatum(:sl,4))
UnipotentClasses(sl₄)
1111<211<22<31<4
┌────┬───────────────────────────────────────────────────────────────┐
│u   │D-R dℬ ᵤ B-C   C_𝐆(u) A₃(Φ₁³) A₁(A₁×A₁Φ₁)/-1 .(A₃)/ζ₄ .(A₃)/ζ₄³│
├────┼───────────────────────────────────────────────────────────────┤
│4   │222    0 222    q³.Z₄     1:4           -1:2    ζ₄:Id    ζ₄³:Id│
│31  │202    1 22.    q⁴.Φ₁   Id:31                                  │
│22  │020    2 2.2 q⁴.A₁.Z₂    2:22          11:11                   │
│211 │101    3 2..  q⁵.A₁Φ₁  Id:211                                  │
│1111│000    6 ...       A₃ Id:1111                                  │
└────┴───────────────────────────────────────────────────────────────┘

The first column of the table gives the name of the unipotent class, here a partition describing the Jordan form. The partial order on unipotent classes given by the Zariski closure is given before the table. The column D-R, which is only shown in good characteristic, gives the Dynkin-Richardson diagram for each class; the column dBu gives the dimension of the variety $ℬ ᵤ$. The column B-C gives the Bala-Carter classification of $u$, that is in the case of $sl₄$ it shows $u$ as a regular unipotent in a Levi subgroup by giving the Dynkin-Richardson diagram of a regular unipotent (all 2's) for the entries corresponding to the Levi and . for the entries which not corresponding to the Levi. The column C(u) describes the group $C_𝐆(u)$: a power $qᵈ$ describes that the unipotent radical of $C_𝐆(u)$ has dimension $d$ (thus $qᵈ$ rational points); then follows a description of the reductive part of the neutral component of $C_𝐆(u)$, given by the name of its root datum. Then if $C_𝐆(u)$ is not connected, the description of $A(u)$ is given using a different vocabulary: a cyclic group of order 4 is given as Z4, and a symmetric group on 3 points would be given as S3.

For example, the first class 4 has $C_𝐆(u)^0$ unipotent of dimension 3 and $A(u)$ equal to Z4, the cyclic group of order 4. The class 22 has $C_G(u)$ with unipotent radical of dimension 4, reductive part of type A1 and $A(u)$ is Z2, the cyclic group of order 2. The other classes have $C_𝐆(u)$ connected. For the class 31 the reductive part of $C_G(u)$ is a torus of rank 1.

Then there is a column for each Springer series, giving for each class the pairs 'a:b' where 'a' is the name of the character of $A(u)$ describing the local system involved and 'b' is the name of the character of the (relative) Weyl group corresponding by the Springer correspondence. At the top of the column is the name of the relative Weyl group, and in brackets the name of the Levi affording a cuspidal local system; then, separated by a /, is a description of the central character associated with the Springer series (omitted if this central character is trivial): all local systems in a given Springer series have the same restriction to the centre of $𝐆$. To see what the picture becomes for another algebraic group in the same isogeny class, for example the adjoint group, one simply discards the Springer series whose central character becomes trivial on the centre of $𝐆$; and each group $A(u)$ must be quotiented by the common kernel of the remaining characters. Here is the table for the adjoint group:

julia> UnipotentClasses(coxgroup(:A,3))
UnipotentClasses(A₃)
1111<211<22<31<4
┌────┬────────────────────────────┐
│u   │D-R dℬ ᵤ B-C  C_𝐆(u) A₃(Φ₁³)│
├────┼────────────────────────────┤
│4   │222    0 222      q³    Id:4│
│31  │202    1 22.   q⁴.Φ₁   Id:31│
│22  │020    2 2.2   q⁴.A₁   Id:22│
│211 │101    3 2.. q⁵.A₁Φ₁  Id:211│
│1111│000    6 ...      A₃ Id:1111│
└────┴────────────────────────────┘

Here is another example:

julia> UnipotentClasses(coxgroup(:G,2))
UnipotentClasses(G₂)
1<A₁<Ã₁<G₂(a₁)<G₂
┌──────┬──────────────────────────────────────────┐
│u     │D-R dℬ ᵤ B-C C_𝐆(u)         G₂(Φ₁²)  .(G₂)│
├──────┼──────────────────────────────────────────┤
│G₂    │ 22    0  22     q²         Id:φ₁‚₀       │
│G₂(a₁)│ 20    1  20  q⁴.S₃ 21:φ′₁‚₃ 3:φ₂‚₁ 111:Id│
│Ã₁    │ 01    2  .2  q³.A₁         Id:φ₂‚₂       │
│A₁    │ 10    3  2.  q⁵.A₁        Id:φ″₁‚₃       │
│1     │ 00    6  ..     G₂         Id:φ₁‚₆       │
└──────┴──────────────────────────────────────────┘

which illustrates that on class G₂(a₁) there are two local systems in the principal series of the Springer correspondence, and another cuspidal local system. It also illustrates how we represent the Bala-Carter classification in general. If a class is attached to (𝐋,𝐏) then the simple roots in the complement of 𝐋 have a .. Those in 𝐋 have a 0 or a 2, the 2s characterising 𝐏. So, from the B-C column, we see that that G₂(a₁) is not in a proper Levi, in which case the Bala-Carter diagram coincides with the Dynkin-Richardson diagram.

Characteristics 2 and 3 are not good for G2. To get the unipotent classes and the Springer correspondence in a bad characteristic, give a second argument to the function UnipotentClasses:

julia> UnipotentClasses(coxgroup(:G,2),3)
UnipotentClasses(G₂,3)
1<A₁,(Ã₁)₃<Ã₁<G₂(a₁)<G₂
┌──────┬───────────────────────────────────────────┐
│u     │dℬ ᵤ B-C C_𝐆(u)  G₂(Φ₁²) .(G₂) .(G₂)  .(G₂)│
├──────┼───────────────────────────────────────────┤
│G₂    │   0  22  q².Z₃   1:φ₁‚₀       ζ₃:Id ζ₃²:Id│
│G₂(a₁)│   1  20  q⁴.Z₂   2:φ₂‚₁ 11:Id             │
│Ã₁    │   2  .2     q⁶  Id:φ₂‚₂                   │
│A₁    │   3  2.  q⁵.A₁ Id:φ″₁‚₃                   │
│(Ã₁)₃ │   3  ??  q⁵.A₁ Id:φ′₁‚₃                   │
│1     │   6  ..     G₂  Id:φ₁‚₆                   │
└──────┴───────────────────────────────────────────┘

The function ICCTable gives the transition matrix between the functions $Xᵪ$ and $Y_ψ$.

julia> uc=UnipotentClasses(coxgroup(:G,2));
julia> t=ICCTable(uc;q=Pol(:q))
Coefficients of Xᵪ on Yᵩ for series L=G₂₍₎=Φ₁² W_G(L)=G₂
┌──────┬─────────────────────────────┐
│X\Y   │G₂ G₂(a₁)⁽²¹⁾ G₂(a₁) Ã₁ A₁  1│
├──────┼─────────────────────────────┤
│Xφ₁‚₀ │ 1          .      1  1  1  1│
│Xφ′₁‚₃│ .          1      .  1  . q²│
│Xφ₂‚₁ │ .          .      1  1  1 Φ₈│
│Xφ₂‚₂ │ .          .      .  1  1 Φ₄│
│Xφ″₁‚₃│ .          .      .  .  1  1│
│Xφ₁‚₆ │ .          .      .  .  .  1│
└──────┴─────────────────────────────┘

An example which illustrates how to get the shoji names of classes

julia> uc=UnipotentClasses(coxgroup(:F,4));

julia> uc.classes[10:end]
7-element Vector{UnipotentClass}:
 UnipotentClass(C₃(a₁))
 UnipotentClass(F₄(a₃))
 UnipotentClass(C₃)
 UnipotentClass(B₃)
 UnipotentClass(F₄(a₂))
 UnipotentClass(F₄(a₁))
 UnipotentClass(F₄)

julia> xdisplay(uc.classes[10:end],shoji=true)
7-element Vector{UnipotentClass}:
 UnipotentClass(A₁+B₂)
 UnipotentClass(A₃+Ã₁)
 UnipotentClass(C₃)
 UnipotentClass(B₃)
 UnipotentClass(C₃+A₁)
 UnipotentClass(B₄)
 UnipotentClass(F₄)

In the ICCTable the row labels and column labels show the two ways of indexing local systems: the row labels give the character of the relative Weyl group and the column labels give the class and the name of the local system as a character of A(u): for example, G2(a1) is the trivial local system of the class G2(a1), while G2(a1)(21) is the local system on this class corresponding to the 2-dimensional character of $A(u)=A₂$.

The data on unipotent classes for arbitrary reductive groups are obtained as follows. We may reduce to the case where the radical is trivial, that is 𝐆 is semisimple. Then 𝐆 is the quotient of a simply connected group by a finite central group. The data for quasi-simple simply connected groups have been entered by hand for each type and a general simply connected group is a direct product of quasi-simple ones. In such a group, a character of A(Z), the group of components of the centre, is attached to each Springer series. For an isogenous reductive group 𝐆' with centre Z' the group A(Z') is a quotient of the group A(Z). The Springer series for 𝐆' are those such that the corresponding character of A(Z) factors through A(Z') (to compute A(Z') see algebraic_center). The geometric unipotent classes of 𝐆 and 𝐆' are in bijection. For u a unipotent element of 𝐆' (which we can also consider as a unipotent element of 𝐆) the group A₁=A(u) in 𝐆' is a quotient of A=A(u) in 𝐆 which we can compute as follows: the Springer correspondence for 𝐆' tells us which characters of A survive in 𝐆'. Then A' is the quotient of A by the common kernel of these characters.

Chevie.Ucl.UnipotentClasses — Type

UnipotentClasses(W[,p])

W should be a FiniteCoxeterGroup record for a Weyl group or rootdatum describing a reductive algebraic group 𝐆. The function returns a record containing information about the unipotent classes of 𝐆 in characteristic p (if omitted, p is assumed to be any good characteristic for 𝐆). This contains the following fields:

group: a pointer to W

p: the characteristic of the field for which the unipotent classes were computed. It is 0 for any good characteristic.

orderclasses: a list describing the Hasse diagram of the partial order induced on unipotent classes by the closure relation. That is .orderclasses[i] is the list of j such that $C̄ⱼ⊋ C̄ᵢ$ and there is no class $Cₖ$ such that $C̄ⱼ⊋ C̄ₖ⊋ C̄ᵢ$.

classes: a list of records holding information for each unipotent class. See the help for UnipotentClass for a description of these records.

springerseries: a list of records, each of which describes a Springer series of 𝐆.

The records describing individual Springer series have the following fields:

levi:the indices of the reflections corresponding to the Levi subgroup 𝐋 where lives the cuspidal local system ι from which the Springer series is induced.

relgroup: The relative Weyl group $N_𝐆(𝐋,ι)/𝐋$. The first series is the principal series for which .levi=[] and .relgroup=W.

locsys: a list of length nconjugacy_classes(.relgroup), holding in i-th position a pair describing which local system corresponds to the i-th character of $N_𝐆(𝐋,ι)$. The first element of the pair is the index of the concerned unipotent class u, and the second is the index of the corresponding character of A(u).

Z: the central character associated to the Springer series, specified by its value on the generators of the center.

julia> W=rootdatum(:sl,4)
sl₄

julia> uc=UnipotentClasses(W);

julia> uc.classes
5-element Vector{UnipotentClass}:
 UnipotentClass(1111)
 UnipotentClass(211)
 UnipotentClass(22)
 UnipotentClass(31)
 UnipotentClass(4)

The show function for unipotent classes accepts all the options of showtable and of charnames. Giving the option mizuno (resp. shoji) uses the names given by (Mizuno, 1977; Mizuno, 1980) (resp. (Shoji, 1982)) for unipotent classes. Moreover, there is also an option fourier which gives the Springer correspondence tensored with the sign character of each relative Weyl group, which is the correspondence obtained via a Fourier-Deligne transform (here we assume that p is very good, so that there is a nondegenerate invariant bilinear form on the Lie algebra, and also one can identify nilpotent orbits with unipotent classes).

Here is how to display the non-cuspidal part of the Springer correspondence of the unipotent classes of E₆ using the notations of (Mizuno, 1977) for the classes and those of (Frame, 1951) for the characters of the Weyl group and of (Spaltenstein, 1985) for the characters of G₂ (this is convenient for checking our data with (Spaltenstein, 1985) which used that same conventions):

julia> uc=UnipotentClasses(rootdatum(:E6sc));

julia> xdisplay(uc;cols=[5,6,7],spaltenstein=true,frame=true,mizuno=true,
      order=false)
┌──────┬─────────────────────────────────────────────────────┐
│u     │             E₆(Φ₁⁶) G₂(A₂×A₂Φ₁²)/ζ₃ G₂(A₂×A₂Φ₁²)/ζ₃²│
├──────┼─────────────────────────────────────────────────────┤
│E₆    │                1:1ₚ            ζ₃:1            ζ₃²:1│
│E₆(a₁)│                1:6ₚ           ζ₃:εₗ           ζ₃²:εₗ│
│D₅    │              Id:20ₚ                                 │
│A₅+A₁ │        -1:15ₚ 1:30ₚ           ζ₃:θ′           ζ₃²:θ′│
│A₅    │              1:15_q           ζ₃:θ″           ζ₃²:θ″│
│D₅(a₁)│              Id:64ₚ                                 │
│A₄+A₁ │              Id:60ₚ                                 │
│D₄    │              Id:24ₚ                                 │
│A₄    │              Id:81ₚ                                 │
│D₄(a₁)│111:20ₛ 3:80ₛ 21:90ₛ                                 │
│A₃+A₁ │              Id:60ₛ                                 │
│2A₂+A₁│               1:10ₛ          ζ₃:ε_c          ζ₃²:ε_c│
│A₃    │             Id:81ₚ′                                 │
│A₂+2A₁│             Id:60ₚ′                                 │
│2A₂   │              1:24ₚ′            ζ₃:ε            ζ₃²:ε│
│A₂+A₁ │             Id:64ₚ′                                 │
│A₂    │      11:15ₚ′ 2:30ₚ′                                 │
│3A₁   │            Id:15_q′                                 │
│2A₁   │             Id:20ₚ′                                 │
│A₁    │              Id:6ₚ′                                 │
│1     │              Id:1ₚ′                                 │
└──────┴─────────────────────────────────────────────────────┘

Chevie.Ucl.UnipotentClass — Type

A struct UnipotentClass representing the class C of a unipotent element u of the reductive group 𝐆 with Weyl group W, contains always the following information

.name The name of C

.parameter A parameter describing C. Sometimes the same as .name; a partition describing the Jordan form, for classical groups.

.dimBu The dimension of the variety of Borel subgroups containing u.

For some classes in types E₆, E₇, E₈ there is a field .mizuno containing the names given by Mizuno for some classes, and for some classes in type F₄ a field .shoji containing the names given by Shoji.

A UnipotentClass contains also some of the following information (all of it for some types and some characteristics but sometimes much less)

.dynkin the Dynkin-Richardson diagram of C (a vector giving a weight 0, 1 or 2 to the simple roots).

.dimred the dimension of the reductive part of C_G(u).

.red a CoxeterCoset recording the type of the reductive part of C_G(u), with the twisting induced by the Frobenius if any.

.Au the group A_G(u)=C_G(u)/C^0_G(u).

.balacarter encodes the Bala-Carter classification of C, which says that u is distinguished in a Levi L (the Richardson class in a parabolic P_L) as a vector listing the indices of the simple roots in L, with those not in P_L negated.

.rep a list of indices for roots such that if U=UnipotentGroup(W) then prod(U,u.rep) is a representative of C (which can be obtained also by representative(W,u)).

.dimunip the dimension of the unipotent part of C_G(u).

.AuAction an ExtendedCoxeterGroup recording the action of A_G(u) on red.

Chevie.Ucl.ICCTable — Type

ICCTable(uc,seriesNo=1;q=Pol())

This function gives the table of decompositions of the functions $X_ι$ in terms of the functions $Y_ι$. Here ι is a 𝐆-equivariant local system on the class C of a unipotent element u. Such a local system is parametrised by the pair (u,ϕ) of u and a character of the group of components A(u) of $C_𝐆 (u)$. The function $Y_ι$ is the characteristic function of this local system and $X_ι$ is the characteristic function of the corresponding intersection cohomology complex on C̄. The Springer correspondence says that the local systems can also be indexed by characters of a relative Weyl group. Since the coefficient of Xᵪ on Yᵩ is 0 if χ and φ are not characters of the same relative Weyl group (are not in the same Springer series), the table is for one Springer series, specified by the argument 'seriesNo' (this defaults to 'seriesNo=1' which is the principal series). The decomposition multiplicities are graded, and are given as polynomials in one variable (specified by the argument q; if not given Pol() is assumed).

julia> uc=UnipotentClasses(coxgroup(:A,3));t=ICCTable(uc)
Coefficients of Xᵪ on Yᵩ for series L=A₃₍₎=Φ₁³ W_G(L)=A₃
┌─────┬────────────────┐
│X\Y  │4 31 22 211 1111│
├─────┼────────────────┤
│X4   │1  1  1   1    1│
│X31  │.  1  1  Φ₂   Φ₃│
│X22  │.  .  1   1   Φ₄│
│X211 │.  .  .   1   Φ₃│
│X1111│.  .  .   .    1│
└─────┴────────────────┘

In the above the multiplicities are given as products of cyclotomic polynomials to display them more compactly. However the format of such a table can be controlled more precisely.

For instance, one can ask to not display the entries as products of cyclotomic polynomials and not display the zeroes as '.'

julia> xdisplay(t;cycpol=false,dotzero=false)
Coefficients of Xᵪ on Yᵩ for A3
┌─────┬──────────────────┐
│X\Y  │4 31 22 211   1111│
├─────┼──────────────────┤
│X4   │1  1  1   1      1│
│X31  │0  1  1 q+1 q²+q+1│
│X22  │0  0  1   1   q²+1│
│X211 │0  0  0   1 q²+q+1│
│X1111│0  0  0   0      1│
└─────┴──────────────────┘

Since show uses the function showtable, all the options of this function are also available. We can use this to restrict the entries displayed to a given sublist of the rows and columns (here the indices correspond to the number in Chevie of the corresponding character of the relative Weyl group of the given Springer series):

julia> uc=UnipotentClasses(coxgroup(:F,4));
julia> t=ICCTable(uc);
julia> sh=[13,24,22,18,14,9,11,19];
julia> xdisplay(t,rows=sh,cols=sh)
Coefficients of Xᵪ on Yᵩ for series L=F₄₍₎=Φ₁⁴ W_G(L)=F₄
┌───────┬────────────────────────────────────────────┐
│X\Y    │A₁+Ã₁ A₂ Ã₂ A₂+Ã₁ Ã₂+A₁ B₂⁽¹¹⁾ B₂ C₃(a₁)⁽¹¹⁾│
├───────┼────────────────────────────────────────────┤
│Xφ₉‚₁₀ │    1  .  .     .     .      .  .          .│
│Xφ″₈‚₉ │    1  1  .     .     .      .  .          .│
│Xφ′₈‚₉ │    1  .  1     .     .      .  .          .│
│Xφ″₄‚₇ │    1  1  .     1     .      .  .          .│
│Xφ′₆‚₆ │   Φ₄  1  1     1     1      .  .          .│
│Xφ₄‚₈  │   q²  .  .     .     .      1  .          .│
│Xφ″₉‚₆ │   Φ₄ Φ₄  .     1     .      .  1          .│
│Xφ′₄‚₇ │   q²  . Φ₄     .     1      .  .          1│
└───────┴────────────────────────────────────────────┘

The ìo option rowlocsys=true will display local systems also for the row labels.

The function 'ICCTable' returns an object with various pieces of information which can help further computations.

.scalar: this contains the table of multiplicities Pᵪᵩ of the Xᵪ on the Yᵩ. One should pay attention that by default, the table is not displayed in the same order as the stored |.scalar|, which is in order in Chevie of the characters in the relative Weyl group; the table is transposed, then lines and rows are sorted by dimBu,class no,index of character in A(u) while displayed.

.group: The group W.

.relgroup: The relative Weyl group for the Springer series.

.series: The index of the Springer series given for W.

.dimBu: The list of $dimℬᵤ$ for each local system (u,φ) in the series.

:L: The matrix of (unnormalised) scalar products of the functions $Yᵩ$ with themselves, that is the $(φ,χ)$ entry is $∑_{g∈𝐆(𝔽_q)} Yᵩ(g) Yᵪ(g)$. This is thus a symmetric, block-diagonal matrix where the diagonal blocks correspond to geometric unipotent conjugacy classes. This matrix is obtained as a by-product of Lusztig's algorithm to compute $Pᵩᵪ$.

Chevie.Ucl.XTable — Type

XTable(uc;classes=false)

This function presents in a different way the information obtained from ICCTable. Let $X̃_{u,ϕ}=q^{1/2(codim C-dim Z(𝐋 ))}X_{u,ϕ}$ where C is the class of u and Z(𝐋) is the center of Levi subgroup on which lives the cuspidal local system attached to the local system (u,ϕ).

Then XTable(uc) gives the decomposition of the functions $X̃_{u,ϕ}$ on local systems. t=XTable(uc,classes==true) gives the values of the functions $X̃_{u,ϕ}$ on unipotent classes. A side effect of classes=true is to compute the cardinal of the unipotent conjugacy classes, available in t.cardClass; in this case displaying t will show the cardinal of the centralizers of unipotent elements, available in t.centClass.

julia> W=coxgroup(:G,2)
G₂

julia> XTable(UnipotentClasses(W))
Values of character sheaves X̃ᵪ of G₂ on local systems φ
┌──────────┬───────────────────────────────────────────┐
│X̃ᵪ|φ      │   1 A₁ Ã₁ G₂(a₁)⁽¹¹¹⁾ G₂(a₁)⁽²¹⁾ G₂(a₁) G₂│
├──────────┼───────────────────────────────────────────┤
│X_φ₁‚₀^G₂ │   1  1  1           .          .      1  1│
│X_φ₁‚₆^G₂ │  q⁶  .  .           .          .      .  .│
│X_φ′₁‚₃^G₂│  q³  .  q           .          q      .  .│
│X_φ″₁‚₃^G₂│  q³ q³  .           .          .      .  .│
│X_φ₂‚₁^G₂ │ qΦ₈  q  q           .          .      q  .│
│X_φ₂‚₂^G₂ │q²Φ₄ q² q²           .          .      .  .│
│X_Id^.    │   .  .  .          q²          .      .  .│
└──────────┴───────────────────────────────────────────┘

The functions X̃ in the first column are decorated by putting as an exponent the relative groups $W_𝐆 (𝐋)$.

julia> t=XTable(UnipotentClasses(W);classes=true)
Values of character sheaves X̃ᵪ of G₂ on unipotent classes
┌──────────┬─────────────────────────────────────────────────────────┐
│X̃ᵪ|class  │           1     A₁     Ã₁ G₂(a₁) G₂(a₁)₍₂₁₎ G₂(a₁)₍₃₎ G₂│
├──────────┼─────────────────────────────────────────────────────────┤
│X_φ₁‚₀^G₂ │           1      1      1      1          1         1  1│
│X_φ₁‚₆^G₂ │          q⁶      .      .      .          .         .  .│
│X_φ′₁‚₃^G₂│          q³      .      q     2q          .        -q  .│
│X_φ″₁‚₃^G₂│          q³     q³      .      .          .         .  .│
│X_φ₂‚₁^G₂ │         qΦ₈      q      q      q          q         q  .│
│X_φ₂‚₂^G₂ │        q²Φ₄     q²     q²      .          .         .  .│
│X_Id^.    │           .      .      .     q²        -q²        q²  .│
├──────────┼─────────────────────────────────────────────────────────┤
│|C_𝐆(u)|  │q⁶Φ₁²Φ₂²Φ₃Φ₆ q⁶Φ₁Φ₂ q⁴Φ₁Φ₂    6q⁴        2q⁴       3q⁴ q²│
└──────────┴─────────────────────────────────────────────────────────┘

julia> XTable(UnipotentClasses(W,2))
Values of character sheaves X̃ᵪ of G₂ on local systems φ
┌──────────┬──────────────────────────────────────────────────┐
│X̃ᵪ|φ      │   1 A₁ Ã₁ G₂(a₁)⁽¹¹¹⁾ G₂(a₁)⁽²¹⁾ G₂(a₁) G₂⁽¹¹⁾ G₂│
├──────────┼──────────────────────────────────────────────────┤
│X_φ₁‚₀^G₂ │   1  1  1           .          .      1      .  1│
│X_φ₁‚₆^G₂ │  q⁶  .  .           .          .      .      .  .│
│X_φ′₁‚₃^G₂│  q³  .  q           .          q      .      .  .│
│X_φ″₁‚₃^G₂│  q³ q³  .           .          .      .      .  .│
│X_φ₂‚₁^G₂ │ qΦ₈  q  q           .          .      q      .  .│
│X_φ₂‚₂^G₂ │q²Φ₄ q² q²           .          .      .      .  .│
│X_Id^.    │   .  .  .          q²          .      .      .  .│
│X_Id^.    │   .  .  .           .          .      .      q  .│
└──────────┴──────────────────────────────────────────────────┘

julia> XTable(UnipotentClasses(rootdatum(:sl,4)))
Values of character sheaves X̃ᵪ of sl₄ on local systems φ
┌────────┬────────────────────────────────────────────┐
│X̃ᵪ|φ    │1111 211 22⁽¹¹⁾ 22 31 4 4^(ζ₄) 4⁽⁻¹⁾ 4^(ζ₄³)│
├────────┼────────────────────────────────────────────┤
│X₁₁₁₁^A₃│  q⁶   .      .  .  . .      .     .       .│
│X₂₁₁^A₃ │q³Φ₃  q³      .  .  . .      .     .       .│
│X₂₂^A₃  │q²Φ₄  q²      . q²  . .      .     .       .│
│X₃₁^A₃  │ qΦ₃ qΦ₂      .  q  q .      .     .       .│
│X₄^A₃   │   1   1      .  1  1 1      .     .       .│
│X₁₁^A₁  │   .   .     q³  .  . .      .     .       .│
│X₂^A₁   │   .   .     q²  .  . .      .     q       .│
│X_Id^.  │   .   .      .  .  . .   q³⁄₂     .       .│
│X_Id^.  │   .   .      .  .  . .      .     .    q³⁄₂│
└────────┴────────────────────────────────────────────┘

Chevie.Ucl.GreenTable — Type

GreenTable(uc;classes=false)

Keeping the same notations as in the description of XTable, this function returns a table of the functions $Q_{wF}$, attached to elements $wF∈ W_𝐆 (𝐋)⋅F$ where $W_𝐆 (𝐋)$ are the relative weyl groups attached to cuspidal local systems. These functions are defined by $Q_{wF}=∑_{u,ϕ} ϕ̃(wF) X̃_{u,ϕ}$. Note that for the principal Springer series, when 𝐋=𝐓 is a maximal torus, the function $Q_{wF}$ coincides with the unipotent values of the Deligne-Lusztig character $R^𝐆_{𝐓_W}(1)$. As for XTable, by default values are given in terms of local systems. If classes=true is given, then the values of the $Q_{wF}$ on conjugacy classes are given.

julia> W=coxgroup(:G,2)
G₂

julia> GreenTable(UnipotentClasses(W))
Values of Green functions Q_wF of G₂ on local systems φ
┌──────────┬──────────────────────────────────────────────────────────┐
│Qᴵ_wF|φ   │        1     A₁       Ã₁ G₂(a₁)⁽¹¹¹⁾ G₂(a₁)⁽²¹⁾ G₂(a₁) G₂│
├──────────┼──────────────────────────────────────────────────────────┤
│Q_A₀^G₂   │  Φ₂²Φ₃Φ₆   Φ₂Φ₃ (2q+1)Φ₂           .          q   2q+1  1│
│Q_Ã₁^G₂   │-Φ₁Φ₂Φ₃Φ₆  -Φ₁Φ₃       Φ₂           .          q      1  1│
│Q_A₁^G₂   │-Φ₁Φ₂Φ₃Φ₆   Φ₂Φ₆      -Φ₁           .         -q      1  1│
│Q_G₂^G₂   │ Φ₁²Φ₂²Φ₃ -Φ₁Φ₂²    -Φ₁Φ₂           .         -q     Φ₂  1│
│Q_A₂^G₂   │ Φ₁²Φ₂²Φ₆  Φ₁²Φ₂    -Φ₁Φ₂           .          q    -Φ₁  1│
│Q_A₁+Ã₁^G₂│  Φ₁²Φ₃Φ₆  -Φ₁Φ₆ (2q-1)Φ₁           .         -q  -2q+1  1│
│Q^.       │        .      .        .          q²          .      .  .│
└──────────┴──────────────────────────────────────────────────────────┘

The functions $Q_{wF}$ depend only on the conjugacy class of wF, so in the first column the indices of Q are the names of the conjugacy classes of $W_𝐆(𝐋)$. The exponents are the names of the groups $W_𝐆(𝐋)$. The validity of the tables is for generic q, which means that some assumptions on q may be necessary. For instance in the last table we need q≡1 (mod 3). For q≡-1 (mod 3) the non-zero entry in the last line is -q².

julia> GreenTable(UnipotentClasses(W);classes=true)
Values of Green functions Q_wF of G₂ on unipotent classes
┌───────────┬────────────────────────────────────────────────────────┐
│Qᴵ_wF|class│        1     A₁       Ã₁ G₂(a₁) G₂(a₁)₍₂₁₎ G₂(a₁)₍₃₎ G₂│
├───────────┼────────────────────────────────────────────────────────┤
│Q_A₀^G₂    │  Φ₂²Φ₃Φ₆   Φ₂Φ₃ (2q+1)Φ₂   4q+1       2q+1        Φ₂  1│
│Q_Ã₁^G₂    │-Φ₁Φ₂Φ₃Φ₆  -Φ₁Φ₃       Φ₂   2q+1          1       -Φ₁  1│
│Q_A₁^G₂    │-Φ₁Φ₂Φ₃Φ₆   Φ₂Φ₆      -Φ₁  -2q+1          1        Φ₂  1│
│Q_G₂^G₂    │ Φ₁²Φ₂²Φ₃ -Φ₁Φ₂²    -Φ₁Φ₂    -Φ₁         Φ₂      2q+1  1│
│Q_A₂^G₂    │ Φ₁²Φ₂²Φ₆  Φ₁²Φ₂    -Φ₁Φ₂     Φ₂        -Φ₁     -2q+1  1│
│Q_A₁+Ã₁^G₂ │  Φ₁²Φ₃Φ₆  -Φ₁Φ₆ (2q-1)Φ₁  -4q+1      -2q+1       -Φ₁  1│
│Q^.        │        .      .        .     q²        -q²        q²  .│
└───────────┴────────────────────────────────────────────────────────┘

julia> GreenTable(UnipotentClasses(rootdatum(:sl,4)))
Values of Green functions Q_wF of sl₄ on local systems φ
┌────────┬──────────────────────────────────────────────────────────────────┐
│Qᴵ_wF|φ │     1111          211 22⁽¹¹⁾       22   31 4 4^(ζ₄) 4⁽⁻¹⁾ 4^(ζ₄³)│
├────────┼──────────────────────────────────────────────────────────────────┤
│Q₁₁₁₁^A₃│  Φ₂²Φ₃Φ₄ (3q²+2q+1)Φ₂      . (2q+1)Φ₂ 3q+1 1      .     .       .│
│Q₂₁₁^A₃ │-Φ₁Φ₂Φ₃Φ₄   -q³+q²+q+1      .       Φ₂   Φ₂ 1      .     .       .│
│Q₂₂^A₃  │  Φ₁²Φ₃Φ₄        -Φ₁Φ₄      .  2q²-q+1  -Φ₁ 1      .     .       .│
│Q₃₁^A₃  │ Φ₁²Φ₂²Φ₄        -Φ₁Φ₂      .    -Φ₁Φ₂    1 1      .     .       .│
│Q₄^A₃   │ -Φ₁³Φ₂Φ₃        Φ₁²Φ₂      .      -Φ₁  -Φ₁ 1      .     .       .│
│Q₁₁^A₁  │        .            .   q²Φ₂        .    . .      .     q       .│
│Q₂^A₁   │        .            .  -q²Φ₁        .    . .      .     q       .│
│Q^.     │        .            .      .        .    . .   q³⁄₂     .       .│
│Q^.     │        .            .      .        .    . .      .     .    q³⁄₂│
└────────┴──────────────────────────────────────────────────────────────────┘

Chevie.Ucl.UnipotentValues — Function

UnipotentValues(uc,classes=false)

This function returns a table of the values of unipotent characters on local systems (by default) or on unipotent classes (if classes=true). These tables are for generic q, that is their validity depends on some assumptions on q (sometimes q large enough, or for G₂ when q≡1 (mod 3)).

julia> W=coxgroup(:G,2)
G₂

julia> UnipotentValues(UnipotentClasses(W);classes=true)
Values of unipotent characters for G₂ on unipotent classes
┌───────┬─────────────────────────────────────────────────────────────┐
│       │        1          A₁     Ã₁   G₂(a₁) G₂(a₁)₍₂₁₎ G₂(a₁)₍₃₎ G₂│
├───────┼─────────────────────────────────────────────────────────────┤
│φ₁‚₀   │        1           1      1        1          1         1  1│
│φ₁‚₆   │       q⁶           .      .        .          .         .  .│
│φ′₁‚₃  │  qΦ₃Φ₆/3    -qΦ₁Φ₂/3      q (q+5)q/3     -qΦ₁/3     qΦ₁/3  .│
│φ″₁‚₃  │  qΦ₃Φ₆/3  (2q²+1)q/3      .    qΦ₁/3     -qΦ₁/3  (q+2)q/3  .│
│φ₂‚₁   │ qΦ₂²Φ₃/6 (2q+1)qΦ₂/6  qΦ₂/2 (q+5)q/6     -qΦ₁/6     qΦ₁/6  .│
│φ₂‚₂   │ qΦ₂²Φ₆/2       qΦ₂/2  qΦ₂/2   -qΦ₁/2      qΦ₂/2    -qΦ₁/2  .│
│G₂[-1] │ qΦ₁²Φ₃/2      -qΦ₁/2 -qΦ₁/2   -qΦ₁/2      qΦ₂/2    -qΦ₁/2  .│
│G₂[1]  │ qΦ₁²Φ₆/6 (2q-1)qΦ₁/6 -qΦ₁/2 (q+5)q/6     -qΦ₁/6     qΦ₁/6  .│
│G₂[ζ₃] │qΦ₁²Φ₂²/3    -qΦ₁Φ₂/3      .    qΦ₁/3     -qΦ₁/3  (q+2)q/3  .│
│G₂[ζ₃²]│qΦ₁²Φ₂²/3    -qΦ₁Φ₂/3      .    qΦ₁/3     -qΦ₁/3  (q+2)q/3  .│
└───────┴─────────────────────────────────────────────────────────────┘

julia> UnipotentValues(UnipotentClasses(W,3);classes=true)
Values of unipotent characters for G₂ on unipotent classes
┌───────┬─────────────────────────────────────────────────────────────────────┐
│       │        1          A₁         Ã₁ G₂(a₁) G₂(a₁)₍₂₎    G₂       G₂_(ζ₃)│
├───────┼─────────────────────────────────────────────────────────────────────┤
│φ₁‚₀   │        1           1          1      1         1     1             1│
│φ₁‚₆   │       q⁶           .          .      .         .     .             .│
│φ′₁‚₃  │  qΦ₃Φ₆/3    -qΦ₁Φ₂/3        q/3  qΦ₂/3    -qΦ₁/3 -2q/3           q/3│
│φ″₁‚₃  │  qΦ₃Φ₆/3  (2q²+1)q/3        q/3  qΦ₂/3    -qΦ₁/3 -2q/3           q/3│
│φ₂‚₁   │ qΦ₂²Φ₃/6 (2q+1)qΦ₂/6  (3q+1)q/6  qΦ₂/6    -qΦ₁/6  2q/3          -q/3│
│φ₂‚₂   │ qΦ₂²Φ₆/2       qΦ₂/2      qΦ₂/2 -qΦ₁/2     qΦ₂/2     .             .│
│G₂[-1] │ qΦ₁²Φ₃/2      -qΦ₁/2     -qΦ₁/2 -qΦ₁/2     qΦ₂/2     .             .│
│G₂[1]  │ qΦ₁²Φ₆/6 (2q-1)qΦ₁/6 (-3q+1)q/6  qΦ₂/6    -qΦ₁/6  2q/3          -q/3│
│G₂[ζ₃] │qΦ₁²Φ₂²/3    -qΦ₁Φ₂/3        q/3  qΦ₂/3    -qΦ₁/3   q/3 (-ζ₃+2ζ₃²)q/3│
│G₂[ζ₃²]│qΦ₁²Φ₂²/3    -qΦ₁Φ₂/3        q/3  qΦ₂/3    -qΦ₁/3   q/3  (2ζ₃-ζ₃²)q/3│
└───────┴─────────────────────────────────────────────────────────────────────┘

┌───────┬─────────────────────────┐
│       │     G₂_(ζ₃²)       (Ã₁)₃│
├───────┼─────────────────────────┤
│φ₁‚₀   │            1           1│
│φ₁‚₆   │            .           .│
│φ′₁‚₃  │          q/3  (2q²+1)q/3│
│φ″₁‚₃  │          q/3    -qΦ₁Φ₂/3│
│φ₂‚₁   │         -q/3 (2q+1)qΦ₂/6│
│φ₂‚₂   │            .       qΦ₂/2│
│G₂[-1] │            .      -qΦ₁/2│
│G₂[1]  │         -q/3 (2q-1)qΦ₁/6│
│G₂[ζ₃] │ (2ζ₃-ζ₃²)q/3    -qΦ₁Φ₂/3│
│G₂[ζ₃²]│(-ζ₃+2ζ₃²)q/3    -qΦ₁Φ₂/3│
└───────┴─────────────────────────┘

Chevie.Ucl.induced_linear_form — Function

induced_linear_form(W, H, h)

This routine can be used to find the unipotent class in the reductive group with Weyl group W which contains a given unipotent class of a reductive subgroup of maximum rank represented by the reflection subgroup H of W.

The argument h is a linear form on the roots of H, given by its value on the simple roots; this linear form is extended to the roots of W by 0 on the orthogonal of the roots of H; and finally the resulting form is conjugated by an element of W so that it takes positive values on the simple roots. If the initial form describes a Dynkin-Richardson diagram for H, the result will describe a Dynkin-Richardson diagram for W.

julia> W=coxgroup(:F,4)
F₄

julia> H=reflection_subgroup(W,[1,3])
F₄₍₁₃₎=A₁×Ã₁Φ₁²

julia> induced_linear_form(W,H,[2,2])
4-element Vector{Int64}:
 0
 1
 0
 0

julia> uc=UnipotentClasses(W);

julia> uc.classes[4].dynkin
4-element Vector{Int64}:
 0
 1
 0
 0

julia> uc.classes[4]
UnipotentClass(A₁+Ã₁)

The example above shows that the class containing the regular class of the Levi subgroup of type A₁×Ã₁ is the class A₁+Ã₁.

Chevie.Ucl.special_pieces — Function

'special_pieces(<uc>)'

The special pieces form a partition of the unipotent variety of a reductive group 𝐆 which was first defined in (Spaltenstein, 1982; chap. III) as the fibres of d^2, where d is a "duality map". Another definition is as the set of classes in the Zariski closure of a special class and not in the Zariski closure of any smaller special class, where a special class is a class such that the trivial local system on it corresponds to a special character by the Springer correspondence.

Since each Lusztig family of characters contains a unique speical character, special pieces are in bijection with lusztig families.

Each piece is a union of unipotent conjugacy classes so is represented in Chevie as a list of class numbers. Thus the list of special pieces is returned as a list of lists of class numbers. The list is sorted by increasing piece dimension, while each piece is sorted by decreasing class dimension, so that the special class is listed first.

julia> W=coxgroup(:G,2)
G₂

julia> special_pieces(UnipotentClasses(W))
3-element Vector{Vector{Int64}}:
 [1]
 [4, 3, 2]
 [5]

julia> special_pieces(UnipotentClasses(W,3))
3-element Vector{Vector{Int64}}:
 [1]
 [4, 3, 2, 6]
 [5]

The example above shows that the special pieces may be different in bad characteristic.

Chevie.Ucl.distinguished_parabolics — Function

distinguished_parabolics(W)

the list of distinguished (in the sense of Richardson) standard parabolic subgroups W_I of W, each given by the list I of indices in the simple reflections. The distinguished unipotent conjugacy classes of the reductive group 𝐆 with Weyl group W consist of the dense unipotent orbit in the unipotent radical of a parabolic subgroup 𝐏_I of 𝐆 associated with a distinguished W_I. Their Dynkin-Richardson diagram contains a 0 at indices I and a 2 in other entries. If 𝐏_I=𝐋𝐔 is a Levi decomposition, 𝐏_I is distinguished iff dim 𝐋∩𝐆'=dim 𝐔/𝐔'.

julia> W=coxgroup(:F,4)
F₄

julia> distinguished_parabolics(W)
4-element Vector{Vector{Int64}}:
 []
 [3]
 [1, 3]
 [1, 3, 4]