Classtypes

Chevie.Gt.closed_subsystems вАФ Function

closed_subsystems(W)

W should be a Weyl group. The function returns the list, sorted by rank, of closed subsystems of the root system of W. Each closed subsystem is represented by the list of indices of its simple roots. If W is the Weyl group of a reductive group рЭРЖ, then closed subsystem correspond to reductive subgroups of maximal rank. And all such groups are obtained this way, apart from some exceptions in characteristics 2 and 3, see (Malle and Testerman, 2011; Proposition 13.4).

julia> W=coxgroup(:G,2)
GвВВ

julia> closed_subsystems(W)
12-element Vector{Vector{Int64}}:
 []
 [1]
 [2]
 [3]
 [4]
 [5]
 [6]
 [1, 2]
 [1, 4]
 [1, 5]
 [2, 6]
 [3, 5]

Chevie.Gt.closed_subsystemsPoset вАФ Function

closed_subsystemsPoset(W)

W should be a Weyl group. The function returns the Poset of closed subsystems of the root system of W. Each closed subsystem is represented by the list of indices of its simple roots. If W is the Weyl group of a reductive group рЭРЖ, then closed subsystem correspond to reductive subgroups of maximal rank. And all such groups are obtained this way, apart from some exceptions in characteristics 2 and 3, see (Malle and Testerman, 2011; Proposition 13.4).

julia> W=coxgroup(:G,2)
GвВВ

julia> closed_subsystemsPoset(W)
1 2<1 4<4<вИЕ
1 2<1 5<1<вИЕ
1 2<2 6<6<вИЕ
1 2<3 5<5<вИЕ
1 4<1
1 5<6
1 5<5
2 6<2<вИЕ
3 5<3<вИЕ

Chevie.Gt.closed_subsystems_reps вАФ Function

closed_subsystems_reps(W)

W should be a Weyl group. The function returns representatives of the W-orbits of closed subsystems of the root system of W, sorted by rank. Each closed subsystem is represented by the list of indices of its simple roots. If W is the Weyl group of a reductive group рЭРЖ, then closed subsystem correspond to reductive subgroups of maximal rank. And all such groups are obtained this way, apart from some exceptions in characteristics 2 and 3, see (Malle and Testerman, 2011; Proposition 13.4).

julia> W=coxgroup(:G,2)
GвВВ

julia> map(I->reflection_subgroup(W,I),closed_subsystems_reps(W))
6-element Vector{FiniteCoxeterSubGroup{Perm{Int16},Int64}}:
 GвВВвВНвВО=ќ¶вВБ¬≤
 GвВВвВНвВБвВО=AвВБќ¶вВБ
 GвВВвВНвВВвВО=AћГвВБќ¶вВБ
 GвВВ
 GвВВвВНвВБвВДвВО=AвВБ√ЧAћГвВБ
 GвВВвВНвВБвВЕвВО=AвВВ

Chevie.Gt.ClassTypes вАФ Type

ClassTypes(G[,p])

G should be a root datum or a twisted root datum representing a finite reductive group $рЭРЖ ^F$ and p should be a prime. The function returns the class types of G in characteristic p (in good characteristic if p is omitted). Two elements of $рЭРЖ ^F$ have the same class type if their centralizers are conjugate. If su is the Jordan decomposition of an element x, the class type of x is determined by the class type of its semisimple part s and the unipotent class of u in $C_рЭРЖ (s)$.

The function ClassTypes is presently only implemented for simply connected groups, where $C_рЭРЖ (s)$ is connected. This section is a bit experimental and may change in the future.

ClassTypes returns a struct which contains a list of classtypes for semisimple elements, which are represented by subspets and contain additionnaly information on the unipotent classes of $C_рЭРЖ (s)$.

Let us give some examples:

julia> t=ClassTypes(rootdatum(:sl,3))
ClassTypes(AвВВ,good characteristic)
вФМвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФђвФАвФАвФАвФАвФАвФАвФАвФАвФАвФР
вФВC_G(s)    вФВ |C_G(s)|вФВ
вФЬвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФЉвФАвФАвФАвФАвФАвФАвФАвФАвФАвФ
вФВAвВВвВНвВО=ќ¶вВБ¬≤  вФВ      ќ¶вВБ¬≤вФВ
вФВAвВВвВНвВО=ќ¶вВБќ¶вВВ вФВ     ќ¶вВБќ¶вВВвФВ
вФВAвВВвВНвВО=ќ¶вВГ   вФВ       ќ¶вВГвФВ
вФВAвВВвВНвВБвВО=AвВБќ¶вВБвФВ   qќ¶вВБ¬≤ќ¶вВВвФВ
вФВAвВВ        вФВq¬≥ќ¶вВБ¬≤ќ¶вВВќ¶вВГвФВ
вФФвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФівФАвФАвФАвФАвФАвФАвФАвФАвФАвФШ

By default, only information about semisimple centralizer types is returned: the type, and its generic order.

julia> xdisplay(t;unip=true)
ClassTypes(AвВВ,good characteristic)
вФМвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФђвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФР
вФВC_G(s)    вФВ    u |C_G(su)|вФВ
вФЬвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФЉвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФ
вФВAвВВвВНвВО=ќ¶вВБ¬≤  вФВ    1       ќ¶вВБ¬≤вФВ
вФВAвВВвВНвВО=ќ¶вВБќ¶вВВ вФВ    1      ќ¶вВБќ¶вВВвФВ
вФВAвВВвВНвВО=ќ¶вВГ   вФВ    1        ќ¶вВГвФВ
вФВAвВВвВНвВБвВО=AвВБќ¶вВБвФВ   11    qќ¶вВБ¬≤ќ¶вВВвФВ
вФВ          вФВ    2       qќ¶вВБвФВ
вФВAвВВ        вФВ  111 q¬≥ќ¶вВБ¬≤ќ¶вВВќ¶вВГвФВ
вФВ          вФВ   21      q¬≥ќ¶вВБвФВ
вФВ          вФВ    3       3q¬≤вФВ
вФВ          вФВ 3_ќґвВГ       3q¬≤вФВ
вФВ          вФВ3_ќґвВГ¬≤       3q¬≤вФВ
вФФвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФівФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФШ

Here we have displayed information on unipotent classes, with their centralizer.

julia> xdisplay(t;nClasses=true)
ClassTypes(AвВВ,good characteristic)
вФМвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФђвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФР
вФВC_G(s)    вФВ       nClasses  |C_G(s)|вФВ
вФЬвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФЉвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФ
вФВAвВВвВНвВО=ќ¶вВБ¬≤  вФВ(q¬≤-5q+2qвВГ+4)/6       ќ¶вВБ¬≤вФВ
вФВAвВВвВНвВО=ќ¶вВБќ¶вВВ вФВ       (q¬≤-q)/2      ќ¶вВБќ¶вВВвФВ
вФВAвВВвВНвВО=ќ¶вВГ   вФВ  (q¬≤+q-qвВГ+1)/3        ќ¶вВГвФВ
вФВAвВВвВНвВБвВО=AвВБќ¶вВБвФВ       (q-qвВГ-1)    qќ¶вВБ¬≤ќ¶вВВвФВ
вФВAвВВ        вФВ             qвВГ q¬≥ќ¶вВБ¬≤ќ¶вВВќ¶вВГвФВ
вФФвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФівФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФШ

Here we have added information on how many semisimple conjugacy classes of рЭРЖ ^F have a given type. The answer in general involves variables of the form qвВР which represent gcd(q-1,a).

Finally an example in bad characteristic:

julia> t=ClassTypes(coxgroup(:G,2),2);xdisplay(t;nClasses=true)
ClassTypes(GвВВ,char. 2)
ClassTypes(GвВВ,char. 2)
вФМвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФђвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФР
вФВC_G(s)    вФВ         nClasses     |C_G(s)|вФВ
вФЬвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФЉвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФ
вФВGвВВвВНвВО=ќ¶вВБ¬≤  вФВ(q¬≤-8q+2qвВГ+10)/12          ќ¶вВБ¬≤вФВ
вФВGвВВвВНвВО=ќ¶вВБќ¶вВВ вФВ        (q¬≤-2q)/4         ќ¶вВБќ¶вВВвФВ
вФВGвВВвВНвВО=ќ¶вВБќ¶вВВ вФВ        (q¬≤-2q)/4         ќ¶вВБќ¶вВВвФВ
вФВGвВВвВНвВО=ќ¶вВЖ   вФВ    (q¬≤-q-qвВГ+1)/6           ќ¶вВЖвФВ
вФВGвВВвВНвВО=ќ¶вВГ   вФВ    (q¬≤+q-qвВГ+1)/6           ќ¶вВГвФВ
вФВGвВВвВНвВО=ќ¶вВВ¬≤  вФВ (q¬≤-4q+2qвВГ-2)/12          ќ¶вВВ¬≤вФВ
вФВGвВВвВНвВБвВО=AвВБќ¶вВБвФВ       (q-qвВГ-1)/2       qќ¶вВБ¬≤ќ¶вВВвФВ
вФВGвВВвВНвВБвВО=AвВБќ¶вВВвФВ       (q-qвВГ+1)/2       qќ¶вВБќ¶вВВ¬≤вФВ
вФВGвВВвВНвВВвВО=AћГвВБќ¶вВБвФВ          (q-2)/2       qќ¶вВБ¬≤ќ¶вВВвФВ
вФВGвВВвВНвВВвВО=AћГвВБќ¶вВВвФВ              q/2       qќ¶вВБќ¶вВВ¬≤вФВ
вФВGвВВ        вФВ                1 qвБґќ¶вВБ¬≤ќ¶вВВ¬≤ќ¶вВГќ¶вВЖвФВ
вФВGвВВвВНвВБвВЕвВО=AвВВ вФВ         (qвВГ-1)/2    q¬≥ќ¶вВБ¬≤ќ¶вВВќ¶вВГвФВ
вФВGвВВвВНвВБвВЕвВО=¬≤AвВВвФВ         (qвВГ-1)/2    q¬≥ќ¶вВБќ¶вВВ¬≤ќ¶вВЖвФВ
вФФвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФівФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФШ

We notice that if q is a power of 2 such that qвЙ°2 (mod 3), so that qвВГ=1, some class types do not exist. We can see what happens by giving a specific value to qвВГ:

julia> xdisplay(t(;q_3=1);nClasses=true)
ClassTypes(GвВВ,char. 2) qвВГ=1
вФМвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФђвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФР
вФВC_G(s)    вФВ     nClasses     |C_G(s)|вФВ
вФЬвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФЉвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФ
вФВGвВВвВНвВО=ќ¶вВБ¬≤  вФВ(q¬≤-8q+12)/12          ќ¶вВБ¬≤вФВ
вФВGвВВвВНвВО=ќ¶вВБќ¶вВВ вФВ    (q¬≤-2q)/4         ќ¶вВБќ¶вВВвФВ
вФВGвВВвВНвВО=ќ¶вВБќ¶вВВ вФВ    (q¬≤-2q)/4         ќ¶вВБќ¶вВВвФВ
вФВGвВВвВНвВО=ќ¶вВЖ   вФВ     (q¬≤-q)/6           ќ¶вВЖвФВ
вФВGвВВвВНвВО=ќ¶вВГ   вФВ     (q¬≤+q)/6           ќ¶вВГвФВ
вФВGвВВвВНвВО=ќ¶вВВ¬≤  вФВ   (q¬≤-4q)/12          ќ¶вВВ¬≤вФВ
вФВGвВВвВНвВБвВО=AвВБќ¶вВБвФВ      (q-2)/2       qќ¶вВБ¬≤ќ¶вВВвФВ
вФВGвВВвВНвВБвВО=AвВБќ¶вВВвФВ          q/2       qќ¶вВБќ¶вВВ¬≤вФВ
вФВGвВВвВНвВВвВО=AћГвВБќ¶вВБвФВ      (q-2)/2       qќ¶вВБ¬≤ќ¶вВВвФВ
вФВGвВВвВНвВВвВО=AћГвВБќ¶вВВвФВ          q/2       qќ¶вВБќ¶вВВ¬≤вФВ
вФВGвВВ        вФВ            1 qвБґќ¶вВБ¬≤ќ¶вВВ¬≤ќ¶вВГќ¶вВЖвФВ
вФФвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФівФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФАвФШ