Number fields

Gapjm.Nf

CyclotomicNumbers.galois

Gapjm.Nf.Aut

Gapjm.Nf.CF

Gapjm.Nf.NF

Gapjm.Nf — Module

Number fields are the finite extensions of ℚ. The only ones that we can handle at the moment are subfields of cyclotomic fields, that is whose elements are Cycs; they are also characterized as the number fields K such that Gal(K/ℚ) is abelian. For example, ℚ (√5) is a number field that is not cyclotomic but contained in the cyclotomic field ℚ (ζ₅).

The default constructor for a number field constructs the smallest number field containing its arguments.

julia> F=NF(E(5)) # the full cyclotomic field prints as CF
CF(5)

julia> K=NF(root(5)) # a subfield
NF(5,-1₅)

julia> conductor(K) # no. of smallest cyclotomic field containing K
5

julia> E(5)+E(5,-1) in NF(root(5)) # test if an element is in the subfield
true

Above in NF(5,-1₅) the -1₅ represents the stabilizer of K in the galois group of F.

julia> elements(galois(F))
4-element Vector{Gapjm.Nf.NFAut}:
 Aut(CF(5),1₅)
 Aut(CF(5),2₅)
 Aut(CF(5),-1₅)
 Aut(CF(5),-2₅)

An element of the galois group marked -2₅ acts by raising the fifth roots of unity to the power -2. Thus NF(5,-1₅) represents the subfield of CF(5) stable by complex conjugacy.

julia> NF(root(3),root(5)) # here the stabilizer needs 2 generators
NF(60,-11₆₀,-1₆₀)

Gapjm.Nf.NF — Function

NF(gens...) or NF(gens::AbstractVector)

return the smallest number field containing the elements gens (which can be Cyc, Root1, Integer, Rational{<:Integer}).

julia> NF(E(3),root(5))
NF(15,4₁₅)

julia> NF([E(3),root(5)])
NF(15,4₁₅)

NF(N::Integer, stab::Group{<:Mod}) fixed field of the subgroup stab of galois(CF(N)) in CF(N). stab should not inject in the multiplicative group of a proper divisor of N.

Gapjm.Nf.CF — Function

CF(N::Integer) the cyclotomic field generated by the N`-th roots of unity.

Gapjm.Nf.Aut — Function

Aut(F::NumberField,k::Union{Integer,Mod})

The Galois automorphisms of the cyclotomic field CF(n) are the linear extensions σₖ of raising n-th roots of unity to some power k prime to n. If F is a subfield of CF(n), a Galois automorphism σₖ is replaced by a canonical representative σₗ of its orbit modulo the stabilizer of F in the Galois group galois(CF(n)). This is the result of Aut(F,k). The number k can be given as an integer or as Mod(k,n).

julia> F=NF(root(5))
NF(5,-1₅)

julia> s=Aut(F,3)
Aut(NF(5,-1₅),2₅)

julia> root(5)^s # action of s on a Cyc
Cyc{Int64}: -√5

CyclotomicNumbers.galois — Method

galois(F::NumberField) galois of F over ℚ

the galois group of F, a number field of conductor n, is the quotient of the galois group of CF(n) (isomorphic to the multiplicative group (ℤ/n)ˣ) by the stabilizer of F. It is given as a group of galois automorphisms (see Aut).

julia> K=CF(5)
CF(5)

julia> F=NF(root(5))
NF(5,-1₅)

julia> galois(K)
Group(Gapjm.Nf.NFAut[Aut(CF(5),2₅)])

julia> elements(galois(K))
4-element Vector{Gapjm.Nf.NFAut}:
 Aut(CF(5),1₅)
 Aut(CF(5),2₅)
 Aut(CF(5),-1₅)
 Aut(CF(5),-2₅)

julia> elements(galois(F))
2-element Vector{Gapjm.Nf.NFAut}:
 Aut(NF(5,-1₅),1₅)
 Aut(NF(5,-1₅),2₅)

julia-repl