Finite reflection groups

Gapjm.PermRoot.Diagram

Gapjm.PermRoot.PRG

Gapjm.PermRoot.PermRootGroup

Gapjm.PermRoot.PermX

Gapjm.PermRoot.bipartite_decomposition

Gapjm.PermRoot.cartan

Gapjm.PermRoot.cartan

Gapjm.PermRoot.cartan

Gapjm.PermRoot.cartan

Gapjm.PermRoot.cartan

Gapjm.PermRoot.coroot

Gapjm.PermRoot.coroots

Gapjm.PermRoot.generic_order

Gapjm.PermRoot.inclusion

Gapjm.PermRoot.invariant_form

Gapjm.PermRoot.invariants

Gapjm.PermRoot.isparabolic

Gapjm.PermRoot.nhyp

Gapjm.PermRoot.nref

Gapjm.PermRoot.parabolic_closure

Gapjm.PermRoot.parabolic_reps

Gapjm.PermRoot.rank

Gapjm.PermRoot.reflchar

Gapjm.PermRoot.reflection

Gapjm.PermRoot.reflection_subgroup

Gapjm.PermRoot.reflection_subgroup

Gapjm.PermRoot.reflection_subgroup

Gapjm.PermRoot.reflectionmat

Gapjm.PermRoot.refleigen

Gapjm.PermRoot.reflrep

Gapjm.PermRoot.reflrep

Gapjm.PermRoot.reflrep

Gapjm.PermRoot.reflrep

Gapjm.PermRoot.refls

Gapjm.PermRoot.restriction

Gapjm.PermRoot.roots

Gapjm.PermRoot.roots

Gapjm.PermRoot.roots

Gapjm.PermRoot.semisimplerank

Gapjm.PermRoot.simple_conjugating

Gapjm.PermRoot.simple_reps

Gapjm.PermRoot.simplecoroots

Gapjm.PermRoot.simpleroots

Gapjm.PermRoot.torus_order

Gapjm.PermRoot — Module

Let V be a vector space over a subfield K of the complex numbers (usually in Julia elements of K are Rationals or Cycs). A complex reflection is an element s∈ GL(V) of finite order whose fixed point set is an hyperplane, the reflecting hyperplane Hₛ of s (we will in the following abbreviate "complex reflection" to reflection; in some literature "reflection" is reserved for the order 2 and otherwise pseudo-reflection is used). Thus a reflection s has a unique eigenvalue ζₛ not equal to 1, which is a root of unity (if K is a subfield of the real numbers, ζₛ is necessarily equal to -1). A reflection s is determined by a vector r∈ V (a root) and a linear form rᵛ on V (a coroot) such that rᵛ(r)=1-ζ. Then s is the map on V defined by s(x)=x-rᵛ(x)r. The root and the coroot for s are unique up to scaling the root and inversely scaling the coroot.

A complex reflection group W is a group generated by a finite number of complex reflections.

Since when W contains a reflection s it contains its powers, W is always generated by reflections s such that ζₛ==E(d) where d is the order of s; we may in addition assume that s is not a power of another reflection with larger order d. Such a reflection is called a distinguished reflection of W; it is a canonical generator of the cyclic subgroup of W centralizing Hₛ. The generators of the complex reflection groups we construct are always distinguished reflections (in a real reflection group all reflections are distinguished).

We deal in this module with finite complex reflection groups. Note that W does not need to be a matrix group. But by definition W has a faithful representation (the reflection representation) where reflections(W) act as reflections (see reflrep). Note that the set of distinguished reflections are in bijection with the reflecting hyperplanes.

It is much more efficient to compute with permutation groups than with matrix groups, when possible; thus the efficient representation that we use for computing with W, is a permutation representation on a W-invariant set of root and coroot vectors for reflections of W; that is, a set R of pairs (r,rᵛ)∈ V×Vᵛ invariant by W and such that each distinguished reflection in W is defined by some pair in R (see reflection). Trying to find a W-invariant set implies that there may be several pairs for each reflection, differing by roots of unity. This generalizes the usual construction for Coxeter groups (the case K=ℝ) where to each reflection of W is associated two roots, a positive and a negative one. For irreducible complex reflection groups, there are at least as many roots on a given line as the order of the center of W.

For these groups the following methods are defined:

gens(W): the (distinguished) reflections which generate W, given as permutations of the roots.

refls(W): the list of distinguished reflections of W, which starts with gens(W). Actually this list is of same length as the roots, and its i-th element is the distinguished reflection for the i-th root.

The finite irreducible complex reflection groups have been completely classified by Shepard and Todd. They consist of one infinite family G(de,e,r) depending on 3 parameters, and 34 exceptional groups which have been given by Shephard and Todd names which range from G₄ to G₃₇. These exceptional groups contain the exceptional Coxeter groups, e.g., coxgroup(:E,8) is the same as G₃₇.

We provide functions to build any finite reflection group, either by giving a list of roots and coroots defining the generating reflections (see PRG), or in terms of the classification (see complex_reflection_group). The output is a permutation group on set of roots . In the context e.g. of Weyl groups, one wants to describe the particular root system chosen in term of the traditional classification of crystallographic root systems. This is done via calls to the function coxgroup (see the module Weyl). There is not yet a general theory on how to construct a nice set of roots for a non-real reflection group; the roots chosen here where obtained case-by-case; however, they satisfy several important properties:

The generating reflections satisfy braid relations which present the braid group associated to W (see Diagram).

The field of definition of W is the field K generated by the traces of the elements of W acting on V. It is a theorem that W may be realized as a reflection group over K.

julia> W=complex_reflection_group(4)
G₄

julia> gens(W) # as a permutation of the 24 roots
2-element Vector{Perm{Int16}}:
 (1,3,9)(2,4,7)(5,10,18)(6,11,16)(8,12,19)(13,15,20)(14,17,21)(22,23,24)
 (1,5,13)(2,6,10)(3,7,14)(4,8,15)(9,16,22)(11,12,17)(18,19,23)(20,21,24)

julia> length(unique(refls(W)))
4

julia> length(refls(W)) # 24=4*(number of roots of unity in ℚ (ζ₃))
24

julia> reflrep(W) # the reflection representations
2-element Vector{Matrix{Cyc{Rational{Int64}}}}:
 [1 0; 0 ζ₃]
 [√-3/3 -ζ₃²√-3/3; (-3+√-3)/3 -ζ₃√-3/3]

julia> braid_relations(W) # the braid relation between generators
1-element Vector{Vector{Vector{Int64}}}:
 [[1, 2, 1], [2, 1, 2]]

julia> Diagram(W) # the same in pictures
③ ——③ G₄
1   2

For almost all irreducible complex reflection groups, the generating matrices for W we give have coefficients in K. Further, the set of matrices for all elements of W is globally invariant under the Galois group of K/ℚ, thus the Galois action induces automorphisms of W. The exceptions are G₂₂, G₂₇ where the matrices are in a degree two extension of K (this is needed to have a globally invariant model, see MarinMichel2010) and some dihedral groups, as well as H₃ and H₄, where the matrices given (the usual Coxeter reflection representation over K) are not globally invariant (but one can get invariant ones by chevieget(:H4,:InvariantModel)).

It turns out that all representations of W are defined over the field of definition K (see. Benard1976 and D.~Bessis' thesis). This has been known for a long time in the case K=ℚ, the case of Weyl groups, which have integral representations.

The Cartan matrix (see cartan) for the generating roots (those which correspond to the generating reflections) has entries in the ring ℤₖ of integers of K, and the roots (resp. coroots) are linear combination with coefficients in ℤₖ of a linearly independent subset of them.

Some additional methods for finite reflection groups:

roots(W): a set of complex roots in V, given as a list of vectors, on which W has a faithful permutation representation.

simpleroots(W): the generating roots as a matrix.

simplecoroots(W): the generating coroots as a matrix.

julia> cartan(W)
2×2 Matrix{Cyc{Rational{Int64}}}:
 ζ₃²√-3     ζ₃²
   -ζ₃²  ζ₃²√-3

julia> simpleroots(W)
2×2 Matrix{Cyc{Rational{Int64}}}:
    0  ζ₃²√-3
 2ζ₃²     ζ₃²

julia> simplecoroots(W)
2×2 Matrix{Cyc{Rational{Int64}}}:
     0      1
 √-3/3  √-3/3

Let SV be the symmetric algebra of V. The invariants of W in SV are called the polynomial invariants of W. They are generated as a polynomial ring by dim V homogeneous algebraically independent polynomials f₁,…,f_{dim V}. The polynomials fᵢ are not uniquely determined but their degrees are. The fᵢ are called the basic invariants of W, and their degrees the reflection degrees of W; the function degrees(W) returns these invariants.

Let I be the ideal generated by the homogeneous invariants of positive degree in SV. Then SV/I is isomorphic to the regular representation of W as a W-module. It is thus a graded (by the degree of elements of SV) version of the regular representation of W. The polynomial which gives the graded multiplicity of a character φ of W in the graded module SV/I is called the fake degree of φ. The function fakedegrees returns these polynomials.

julia> degrees(W)
2-element Vector{Int64}:
 4
 6

julia> fakedegrees(W,Pol(:x))
7-element Vector{Pol{Int64}}:
 1
 x⁴
 x⁸
 x⁷+x⁵
 x⁵+x³
 x³+x
 x⁶+x⁴+x²

Gapjm.PermRoot.PermRootGroup — Type

PermRootGroup is the type of reflection group represented as a permutation group on a set of roots.

Gapjm.ComplexR.complex_reflection_group — Function

complex_reflection_group(STnumber) or crg(STnumber)

complex_reflection_group(p,q,r) or crg(p,q,r)

The first form of complex_reflection_group returns the complex reflection group which has Shephard-Todd number STnumber, see Shephard-Todd1954. The second form returns the imprimitive complex reflection group G(p,q,r).

julia> G=complex_reflection_group(4)
G₄

julia> degrees(G)
2-element Vector{Int64}:
 4
 6

julia> length(G)
24

julia> W*coxgroup(:A,2) # how to make a non-irreducible group
G₄×A₂

julia> complex_reflection_group(1,1,3) # another way to enter A₂
A₂

julia> crg(4) # there is also a short alias
G₄

Gapjm.PermRoot.PRG — Type

PRG(roots,coroots) complex reflection group with given roots and coroots

Gapjm.PermRoot.rank — Function

rank(W::ComplexReflectionGroup)

This function returns the rank of a reflection group, which is the dimension of the space where it acts.

julia> rank(complex_reflection_group(31))
4

Gapjm.PermRoot.semisimplerank — Function

semisimpleRank(W::ComplexReflectionGroup)

This function returns the semisimple rank of the reflection group W, which is the dimension of the space where it effectively acts. If Wis a reflection group acting on V, and V₁ is the subspace generated by the roots(W), then the dimension of V₁ is the semisimple rank. W is called essential if V₁=V.

julia> W=reflection_subgroup(coxgroup(:A,3),[1,3])
A₃₍₁₃₎=A₁×A₁Φ₁

julia> semisimplerank(W)
2

julia> rank(W)
3

Gapjm.PermRoot.nref — Function

nref(W::ComplexReflectionGroup) the number of reflections of W

Gapjm.PermRoot.nhyp — Function

nhyp(W::ComplexReflectionGroup)

The number of reflecting hyperplanes of W

Gapjm.PermRoot.roots — Method

roots(W::PermRootGroup) the roots of W

Gapjm.PermRoot.roots — Method

roots(W::PermRootGroup,i) same as but better than roots(W)[i]

Gapjm.PermRoot.coroots — Function

coroots(W) the list of coroots of W (listed in the same order as the roots)

coroots(W,i) same as but better than coroots(W)[i]

Gapjm.PermRoot.coroot — Function

coroot(r,λ=1)

the coroot for an orthogonal reflection of root r and eigenvalue λ.

Gapjm.PermRoot.inclusion — Function

inclusion(W::PermRootGroup)

the indices of the roots of W in the roots of parent(W).

inclusion(W::PermRootGroup,i::Integer) inclusion(W::PermRootGroup,v::AbstractVector{<:Integer})

same as inclusion(W)[i] or inclusion(W)[v] (but more efficient).

Gapjm.PermRoot.restriction — Function

restriction(W::PermRootGroup)

A list for each root of parent(W), which hold 0if the root is not a root of W and i if the root is the i-th root of W.

restriction(W::PermRootGroup,i::Integer) restriction(W::PermRootGroup,v::AbstractVector{<:Integer})

same as restriction(W)[i] or restriction(W)[v] (but more efficient).

Gapjm.CoxGroups.degrees — Method

degrees(W::ComplexReflectionGroup)

returns a list holding the degrees of W as a reflection group on the vector space V on which it acts. These are the degrees d₁,…,dₙ where n is the dimension of V of the basic invariants of W in SV. They reflect various properties of W; in particular, their product is the cardinality of W.

julia> W=complex_reflection_group(30)
H₄

julia> degrees(W)
4-element Vector{Int64}:
  2
 12
 20
 30

julia> length(W)
14400

Gapjm.ComplexR.codegrees — Function

codegrees(W::ComplexReflectionGroup)

returns the vector of codegrees of W as a reflection group on the space V of reflrep(W). These are one less than the degrees of the basic derivations of W on SV⊗ V^vee.

julia> W=complex_reflection_group(4)
G₄

julia> codegrees(W)
2-element Vector{Int64}:
 0
 2

Gapjm.PermRoot.cartan — Method

cartan(W::PermRootGroup) Cartan matrix of W.

Let s₁,…,sₙ be reflections with associated roots rᵢ and coroots cᵢ. The matrix C with entries Cᵢ,ⱼ=cᵢ(rⱼ) is called a Cartan matrix of s₁,…,sₙ. It is uniquely determined by s₁,…,sₙ up to conjugation by a diagonal matrix.

If s₁,…,sₙ generate a reflection group W, then C up to conjugation by a diagonal matrix is an invariant of the reflection representation of W. If invertible, the matrix C determines this representation since then the rᵢ form a basis in which the matrix for sᵢ differs from the identity only on the i-th line, where the corresponding line of C has been subtracted.

julia> W=coxgroup(:A,3)
A₃

julia> cartan(W)
3×3 Matrix{Int64}:
  2  -1   0
 -1   2  -1
  0  -1   2

Gapjm.PermRoot.cartan — Method

cartan(W::PermRootGroup,i,j) the cartan coefficient cᵢ(rⱼ), the value of the linear form given by the i-th coroot cᵢ on the j-th root rᵢ of W

Gapjm.PermRoot.simpleroots — Function

simpleroots(W::ComplexReflectionGroup) the simple roots of W (those corresponding to gens(W)) as a matrix (each root is a row)

Gapjm.PermRoot.simplecoroots — Function

simplecoroots(W::ComplexReflectionGroup) the simple coroots of W (those corresponding to gens(W)) as a matrix (each coroot is a row)

Gapjm.PermRoot.reflection — Function

reflection(s::Matrix [,r::AbstractVector])

s should be is a square matrix, and if given r should be a vector of length size(s,1). The function determines if s is the matrix of a reflection (resp. if r is given if it is the matrix of a reflection of root r; the point of giving r is to specify exactly the desired root and coroot, which otherwise are determined only up to a scalar and its inverse). The function returns nothing if s if is not a reflection (resp. not a reflection with root r), and otherwise returns a named tuple with four fields:

.root: the root of the reflection s (equal to r if given)

.coroot: the coroot of s

.eigenvalue: the non-trivial eigenvalue of s

.isOrthogonal: a boolean which is true if and only if s is orthogonal with respect to the usual scalar product (then s is determined by the root and the eigenvalue as reflectionmat(.root,.eigenvalue))

julia> reflection([-1 0 0;1 1 0;0 0 1])
(root = [2, 0, 0], coroot = Rational{Int64}[1//1, -1//2, 0//1], eig = -1, isOrthogonal = false)

julia> reflection([-1 0 0;1 1 0;0 0 1],[1,0,0])
(root = [1, 0, 0], coroot = Rational{Int64}[2//1, -1//1, 0//1], eig = -1, isOrthogonal = false)

Gapjm.PermRoot.reflectionmat — Function

reflectionmat(root, coroot) the matrix of the reflection with a given root and coroot.

A (complex) reflection is a finite order element s of GL(V), the linear group of a vector space over a subfield of the complex numbers, whose fixed points form a hyperplane H (the reflecting hyperplane of s); an eigenvector r for the non-trivial eigenvalue ζ (a root of unity) is called a root of s. If we choose a linear form rᵛ defining H such that rᵛ(r)=1-ζ (a coroot of s) then s is given by x↦ x-rᵛ(x)r.

A way of specifying s is by giving a root and a coroot, which are uniquely determined by s up to multiplication of the root by a scalar and of the coroot by the inverse scalar. The function reflectionmat gives the matrix of the corresponding reflection in the standard basis of V, where the root and the coroot are vectors given in the standard bases of V and Vᵛ, so the pairing rᵛ(r) is obtained as transpose(root)*coroot.

julia> r=reflectionmat([1,0,0],[2,-1,0])
3×3 Matrix{Int64}:
 -1  0  0
  1  1  0
  0  0  1

julia> r==reflrep(coxgroup(:A,3),1)
true

julia> r*[2,-1,0]
3-element Vector{Int64}:
 -2
  1
  0

julia> [1 0 0]*r
1×3 Matrix{Int64}:
 -1  0  0

As we see in the last lines, in Julia a matrix operates from the right on the vector space V of row vectors and from the left on the dual space Vᵛ of column vectors.

reflectionmat(r, ζ=-1)

returns the matrix of the complex reflection determined by the root r and the eigenvalue ζ when the vector space and its dual are identified via the scalar product <x,y>=transpose(x)*y; the coroot rᵛ is then equal to the linear form x->(1-ζ)<x,r>/<r,r>.

julia> reflectionmat([1,0,-E(3,2)])
3×3 Matrix{Cyc{Rational{Int64}}}:
  0  0  ζ₃²
  0  1    0
 ζ₃  0    0

Gapjm.PermRoot.refls — Function

refls(W::ComplexReflectionGroup)

a list of same length as W.roots giving the corresponding distinguished reflections. In particular this list is much longer than unique(refls(W)) since in general there are several roots corresponding to a reflection.

refls(W,i)

reflection for i-th root(s) of W (i can be an index or a vector of indices)

PermGroups.Perms.reflength — Method

reflength(W::PermRootGroup,w::Perm)

This function returns the number of eigenvalues of w in the reflection representation which are not equal to 1. For a finite Coxeter group, this is equal to the reflection length of w, that is the minimum number of reflections of which w is a product. This also holds in general in a well-generated complex reflection group if w divides a Coxeter element for the reflection length.

julia> W=coxgroup(:A,4)
A₄

julia> reflength(W,longest(W))
2

julia> reflength(W,W(1,2,3,4))
4

Gapjm.PermRoot.simple_conjugating — Function

simple_conjugating(W)

For each index i of a root, an element w∈ W such that restriction(W,inclusion(W,simple_reps(W,i))^w)==i (or action(W,simple_reps(W,i))==i)

simple_conjugating(W,i)

an element w∈ W such that restriction(W,inclusion(W,simple_reps(W,i))^w)==i (or action(W,simple_reps(W,i))==i)

Gapjm.PermRoot.simple_reps — Function

simple_reps(W)

for each root, the index of the first simple root conjugate to it

simple_reps(W,i)

the smallest index of a root in the same W-orbit as the i-th root.

Gapjm.PermRoot.Diagram — Type

Diagram(W) diagram of finite reflection group W

julia> Diagram(crg(33))
G₃₃     3
       /^1---2---4---5 423423==342342

Gapjm.ComplexR.hyperplane_orbits — Function

hyperplane_orbits(W)

returns a list of named tuples, one for each hyperplane orbit of the reflection group W. If o is the named tuple for such an orbit, and s is the first element of gens(W) whose hyperplane is in the orbit, it contains the following fields

o.s: index of s in gens(W)

o.cl_s: map(i->position_class(W,s^i),1:o.order-1)

o.order: order of s

.N_s: Size of orbit

.det_s: for i in 1:o.order-1, position in CharTable of (det_s)^i

julia> W=coxgroup(:B,2)
B₂

julia> hyperplane_orbits(W)
2-element Vector{NamedTuple{(:s, :cl_s, :order, :N_s, :det_s), Tuple{Int64, Vector{Int64}, Int64, Int64, Vector{Int64}}}}:
 (s = 1, cl_s = [2], order = 2, N_s = 2, det_s = [5])
 (s = 2, cl_s = [4], order = 2, N_s = 2, det_s = [1])

Gapjm.ComplexR.Reflection — Type

Reflection is a struct representing a reflection r in a reflection group W. The fields are W, the index of a root for r, the non-trivial eigenvalue of r, and a word for r in the generating reflections of W.

julia> r=reflections(crg(8))[2]
Reflection(G₈,1,-1)

julia> r.eigen # the non-trival eigenvalue, as a Root1
Root1: -1

julia> root(r)
2-element Vector{Cyc{Rational{Int64}}}:
  0
 ζ₄

julia> coroot(r)
2-element Vector{Cyc{Rational{Int64}}}:
    0
 -2ζ₄

julia> Matrix(r)
2×2 Matrix{Cyc{Rational{Int64}}}:
 1   0
 0  -1

julia> isdistinguished(r) # r is not distinguished
false

julia> exponent(r) # which power of a distinguished reflection it is
2

julia> Perm(r)
(1,8)(2,9)(3,16)(4,15)(5,17)(6,18)(7,19)(10,22)(11,21)(12,23)

julia> hyperplane_orbit(r) # r is in the first hyperplane orbit
1

julia> position_class(r) # the index of the conjugacy class of r in W 
15

julia> word(r) # a word in the generators for r
2-element Vector{Int64}:
 1
 1

Gapjm.ComplexR.reflections — Function

reflections(W) the list of all reflections of the reflection group W (including the non-distinguished ones), given as a Vector{Reflection} (see Reflection).

julia> W=crg(4)
G₄

julia> reflections(W)
8-element Vector{Reflection{PRG{Cyc{Rational{Int64}}, Int16}}}:
 Reflection(G₄,1,ζ₃)
 Reflection(G₄,1,ζ₃²)
 Reflection(G₄,2,ζ₃)
 Reflection(G₄,2,ζ₃²)
 Reflection(G₄,4,ζ₃)
 Reflection(G₄,4,ζ₃²)
 Reflection(G₄,5,ζ₃)
 Reflection(G₄,5,ζ₃²)

Gapjm.CoxGroups.braid_relations — Function

braid_relations(W)

this function returns the relations which present the braid group of the reflection group W. These are homogeneous (both sides of the same length) relations between generators in bijection with the generating reflections of W. A presentation of W is obtained by adding relations specifying the order of the generators.

julia> W=complex_reflection_group(29)
G₂₉

julia> braid_relations(W)
7-element Vector{Vector{Vector{Int64}}}:
 [[1, 2, 1], [2, 1, 2]]
 [[2, 4, 2], [4, 2, 4]]
 [[3, 4, 3], [4, 3, 4]]
 [[2, 3, 2, 3], [3, 2, 3, 2]]
 [[1, 3], [3, 1]]
 [[1, 4], [4, 1]]
 [[4, 3, 2, 4, 3, 2], [3, 2, 4, 3, 2, 4]]

each relation is represented as a pair of lists, specifying that the product of the generators according to the indices on the left side is equal to the product according to the indices on the right side. See also Diagram.

Gapjm.PermRoot.reflrep — Method

reflrep(W,w) Reflection representation

Let W be a finite reflection group on the space V and let w be a permutation of the roots of W. The function reflrep returns the matrix of w acting on V (recall that matrices operate from the right on a vector space in Gapjm). This is the linear transformation of V which acts trivially on the orthogonal of the coroots and has same effect as w on the simple roots. The function makes sense more generally for an element of the normalizer of W in the whole permutation group of the roots.

julia> W=reflection_subgroup(rootdatum("E7sc"),1:6)
E₇₍₁₂₃₄₅₆₎=E₆Φ₁

julia> reflrep(W,longest(W))
7×7 Matrix{Int64}:
  0   0   0   0   0  -1  2
  0  -1   0   0   0   0  2
  0   0   0   0  -1   0  3
  0   0   0  -1   0   0  4
  0   0  -1   0   0   0  3
 -1   0   0   0   0   0  2
  0   0   0   0   0   0  1

Gapjm.PermRoot.reflrep — Method

reflrep(W::ComplexReflectionGroup) returns the list of reflrep(W,x) for x in gens(W), that is the generators of W as matrices.

Gapjm.PermRoot.reflrep — Method

reflrep(W,i) same as but better than reflrep(W,W(i))

Gapjm.PermRoot.refleigen — Function

refleigen(W) Reflection eigenvalues

Let W be a reflection group on the vector space V. refleigen(W) returns for each conjugacy class representative x of W (see classreps) the eigenvalues of x on V, as a list of Root1.

julia> refleigen(coxgroup(:B,2))
5-element Vector{Vector{Root1}}:
 [1, 1]
 [-1, 1]
 [-1, -1]
 [-1, 1]
 [ζ₄³, ζ₄]

refleigen(W,i) faster than refleigen(W)[i]

Gapjm.PermRoot.bipartite_decomposition — Function

bipartite_decomposition(W)

Returns a bipartite decomposition [L,R] of the indices of the generators of the reflection group W, such that reflection_subgroup(W,L) and reflection_subgroup(W,R) are abelian subgroups, and W=reflection_subgroup(W, vcat(L,R)). Gives an error if no such decomposition is possible.

julia> bipartite_decomposition(coxgroup(:E,8))
([1, 4, 6, 8], [3, 2, 5, 7])

Combinat.catalan — Method

Catalan(W::ComplexReflectionGroup)

returns the Catalan Number of the irreducible complex reflection group W. For well-generated groups, this number is equal to the number of simples in the dual Braid monoid. For other groups it was defined by Gordon and Griffeth2012. For Weyl groups, it also counts the number of antichains of roots.

julia> catalan(coxgroup(:A,7))
1430

Catalan(W,i)

returns the i-th Fuss-Catalan Number of the irreducible complex reflection group W. For well-generated groups, this number is equal to the number of chains s₁,…,sᵢ of simples in the dual monoid where sⱼ divides sⱼ₊₁. For these groups, it is also equal to ∏ⱼ(ih+dⱼ)/dⱼ where the product runs over the reflection degrees of W, and where h is the Coxeter number of W. For non-well generated groups, the definition is in Gordon and Griffeth2012.

julia> catalan(complex_reflection_group(7),2)
16

Catalan(W;q=1), resp. Catalan(W,i;q=1)

for q a variable (like Pol() or an 'Mvp') returns the q-Catalan number (resp. the i-th q-Fuss Catalan number) of W. Again the definitions in general are in Gordon and Griffeth2012.

julia> catalan(complex_reflection_group(7),2;q=Pol())
Pol{Int64}: q⁷²+2q⁶⁰+3q⁴⁸+4q³⁶+3q²⁴+2q¹²+1

Gapjm.PermRoot.PermX — Function

PermX(W::ComplexReflectionGroup,M::AbstractMatrix)

Let M be an invertible linear map of the reflection representation of W which preserves the set of roots of parent(W), and normalizes W (for the action of matrices on the right). PermX returns the corresponding permutation of the roots of parent(W); it returns nothing if M does not normalize the set of roots of parent(W).

julia> W=reflection_subgroup(rootdatum("E7sc"),1:6)
E₇₍₁₂₃₄₅₆₎=E₆Φ₁

julia> PermX(W,reflrep(W,longest(W)))==longest(W)
true

Gapjm.PermRoot.reflchar — Function

reflchar(W::ComplexReflectionGroup,w) Reflection character

Returns the trace of the element w of W as an endomorphism of the vector space V on which W acts. This is the same as trace(reflrep(W,w)).

julia-repl``` julia> W=coxgroup(:B,3) B₃

julia> reflchar(W,longest(W)) -3 julia-repl```

reflchar(W::ComplexReflectionGroup)

Returns the reflection character of W. This is the same as map(c->reflchar(W,c),class-reps(W)). When W is irreducible, it is also CharTable(W).irr[charinfo(W).extRefl[2]].

julia> reflchar(coxgroup(:A,3))
5-element Vector{Int64}:
  3
  1
 -1
  0
 -1

Gapjm.PermRoot.invariant_form — Function

invariant_form(W::ComplexReflectionGroup)

This function returns the matrix F of an Hermitian form invariant under the action of the reflection group W. That is, if M is the matrix of an element of W, then M*F*M'=F.

julia> W=complex_reflection_group(4)
G₄

julia> invariant_form(W)
2×2 Matrix{Int64}:
 1  0
 0  2

Gapjm.PermRoot.invariants — Function

invariants(W::ComplexReflectionGroup)

returns the fundamental invariants of W in its reflection representation V. That is, returns a set of algebraically independent elements of the symmetric algebra of the dual of V which generate the W-invariant polynomial functions on V. Each such invariant function is returned as a function: if e₁,…,eₙ is a basis of V and f is the function, then the value of the polynomial function on a₁e₁+…+aₙeₙ is obtained by calling f(a₁,…,aₙ). This function depends on the classification, and is dependent on the exact reflection representation of W. So for the moment it is only implemented when the reflection representation for the irreducible components has the same Cartan matrix as the one provided by Chevie for the corresponding irreducible group. The polynomials are invariant for the natural action of the group elements as matrices; that is, if m==reflrep(W,w) for some w in W, then an invariant f satisfies f(a₁,…,aₙ)=f(v₁,…,vₙ) where [v₁,…,vₙ]=[a₁,…,aₙ]×m. This action is implemented on Mvps by the function ^.

julia> W=coxgroup(:A,2)
A₂

julia> @Mvp x,y,z

julia> i=invariants(W);

julia> i[1](x,y)
Mvp{Int64}: -2x²+2xy-2y²

julia> i[2](x,y)
Mvp{Int64}: 6x²y-6xy²

julia> W=complex_reflection_group(24)
G₂₄

julia> i=invariants(W)[1];

julia> p=i(x,y,z)
Mvp{Rational{Int64}}: (14//1)x⁴+(-12//1)x²y²+(-42//1)x²yz+(21//2)x²z²+(18//7)y⁴+(-6//1)y³z+(-9//2)y²z²+(-21//8)z⁴

julia> p^reflrep(W,1)-p
Mvp{Cyc{Rational{Int64}}}: 0

Gapjm.PermRoot.generic_order — Function

generic_order(W,q)

returns the "compact" generic order of W as a polynomial in q. This is q^(Nₕ)Πᵢ(q^(dᵢ)-1) where dᵢ are the reflection degrees and Nₕ the number of reflecting hyperplanes. For a Weyl group, it is the order of the associated semisimple finite reductive group over the field with q elements.

julia> @Pol q
Pol{Int64}: q

julia> PermRoot.generic_order(complex_reflection_group(4),q)
Pol{Int64}: q¹⁴-q¹⁰-q⁸+q⁴

Gapjm.PermRoot.torus_order — Function

torus_order(W::ComplexReflectionGroup,i,q)

returns as a polynomial in q the toric order of the i-th conjugacy class of W. This is the characteristic polynomial of an element of that class on the reflection representation of W. It is the same as the generic order of the reflection subcoset torus(W,i) of W determined by the trivial subgroup and a representative of the i-th conjugacy class.

julia> W=complex_reflection_group(4)

julia> torus_order.(Ref(W),1:nconjugacy_classes(W),Pol(:q))
7-element Vector{Pol{Cyc{Int64}}}:
 q²-2q+1
 q²+2q+1
 q²+1
 q²-ζ₃q+ζ₃²
 q²+ζ₃q+ζ₃²
 q²+ζ₃²q+ζ₃
 q²-ζ₃²q+ζ₃

Gapjm.Weyl.standard_parabolic — Method

standard_parabolic(W::PermRootGroup, H)

Let H be a reflection subgroup of W. Returns an element w∈ W such that H^w is a standard parabolic subgroup of W (or nothing if H is not parabolic or not conjugate to a standard parabolic).

Gapjm.PermRoot.parabolic_reps — Function

parabolic_reps(W)

Returns a list of indices of roots of W describing representatives of orbits of parabolic subgroups under conjugation by W. For Coxeter groups, each orbit has a representative whose indices is a subset of eachindex(gens(W)). This may not be the case in general.

julia> parabolic_reps(coxgroup(:A,4))
7-element Vector{Vector{Int64}}:
 []
 [1]
 [1, 2]
 [1, 3]
 [1, 2, 3]
 [1, 2, 4]
 [1, 2, 3, 4]

julia> parabolic_reps(complex_reflection_group(3,3,3))
7-element Vector{Vector{Int64}}:
 []
 [1]
 [1, 2]
 [1, 3]
 [1, 9]
 [2, 3]
 [1, 2, 3]

parabolic_reps(W,r)

If a second argument r is given, returns only representatives of the parabolic subgroups of semisimple rank r.

julia> parabolic_reps(coxgroup(:A,4),2)
2-element Vector{Vector{Int64}}:
 [1, 2]
 [1, 3]

julia> parabolic_reps(complex_reflection_group(3,3,3),2)
4-element Vector{Vector{Int64}}:
 [1, 2]
 [1, 3]
 [1, 9]
 [2, 3]

Gapjm.PermRoot.parabolic_closure — Function

parabolic_closure(W,I)

I should be a list of indices of reflections of W. Returns J such that reflection_subgroup(W,J) is the smallest parabolic subgroup of W containing reflection_subgroup(W,I).

julia> W=complex_reflection_group(7)
G₇

julia> parabolic_closure(W,[1])
1-element Vector{Int64}:
 1

julia> parabolic_closure(W,[1,2])
3-element Vector{Int64}:
 1
 2
 3

Gapjm.PermRoot.isparabolic — Function

isparabolic(W)

whether the reflection subgroup W is a parabolic subgroup of parent(W).

julia> W=complex_reflection_group(7)
G₇

julia> isparabolic(reflection_subgroup(W,[1,2]))
false

julia> isparabolic(reflection_subgroup(W,[1]))
true