Posets

Gapjm.Posets.Poset

Gapjm.Posets.Poset

Gapjm.Posets.covering_chains

Gapjm.Posets.hasse

Gapjm.Posets.incidence

Gapjm.Posets.isjoinlattice

Gapjm.Posets.ismeetlattice

Gapjm.Posets.linear_extension

Gapjm.Posets.moebius

Gapjm.Posets.moebiusmatrix

Gapjm.Posets.partition

Gapjm.Posets.restricted

Gapjm.Posets.transitive_closure

Gapjm.Posets — Module

Posets are constructed from one of the two following fields:

incidence: a boolean matrix such that incidence[i,j]==true iff i<=j in the poset. This is sometimes called the ζ matrix of P.

hasse: a list representing the Hasse diagram of the poset: the i-th entry is the list of indices of elements which cover (are immediate successors of) the i-th element, that is the list of j such that i<j and there is no k such that i<k<j.

By default a Poset P is a poset on 1:length(P) where length(P) is the cardinality of P. To make a poset with other elements, the elements should be given in the constructor. There are thus two constructors,

Poset(I::Matrix{Bool}[,elements]) from the incidence matrix I. If given one should have length(elements)==size(I,1)==size(I,2).

Poset(H::Vector{<:Vector{<:Integer}}[,elements]) from the Hasse diagram. If given one should have length(elements)==length(H).

For convenience there is another constructor

Poset(isless::Function,elements); this begins by constructing the incidence matrix from the isless function applied to each pair of elements which may be expensive. For isless one can give either a function implementing < or a function implementing ≤ (it is ored with = in any case).

julia> l=vec(collect(Iterators.product(1:2,1:2)))
4-element Vector{Tuple{Int64, Int64}}:
 (1, 1)
 (2, 1)
 (1, 2)
 (2, 2)

julia> P=Poset((x,y)->all(map(<=,x,y)),l)
1<2,3<4

julia> length(P) # the number of elements of the Poset
4

Posets are printed at the REPL as a list of covering chains. Elements which are equivalent for the Poset are printed together separated by commas.

By default printing shows the index in 1:length(P) for the elements. This can be changed by giving the keyword argument show=:elements.

julia> P=Poset((x,y)->all(map(<=,x,y)),l;show=:elements)
(1, 1)<(2, 1),(1, 2)<(2, 2)

julia> hasse(P)
4-element Vector{Vector{Int64}}:
 [2, 3]
 [4]
 [4]
 []

julia> incidence(P)
4×4 Matrix{Bool}:
 1  1  1  1
 0  1  0  1
 0  0  1  1
 0  0  0  1

More flexibility on printing is obtained by setting a function show_element which takes as arguments an IO, the poset, and the index of the element to print:

julia> P.show_element=(io,x,n)->join(io,x.elements[n],".");

julia> P
1.1<2.1,1.2<2.2

The syntax p[a:b] gives the interval between a and b in the poset.

julia> p=Poset((i,j)->j%i==0,1:6)
1<5
1<2<4
1<3<6
2<6

julia> p[2:6]   # elements between 2 and 6
2-element Vector{Int64}:
 2
 6

julia> p[2:end]  # everything above 2
3-element Vector{Int64}:
 2
 4
 6

julia> p[begin:6] # everything below 6
4-element Vector{Int64}:
 1
 2
 3
 6

julia> p[2:3] # nothing between 2 and 3
Int64[]

see the on-line help on linear_extension, hasse, incidence, partition, covering_chains, transitive_closure, isjoinlattice, ismeetlattice, moebius, moebiusmatrix,reverse, restricted, minimum, maximum for more information

Gapjm.Posets.Poset — Type

Poset(m::Matrix{Bool})

Creates a poset from an incidence matrix m, that is m[i,j]==true if and only if i≤j in the poset,

julia> Poset(Bool[1 1 1 1 1;0 1 0 1 1;0 0 1 1 1;0 0 0 1 0;0 0 0 0 1])
1<2,3<4,5

Poset(h::Vector{<:Vector{<:Integer}})

Creates a poset from a Hasse diagram given as a Vector whose i-th entry is the list of indices of elements which are immediate successors (covers) of the i-th element, that is h[i] is the list of j such that i<j in the poset and such that there is no k such that i<k<j.

julia> Poset([[2,3],[4,5],[4,5],Int[],Int[]])
1<2,3<4,5

Poset(W::CoxeterGroup,w=longest(W))

returns as a poset the Bruhat interval [1,w]of W. If w is not given, the whole Bruhat Poset of W is returned (W must then be finite).

julia> W=coxgroup(:A,2)
A₂

julia> Poset(W)
.<1,2<21,12<121

The above poset is constructed efficiently by constructing the Hasse diagram, but it could be constructed naively as follows:

julia> Poset(W)
.<1,2<21,12<121

julia> p=Poset((x,y)->bruhatless(W,x,y),elements(W))
1<2,3<4,5<6

The element printing is not so nice. This can be remedied by giving a function:

julia> p.show_element=(io,x,n)->(e=x.elements[n];isone(e) ? print(io,".") : print(io,joindigits(word(W,e))));

julia> p
.<2,1<21,12<121

julia> W=coxgroup(:A,3)
A₃

julia> Poset(W,W(1,3))
.<3,1<13

Gapjm.Posets.hasse — Function

hasse(P::Poset)

the Hasse diagram of P.

julia> p=Poset((i,j)->j%i==0,1:5)
1<3,5
1<2<4

julia> hasse(p)
5-element Vector{Vector{Int64}}:
 [2, 3, 5]
 [4]      
 []       
 []       
 []       

Gapjm.Posets.incidence — Function

incidence(P::Poset)

returns the incidence matrix (also called the ζ matrix) of P.

julia> p=Poset([i==6 ? Int[] : [i+1] for i in 1:6])
1<2<3<4<5<6

julia> incidence(p)
6×6 Matrix{Bool}:
 1  1  1  1  1  1
 0  1  1  1  1  1
 0  0  1  1  1  1
 0  0  0  1  1  1
 0  0  0  0  1  1
 0  0  0  0  0  1

Gapjm.Posets.transitive_closure — Function

transitive_closure(M)

M should be a square boolean matrix representing a relation; returns a boolean matrix representing the transitive closure of this relation. The transitive closure is computed by the Floyd-Warshall algorithm, which is quite fast even for large matrices.

julia> m=[j-i in [0,1] for i in 1:5, j in 1:5]
5×5 Matrix{Bool}:
 1  1  0  0  0
 0  1  1  0  0
 0  0  1  1  0
 0  0  0  1  1
 0  0  0  0  1

julia>transitive_closure(m)
5×5 Matrix{Bool}:
 1  1  1  1  1
 0  1  1  1  1
 0  0  1  1  1
 0  0  0  1  1
 0  0  0  0  1

Gapjm.Posets.linear_extension — Function

linear_extension(P)

returns a linear extension of the poset P, that is a vector l containing a permutation of the integers 1:length(P) such that if i<j in P, then findfirst(==(i),l)<findfirst(==(j),l). This is also called a topological sort of P.

julia> p=Poset((i,j)->j%i==0,1:6)
1<5
1<2<4
1<3<6
2<6

julia> linear_extension(p)
6-element Vector{Int64}:
 1
 2
 3
 5
 4
 6

Base.reverse — Function

reverse(P::Poset)

the opposed poset to P.

julia> p=Poset((i,j)->i%4<j%4,1:8)
4,8<1,5<2,6<3,7

julia> reverse(p)
3,7<2,6<1,5<4,8

Gapjm.Posets.partition — Function

partition(P::Poset)

returns the partition of 1:length(P) determined by the equivalence relation associated to P; that is, i and j are in the same part of the partition if the k such that i<k and j<k are the same as well as the k such that k<i and k<j.

julia> p=Poset([i==j || i%4<j%4 for i in 1:8, j in 1:8])
4,8<1,5<2,6<3,7

julia> partition(p)
4-element Vector{Vector{Int64}}:
 [4, 8]
 [2, 6]
 [3, 7]
 [1, 5]

Gapjm.Posets.covering_chains — Function

covering_chains(P::Poset)

A (greedy: the first is longest possible) list of covering chains for P.

Gapjm.Posets.restricted — Method

restricted(P::Poset,inds::AbstractVector{<:Integer};show=:indices)

returns the sub-poset of P determined by inds, which must be a sublist of 1:length(P). The indices in this sub-poset will be renumbered to their position in inds, but the elements set to P.elements[inds]. The keyword show is transmitted to the constructed poset.

julia> p=Poset((i,j)->i%4<j%4,1:8)
4,8<1,5<2,6<3,7

julia> restricted(p,2:6)
3<4<1,5<2

julia> restricted(p,2:6;show=:elements)
4<5<2,6<3

Gapjm.Posets.isjoinlattice — Function

isjoinlattice(P::Poset)

returns true if P is a join semilattice, that is any two elements of P have a unique smallest upper bound; returns false otherwise.

julia> p=Poset((i,j)->j%i==0,1:8)
1<5,7
1<2<4<8
1<3<6
2<6

julia> isjoinlattice(p)
false

Gapjm.Posets.ismeetlattice — Function

ismeetlattice(P)

returns true if P is a meet semilattice, that is any two elements of P have a unique highest lower bound; returns false otherwise.

julia> p=Poset((i,j)->j%i==0,1:8)
1<5,7
1<2<4<8
1<3<6
2<6

julia> ismeetlattice(p)
true

Gapjm.Posets.moebius — Function

moebius(P,y=maximum(P))

the vector of values μ(x,y) of the Moebius function of P for x varying. Here is an example giving the ususal Moebius function on integers.

julia> p=Poset((i,j)->i%j==0,1:8)
5,7<1
6<2<1
6<3<1
8<4<2

julia> moebius(p)
8-element Vector{Int64}:
  1
 -1
 -1
  0
 -1
  1
 -1
  0

Gapjm.Posets.moebiusmatrix — Function

moebiusmatrix(P::Poset) the matrix of Moebius function μ(x,y)

Base.minimum — Method

minimum(p::Poset) the minimum element of P if there is one.

Base.maximum — Method

maximum(p::Poset) the maximum element of P if there is one.

Gapjm.Posets.Poset — Type

Poset(W::CoxeterGroup,w=longest(W))

returns as a poset the Bruhat interval [1,w]of W. If w is not given, the whole Bruhat Poset of W is returned (W must then be finite).

julia> W=coxgroup(:A,2)
A₂

julia> Poset(W)
.<1,2<21,12<121

The above poset is constructed efficiently by constructing the Hasse diagram, but it could be constructed naively as follows:

julia> Poset(W)
.<1,2<21,12<121

julia> p=Poset((x,y)->bruhatless(W,x,y),elements(W))
1<2,3<4,5<6

The element printing is not so nice. This can be remedied by giving a function:

julia> p.show_element=(io,x,n)->(e=x.elements[n];isone(e) ? print(io,".") : print(io,joindigits(word(W,e))));

julia> p
.<2,1<21,12<121

julia> W=coxgroup(:A,3)
A₃

julia> Poset(W,W(1,3))
.<3,1<13