Presentations

Gapjm.Presentations

Gapjm.Presentations.AbsWord

Gapjm.Presentations.FpGroup

Gapjm.Presentations.Presentation

Gapjm.Presentations.conjugate

Gapjm.Presentations.relators

Gapjm.Presentations.simplify

Gapjm.Presentations.tryconjugate

Gapjm.Presentations.@AbsWord

Gapjm.Presentations — Module

This is a port of some GAP3/VkCurve functionality on presentations and finitely presented groups.

finitely presented groups are distinguished from group presentations which are objects of their own. The reason is that when a presentation is changed (e.g. simplified) by Tietze transformations, new generators and relators are introduced; thus all words in a finitely presented group would also have to be changed if such a Tietze transformation were applied to the group. Therefore, it is better to work separately with the presentation and reflect only carefully the changes on the group.

In order to speed up the algorithms, the relators in a presentation are not represented by AbsWords, but by lists of positive or negative generator numbers which we call Tietze words.

We now describe the available functions; for more information look at individual docstrings. The functions Presentation and FpGroup create a presentation from a finitely presented group or, vice versa.

Since it could be large, by default a presentation is printed as a summary of the number of generators, the number of relators, and the total length of all relators. We illustrate below functions displaying more of a presentation.

julia> @AbsWord a,b

julia> F=FpGroup([a,b])
FreeGroup(a,b)

julia> G=F/[a^2,b^7,comm(a,a^b),comm(a,a^(b^2))*inv(b^a)]
FreeGroup(a,b)/[a²,b⁷,a⁻¹b⁻¹a⁻¹bab⁻¹ab,a⁻¹b⁻²a⁻¹b²ab⁻²ab²a⁻¹b⁻¹a]

julia> P=Presentation(G)
Presentation: 2 generators, 4 relators, total length 30

julia> relators(P)
4-element Vector{AbsWord}:
 a²
 b⁷
 ab⁻¹abab⁻¹ab
 b⁻²ab²ab⁻²ab²ab⁻¹

julia> Presentations.PrintGenerators(P)
#I 1. a 10 occurrences involution
#I 2. b 20 occurrences

julia> Presentations.display_balanced(P)
1: a=A
2: bbbbbbb=1
3: aBab=BAbA
4: BBabbaBBabbaB=1

for more information look at the help strings of AbsWord, Presentation, FpGroup, Go, GoGo, conjugate, tryconjugate, simplify, relators, display_balanced

Gapjm.Presentations.AbsWord — Type

An AbsWord is an abstract word in some generators, that is a product of some of these generators raised to positive or negative powers. The generators are indexed by Symbols in the julia implementation (as a Vector{Pair{Symbol,Int}})

Gapjm.Presentations.@AbsWord — Macro

@AbsWord x,y,z defines the variables x,y,z to be AbsWords

Gapjm.Presentations.FpGroup — Type

FpGroup(P::Presentation)

returns the finitely presented group defined by the presentation P.

Gapjm.Presentations.Presentation — Method

Presentation( G::FpGroup[, printlevel])

Presentation returns a presentation containing a copy of the presentation of the given finitely presented group G on the same set of generators.

The optional printlevel parameter can be used to restrict or to extend the amount of output provided by Tietze transformation functions when being applied to the created presentation. The default value 1 is designed for interactive use and implies explicit messages to be displayed by most of these functions. A printlevel value of 0 will suppress these messages, whereas a printlevel value of 2 will enforce some additional output.

Gapjm.Presentations.relators — Function

relators(P::Presentation)

prints the list of relators of P.

Gapjm.Presentations.simplify — Function

simplify(G)

simplify applies Tietze transformations to a copy of the presentation of the given finitely presented group G in order to reduce it with respect to the number of generators, the number of relators, and the relator lengths.

simplify returns the resulting finitely presented group (which is isomorphic to G).

julia> @AbsWord a,b,c,d,e,f

julia> F=FpGroup([a,b,c,d,e,f])
FreeGroup(a,b,c,d,e,f)

julia> G=F/[a^2,b^2,d*f^-1,e^2,f^2,a*b^-1*c,a*e*c^-1,b*d^-1*c,c*d*e^-1,a*f*c^-2,c^4]
FreeGroup(a,b,c,d,e,f)/[a²,b²,df⁻¹,e²,f²,ab⁻¹c,aec⁻¹,bd⁻¹c,cde⁻¹,afc⁻²,c⁴]

julia> simplify(G)
FreeGroup(a,c)/[a²,ac⁻¹ac⁻¹,c⁴]

In fact, the call

julia> simplify(G)

is an abbreviation of the call sequence

julia> P=Presentation(G,0);simplify(P);FpGroup(P)

which applies a rather simple-minded strategy of Tietze transformations to the intermediate presentation P. If for some concrete group the resulting presentation is unsatisfying, then you should try a more sophisticated, interactive use of the available Tietze transformation functions (see "Tietze Transformations").

Gapjm.Presentations.conjugate — Function

conjugate(p,conjugation)

This program modifies a presentation by conjugating a generator by another. The conjugation to apply is described by a length-3 string of the same style as the result of display_balanced, that is "abA" means replace the second generator by its conjugate by the first, and "Aba" means replace it by its conjugate by the inverse of the first.

julia> Presentations.display_balanced(P)
1: dabcd=abcda
2: dabcdb=cabcda
3: bcdabcd=dabcdbc

julia> Presentations.display_balanced(conjugate(P,"Cdc"))
<< presentation with 4 generators, 3 relators of total length 36>>
1: dcabdc=cabdca
2: abdcab=cabdca
3: bdcabd=cabdca

Gapjm.Presentations.tryconjugate — Function

tryconjugate(p[,goal[,printlevel]])

This program tries to simplify group presentations by applying conjugations to the generators. The algorithm depends on random numbers, and on tree-searching, so is not reproducible. By default the program stops as soon as a shorter presentation is found. Sometimes this does not give the desired presentation. One can give a second argument goal, then the program will only stop when a presentation of length less than goal is found. Finally, a third argument can be given and then all presentations the programs runs over which are of length less than or equal to this argument are displayed. Due to the non-deterministic nature of the program, it may be useful to run it several times on the same input. Upon failure (to improve the presentation), the program returns p.

julia> display_balanced(p)
1: ba=ab
2: dbd=bdb
3: cac=aca
4: bcb=cbc
5: dAca=Acad
6: dcdc=cdcd
7: adad=dada
8: dcDbdc=bdcbdB
9: dcdadc=adcdad
10: adcDad=dcDadc
11: BcccbdcAb=dcbACdddc
julia> p=tryconjugate(p)
Presentation: 4 generators, 11 relators, total length 100
dcD=> Presentation: 4 generators, 10 relators, total length 90
# dcD gives Presentation: 4 generators, 10 relators, total length 90
Presentation: 4 generators, 10 relators, total length 90

julia> p=tryconjugate(p)
Dcd=> Presentation: 4 generators, 10 relators, total length 88
# Dcd gives Presentation: 4 generators, 10 relators, total length 88
Presentation: 4 generators, 10 relators, total length 88

julia> p=tryconjugate(p)
dcD=> Presentation: 4 generators, 10 relators, total length 90
Dbd=> Presentation: 4 generators, 10 relators, total length 96
Aca=> Presentation: 4 generators, 9 relators, total length 84
Presentation: 4 generators, 8 relators, total length 76
# Aca gives Presentation: 4 generators, 8 relators, total length 76
Presentation: 4 generators, 8 relators, total length 76

julia> p=tryconjugate(p)
Bcb=> Presentation: 4 generators, 8 relators, total length 70
# Bcb gives Presentation: 4 generators, 8 relators, total length 70
Presentation: 4 generators, 8 relators, total length 70

julia> p=tryconjugate(p)
Cac=> Presentation: 4 generators, 8 relators, total length 64
# Cac gives Presentation: 4 generators, 8 relators, total length 64
Presentation: 4 generators, 8 relators, total length 64

julia> p=tryconjugate(p)
caC=> Presentation: 4 generators, 8 relators, total length 58
# caC gives Presentation: 4 generators, 8 relators, total length 58
Presentation: 4 generators, 8 relators, total length 58

julia> p=tryconjugate(p)
Cac=> Presentation: 4 generators, 8 relators, total length 64
Cbc=> Presentation: 4 generators, 7 relators, total length 50
# Cbc gives Presentation: 4 generators, 7 relators, total length 50
Presentation: 4 generators, 7 relators, total length 50

julia> p=tryconjugate(p)
cdC=> Presentation: 4 generators, 7 relators, total length 56
Dcd=> Presentation: 4 generators, 7 relators, total length 54
Cac=> Presentation: 4 generators, 7 relators, total length 48
# Cac gives Presentation: 4 generators, 7 relators, total length 48
Presentation: 4 generators, 7 relators, total length 48

julia> p=tryconjugate(p)
caC=> Presentation: 4 generators, 7 relators, total length 50
Cdc=> Presentation: 4 generators, 7 relators, total length 50
Dbd=> Dcd=> Presentation: 4 generators, 7 relators, total length 60
Bab=> Aba=> Aca=> Presentation: 4 generators, 7 relators, total length 46
# Aca gives Presentation: 4 generators, 7 relators, total length 46
Presentation: 4 generators, 7 relators, total length 46

julia> display_balanced(p)
1: db=bd
2: ba=ab
3: cac=aca
4: ada=dad
5: bcb=cbc
6: cdcd=dcdc
7: AdCacd=cAdCac