d-Harish-Chandra series

Chevie.dSeries

Chevie.dSeries.Series

Chevie.Symbols.ennola

Chevie.Uch.cuspidal_data

Chevie.dSeries вҖ” Module

d-Harish-Chandra series describe unipotent l-blocks of a finite reductive group $рқҗҶ(рқ”Ҫ_q)$ for $l|ОҰ_d(q)$ (at least, when l is not too small which means mostly not a bad prime for рқҗҶ). Some of the facts stated below are still partly conjectural, we do not try to distinguish precisely what has been established and what is still conjectural.

If (рқҗӢ,О») is a d-cuspidal pair then the constituents of the Lusztig induced $R_рқҗӢ^рқҗҶ(О»)$ are called a d-Harish-Chandra series; they form the unipotent part of an l-block of $рқҗҶ^F$ (BrouГ© et al., 1993; 5.24). It is conjectured (BrouГ© and Michel, 1997; 5.7) (and proven in some cases) that the $рқҗҶ^F$-endomorphism algebra of the l-adic cohomology of the variety рқҗ— which defines the Lusztig induction is a d-cyclotomic Hecke algebra $H_рқҗҶ(рқҗӢ,О»)$ for the group $W_рқҗҶ(рқҗӢ,О»):=N_рқҗҶ(рқҗӢ,О»)/рқҗӢ$, which is a complex reflection group вҖ“- here d-cyclotomic means that the parameters of $H_рқҗҶ(рқҗӢ,О»)$ are monomials in q and that $H_рқҗҶ(рқҗӢ,О»)$ specializes to the algebra of $W_рқҗҶ(рқҗӢ,О»)$ for $qвҶҰО¶_d$.

It follows that the decomposition of the Lusztig induction is of the form $R_рқҗӢ^рқҗҶ(О»)=вҲ‘_{П•вҲҲIrr(W_рқҗҶ(рқҗӢ,О»))}(-1)^{nбө©} П•(1)Оібө©,$ where Оібө© is a unipotent character of рқҗҶ^F attached to П• and where nбө© is the degree $H^{nбө©}_c(рқҗ—)$ where Оібө© occurss; and further for any П• we have $R_рқҗӢ^рқҗҶ(О»)(1)= (-1)^{nбө©} Оібө©(1)Sбө©$ where Sбө© is the Schur element of the character of $H_рқҗҶ(рқҗӢ,О»)$ which deforms to П•. The function Series allows to explore a d-Harish-Chandra series.

julia> W=rootdatum("3D4")
ВіDвӮ„

julia> l=cuspidal_data(W,3)
2-element Vector{@NamedTuple{levi::Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}, cuspidal::Int64, d::Root1}}:
 (levi = ВіDвӮ„, cuspidal = 8, d = О¶вӮғ)
 (levi = ВіDвӮ„вӮҚвӮҺ=ОҰвӮғВІ, cuspidal = 1, d = О¶вӮғ)

julia> Series(W,l[2]...)
О¶вӮғ-series R^ВіDвӮ„_{ВіDвӮ„вӮҚвӮҺ=ОҰвӮғВІ}(О»==Id)  H_G(L,О»)==hecke(GвӮ„,Vector{Mvp{Cyc{Int64}, Int64}}[[О¶вӮғqВІ, О¶вӮғ, О¶вӮғq]])
в”Ңв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯ в”Ӯ    Оібө©    ПҶ  Оө family nвҒ°в”Ӯ
в”ңв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”Ӯ1в”Ӯ  ПҶвӮҒвҖҡвӮҖ ПҶвӮҒвҖҡвӮҖ  1         1в”Ӯ
в”Ӯ2в”Ӯ  ПҶвӮҒвҖҡвӮҶ ПҶвӮҒвҖҡвӮ„  1         2в”Ӯ
в”Ӯ3в”Ӯ  ПҶвӮӮвҖҡвӮӮ ПҶвӮҒвҖҡвӮҲ -1         5в”Ӯ
в”Ӯ6в”Ӯ ПҶвҖівӮҒвҖҡвӮғ ПҶвӮӮвҖҡвӮ…  1         4в”Ӯ
в”Ӯ5в”Ӯ ПҶвҖІвӮҒвҖҡвӮғ ПҶвӮӮвҖҡвӮғ -1         3в”Ӯ
в”Ӯ7в”Ӯ  ПҶвӮӮвҖҡвӮҒ ПҶвӮӮвҖҡвӮҒ -1         5в”Ӯ
в”Ӯ4в”ӮВіDвӮ„[1] ПҶвӮғвҖҡвӮӮ  1         5в”Ӯ
в””в”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

Above we explore the 3-series corresponding to $R_рқҗ“^рқҗҶ(Id)$ where рқҗҶ is the triality group and рқҗ“ is the torus of type (qВІ+q+1)ВІ. The group $W_рқҗҶ(рқҗ“)$ is the complex reflection group GвӮ„. The displays shows in the column 'Оібө©' the name of the unipotent characters constituents of $R_рқҗ“^рқҗҶ(Id)$, and in the first column the number of these characters in the list of unipotent characters. In the column 'ПҶ' the name of the character of $W_рқҗҶ(рқҗ“)$ corresponding to the unipotent character Оібө© is shown; in the column 'Оө' we show the sign $(-1)^{nбө©}$. Finally in the last column we show in which family of unipotent characters is Оібө©.

The theory of d-Harish-Chandra series can be generalized to spetsial complex reflection groups using some axioms. We show below an example.

julia> W=complex_reflection_group(4)
GвӮ„

julia> l=cuspidal_data(W,3)
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = GвӮ„, cuspidal = 3, d = О¶вӮғ)
 (levi = GвӮ„, cuspidal = 6, d = О¶вӮғ)
 (levi = GвӮ„, cuspidal = 7, d = О¶вӮғ)
 (levi = GвӮ„, cuspidal = 10, d = О¶вӮғ)
 (levi = GвӮ„вӮҚвӮҺ=ОҰвӮҒОҰвҖІвӮғ, cuspidal = 1, d = О¶вӮғ)

julia> Series(W,l[5]...)
О¶вӮғ-series R^GвӮ„_{GвӮ„вӮҚвӮҺ=ОҰвӮҒОҰвҖІвӮғ}(О»==Id)  W_G(L,О»)==ZвӮҶ
в”Ңв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯ в”Ӯ   Оібө© ПҶ(mod 3)  Оө parameter family nвҒ°в”Ӯ
в”ңв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”Ӯ1в”Ӯ ПҶвӮҒвҖҡвӮҖ        1  1      О¶вӮғqВІ         1в”Ӯ
в”Ӯ5в”Ӯ ПҶвӮӮвҖҡвӮғ       О¶вӮҶ  1      -О¶вӮғq         2в”Ӯ
в”Ӯ2в”Ӯ ПҶвӮҒвҖҡвӮ„       О¶вӮғ -1        О¶вӮғ         4в”Ӯ
в”Ӯ8в”Ӯ ZвӮғ:2       -1 -1     -О¶вӮғВІq         2в”Ӯ
в”Ӯ9в”ӮZвӮғ:11      О¶вӮғВІ -1       О¶вӮғВІ         4в”Ӯ
в”Ӯ4в”Ӯ ПҶвӮӮвҖҡвӮ…      О¶вӮҶвҒө -1       -О¶вӮғ         4в”Ӯ
в””в”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

Above we explore the 3-series corresponding to the trivial character of the torus of type (q-1)(q-О¶вӮғ). For cyclic groups $W_рқҗҶ(рқҗӢ,О»)$ we display the parameters in the table since they are associated to characters of $W_рқҗҶ(рқҗӢ,О»)$. Finally the mention (mod 3) which appears in the ПҶ column means that in this case the axioms leave an ambiguity in the correspondence between unipotent characters Оібө© and characters П• (as well as with parameters): the correspondence is known only up to a translation by 3 (in this case, the same as a global multiplication of all П• by -1).

Finally, we should note that if the reflection group or coset W is not defined over the integers, what counts is not cyclotomic polynomials but factors of them over the field of definition of W. In this case, one should not give as argument an integer d representing $О¶_d$ but specify a root of unity. For instance, in the above case we get a different answer with:

julia> cuspidal_data(W,E(3,2))
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = GвӮ„, cuspidal = 2, d = О¶вӮғВІ)
 (levi = GвӮ„, cuspidal = 5, d = О¶вӮғВІ)
 (levi = GвӮ„, cuspidal = 7, d = О¶вӮғВІ)
 (levi = GвӮ„, cuspidal = 10, d = О¶вӮғВІ)
 (levi = GвӮ„вӮҚвӮҺ=ОҰвӮҒОҰвҖівӮғ, cuspidal = 1, d = О¶вӮғВІ)

Chevie.Uch.cuspidal_data вҖ” Function

cuspidal_data(W[,d[,ad]];proper=false,all=false)

returns named tuples (levi=LF,cuspidal=О»,d=d) where LF is a d-split Levi (with d-center of dimension ad if ad is given) and О» is a d-cuspidal character of LF. If d=1 this returns ordinary cuspidal characters. The character О» is given as its index in the list of unipotent characters. If d was given as an integer, it is returned as a Root1 representing E(d).

If the keyword proper=true is given, only the data where LF!=W (or equivalently ad>0) are returned.

If d is omitted, data for all d orders of eigenvalues of elements of W is returned. If in addition the keyword argument all=true is given, data for all eigenvalues of elements of W is returned.

julia> cuspidal_data(coxgroup(:F,4),1)
9-element Vector{@NamedTuple{levi::Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}, cuspidal::Int64, d::Root1}}:
 (levi = FвӮ„, cuspidal = 31, d = 1)
 (levi = FвӮ„, cuspidal = 32, d = 1)
 (levi = FвӮ„, cuspidal = 33, d = 1)
 (levi = FвӮ„, cuspidal = 34, d = 1)
 (levi = FвӮ„, cuspidal = 35, d = 1)
 (levi = FвӮ„, cuspidal = 36, d = 1)
 (levi = FвӮ„, cuspidal = 37, d = 1)
 (levi = FвӮ„вӮҚвӮӮвӮғвӮҺ=BвӮӮвӮҚвӮӮвӮҒвӮҺОҰвӮҒВІ, cuspidal = 6, d = 1)
 (levi = FвӮ„вӮҚвӮҺ=ОҰвӮҒвҒҙ, cuspidal = 1, d = 1)

julia> cuspidal_data(complex_reflection_group(4),3)
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = GвӮ„, cuspidal = 3, d = О¶вӮғ)
 (levi = GвӮ„, cuspidal = 6, d = О¶вӮғ)
 (levi = GвӮ„, cuspidal = 7, d = О¶вӮғ)
 (levi = GвӮ„, cuspidal = 10, d = О¶вӮғ)
 (levi = GвӮ„вӮҚвӮҺ=ОҰвӮҒОҰвҖІвӮғ, cuspidal = 1, d = О¶вӮғ)

Chevie.dSeries.Series вҖ” Type

Series(W, L, cuspidal, d)

If the reflection coset or group W corresponds to the algebraic group рқҗҶ and cuspidal is the index of a d-cuspidal unipotent character О» of the d-split Levi рқҗӢ, constructs the d-series $R_рқҗӢ^рқҗҶ(О»)$. The result s has the following properties and functions:

s.spets: the reflection group or coset W.

s.levi: the subcoset L.

s.cuspidal: the index of О» in UnipotentCharacters(L).

s.d: the value of d (a Root1).

relative_group(s): the group $W_рқҗҶ(рқҗӢ,О»)$.

s.e: the order of $W_рқҗҶ(рқҗӢ,О»)$.

dSeries.RLG(s): the UnipotentCharacter given by $R_рқҗӢ^рқҗҶ(О»)$.

signs(s): for each character ПҶ of relative_group(s) the sign $(-1)^{n_ПҶ}$ in RLG(s) of the corresponding constituent Оібө© of RLG(s).

degree(s): the generic degree of RLG(s), as a CycPol.

charnumbers(s): the indices in UnipotentCharacters(W) of the

constituents of RLG(s).

hecke(s): the hecke algebra $H_рқҗҶ(рқҗӢ,О»)$.

The function Series has another form:

Series(W [,d [,ad]];k...)

where it returns a vector of Series corresponding to the cuspidal_data described by the arguments and the keywords.

julia> W=complex_reflection_group(4)
GвӮ„

julia> Series(W,3;proper=true)
1-element Vector{Series}:
 О¶вӮғ-series R^GвӮ„_{GвӮ„вӮҚвӮҺ=ОҰвӮҒОҰвҖІвӮғ}(О»==Id)  W_G(L,О»)==ZвӮҶ

julia> s=Series(W,3,1)[1]
О¶вӮғ-series R^GвӮ„_{GвӮ„вӮҚвӮҺ=ОҰвӮҒОҰвҖІвӮғ}(О»==Id)  W_G(L,О»)==ZвӮҶ
в”Ңв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯ в”Ӯ   Оібө© ПҶ(mod 3)  Оө parameter family nвҒ°в”Ӯ
в”ңв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”Ӯ1в”Ӯ ПҶвӮҒвҖҡвӮҖ        1  1      О¶вӮғqВІ         1в”Ӯ
в”Ӯ5в”Ӯ ПҶвӮӮвҖҡвӮғ       О¶вӮҶ  1      -О¶вӮғq         2в”Ӯ
в”Ӯ2в”Ӯ ПҶвӮҒвҖҡвӮ„       О¶вӮғ -1        О¶вӮғ         4в”Ӯ
в”Ӯ8в”Ӯ ZвӮғ:2       -1 -1     -О¶вӮғВІq         2в”Ӯ
в”Ӯ9в”ӮZвӮғ:11      О¶вӮғВІ -1       О¶вӮғВІ         4в”Ӯ
в”Ӯ4в”Ӯ ПҶвӮӮвҖҡвӮ…      О¶вӮҶвҒө -1       -О¶вӮғ         4в”Ӯ
в””в”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

julia> s.spets
GвӮ„

julia> s.levi
GвӮ„вӮҚвӮҺ=ОҰвӮҒОҰвҖІвӮғ

julia> s.cuspidal
1

julia> s.d
Root1: О¶вӮғ

julia> hecke(s)
hecke(GвӮҶвҖҡвӮҒвҖҡвӮҒ,Vector{Mvp{Cyc{Int64}, Int64}}[[О¶вӮғqВІ, -О¶вӮғq, О¶вӮғ, -О¶вӮғВІq, О¶вӮғВІ, -О¶вӮғ]])

julia> degree(s)
О¶вӮғОҰвӮҒОҰвӮӮВІОҰвҖівӮғОҰвӮ„ОҰвӮҶ

julia> dSeries.RLG(s)
[GвӮ„]:<ПҶвӮҒвҖҡвӮҖ>-<ПҶвӮҒвҖҡвӮ„>-<ПҶвӮӮвҖҡвӮ…>+<ПҶвӮӮвҖҡвӮғ>-<ZвӮғ:2>-<ZвӮғ:11>

julia> charnumbers(s)
6-element Vector{Int64}:
 1
 5
 2
 8
 9
 4

julia> signs(s)
6-element Vector{Int64}:
  1
  1
 -1
 -1
 -1
 -1

julia> relative_group(s)
GвӮҶвҖҡвӮҒвҖҡвӮҒ

Chevie.Symbols.ennola вҖ” Function

ennola(S::CharSymbol)

Ennola of e-symbol S (of content 1 or 0) The order of Ennola (order of center of reflection group) is computed automatically: it is e for content 1 and gcd(e,rank(S)) for content 0.

ennola(W[,z=E(gcd(degrees(W)))])

Let W be an irreducible spetsial reflection group or coset, and z the generator of the center of W, viewed as the root of unity E(gcd(degrees(W)))]). Let рқ”ҫ be the spets attached to W. A property checked case-by case is that, for a unipotent character Оі of рқ”ҫ with polynomial generic degree deg Оі(q) , deg Оі(zq) is equal to Вұdeg Оі'(q) for another unipotent character Оі'; ВұОі' is called the Ennola transform of Оі. For W a Weyl group, the spets рқ”ҫ is a finite reductive group, in which case z=-1 if -1 is in W and z=1 otherwise. The function returns the signed permutation e done by ennola on the unipotent degrees (as an SPerm of 1:length(UnipotentCharacters(W))).

The SPerm e is not uniquely determined by the degrees since two degrees may be equal, but is uniquely determined by some additional axioms that we do not recall here (they include a description of the Ennola-permutation in terms of the Z-based rings attached to each Lusztig family).

If a second argument z is given, it should be a power of the default z and the corresponding power of e is returned.

julia> ennola(rootdatum("3D4"))
SPerm{Int64}: (3,-4)(5,-5)(6,-6)(7,-8)

julia> ennola(complex_reflection_group(14))
SPerm{Int64}: (2,43,-14,16,41,34)(3,35,40,18,-11,42)(4,-37,25,-17,-26,-36)(5,-6,-79)(7,-7)(8,-74)(9,-73)(10,-52,13,31,-50,29)(12,53,15,32,-51,-30)(19,71,70,21,67,68,20,69,72)(22,-39,27,-33,-28,-38)(23,24,-66,-23,-24,66)(44,46,49,-44,-46,-49)(45,48,47,-45,-48,-47)(54,-63,-55,-57,62,-56)(58,-65,-59,-61,64,-60)(75,-77)(76,-76)(78,-78)

The above example shows that it may happen that the order of z-Ennola (here 18) is greater than the order of z (here 6); this is related to the presence of irrationalities qв…“ in the character table of the spetsial Hecke algebra of W.

For a non-irreducible group, z-ennola is defined if z can be considered an element of the the centre of each irreducible component.