d-Harish-Chandra series

Chevie.dSeries

Chevie.dSeries.Series

Chevie.Symbols.ennola

Chevie.Uch.cuspidal_data

Chevie.dSeries — Module

d-Harish-Chandra series describe unipotent l-blocks of a finite reductive group $𝐆(𝔽_q)$ for $l|Φ_d(q)$ (at least, when l is not too small which means mostly not a bad prime for 𝐆). Some of the facts stated below are still partly conjectural, we do not try to distinguish precisely what has been established and what is still conjectural.

If (𝐋,λ) is a d-cuspidal pair then the constituents of the Lusztig induced $R_𝐋^𝐆(λ)$ are called a d-Harish-Chandra series; they form the unipotent part of an l-block of $𝐆^F$ (Broué et al., 1993; 5.24). It is conjectured (Broué and Michel, 1997; 5.7) (and proven in some cases) that the $𝐆^F$-endomorphism algebra of the l-adic cohomology of the variety 𝐗 which defines the Lusztig induction is a d-cyclotomic Hecke algebra $H_𝐆(𝐋,λ)$ for the group $W_𝐆(𝐋,λ):=N_𝐆(𝐋,λ)/𝐋$, which is a complex reflection group –- here d-cyclotomic means that the parameters of $H_𝐆(𝐋,λ)$ are monomials in q and that $H_𝐆(𝐋,λ)$ specializes to the algebra of $W_𝐆(𝐋,λ)$ for $q↦ζ_d$.

It follows that the decomposition of the Lusztig induction is of the form $R_𝐋^𝐆(λ)=∑_{ϕ∈Irr(W_𝐆(𝐋,λ))}(-1)^{nᵩ} ϕ(1)γᵩ,$ where γᵩ is a unipotent character of 𝐆^F attached to ϕ and where nᵩ is the degree $H^{nᵩ}_c(𝐗)$ where γᵩ occurss; and further for any ϕ we have $R_𝐋^𝐆(λ)(1)= (-1)^{nᵩ} γᵩ(1)Sᵩ$ where Sᵩ is the Schur element of the character of $H_𝐆(𝐋,λ)$ which deforms to ϕ. The function Series allows to explore a d-Harish-Chandra series.

julia> W=rootdatum("3D4")
³D₄

julia> l=cuspidal_data(W,3)
2-element Vector{@NamedTuple{levi::Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}, cuspidal::Int64, d::Root1}}:
 (levi = ³D₄, cuspidal = 8, d = ζ₃)
 (levi = ³D₄₍₎=Φ₃², cuspidal = 1, d = ζ₃)

julia> Series(W,l[2]...)
ζ₃-series R^³D₄_{³D₄₍₎=Φ₃²}(λ==Id)  H_G(L,λ)==hecke(G₄,Vector{Mvp{Cyc{Int64}, Int64}}[[ζ₃q², ζ₃, ζ₃q]])
┌─┬────────────────────────┐
│ │    γᵩ    φ  ε family n⁰│
├─┼────────────────────────┤
│1│  φ₁‚₀ φ₁‚₀  1         1│
│2│  φ₁‚₆ φ₁‚₄  1         2│
│3│  φ₂‚₂ φ₁‚₈ -1         5│
│6│ φ″₁‚₃ φ₂‚₅  1         4│
│5│ φ′₁‚₃ φ₂‚₃ -1         3│
│7│  φ₂‚₁ φ₂‚₁ -1         5│
│4│³D₄[1] φ₃‚₂  1         5│
└─┴────────────────────────┘

Above we explore the 3-series corresponding to $R_𝐓^𝐆(Id)$ where 𝐆 is the triality group and 𝐓 is the torus of type (q²+q+1)². The group $W_𝐆(𝐓)$ is the complex reflection group G₄. The displays shows in the column 'γᵩ' the name of the unipotent characters constituents of $R_𝐓^𝐆(Id)$, and in the first column the number of these characters in the list of unipotent characters. In the column 'φ' the name of the character of $W_𝐆(𝐓)$ corresponding to the unipotent character γᵩ is shown; in the column 'ε' we show the sign $(-1)^{nᵩ}$. Finally in the last column we show in which family of unipotent characters is γᵩ.

The theory of d-Harish-Chandra series can be generalized to spetsial complex reflection groups using some axioms. We show below an example.

julia> W=complex_reflection_group(4)
G₄

julia> l=cuspidal_data(W,3)
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = G₄, cuspidal = 3, d = ζ₃)
 (levi = G₄, cuspidal = 6, d = ζ₃)
 (levi = G₄, cuspidal = 7, d = ζ₃)
 (levi = G₄, cuspidal = 10, d = ζ₃)
 (levi = G₄₍₎=Φ₁Φ′₃, cuspidal = 1, d = ζ₃)

julia> Series(W,l[5]...)
ζ₃-series R^G₄_{G₄₍₎=Φ₁Φ′₃}(λ==Id)  W_G(L,λ)==Z₆
┌─┬─────────────────────────────────────┐
│ │   γᵩ φ(mod 3)  ε parameter family n⁰│
├─┼─────────────────────────────────────┤
│1│ φ₁‚₀        1  1      ζ₃q²         1│
│5│ φ₂‚₃       ζ₆  1      -ζ₃q         2│
│2│ φ₁‚₄       ζ₃ -1        ζ₃         4│
│8│ Z₃:2       -1 -1     -ζ₃²q         2│
│9│Z₃:11      ζ₃² -1       ζ₃²         4│
│4│ φ₂‚₅      ζ₆⁵ -1       -ζ₃         4│
└─┴─────────────────────────────────────┘

Above we explore the 3-series corresponding to the trivial character of the torus of type (q-1)(q-ζ₃). For cyclic groups $W_𝐆(𝐋,λ)$ we display the parameters in the table since they are associated to characters of $W_𝐆(𝐋,λ)$. Finally the mention (mod 3) which appears in the φ column means that in this case the axioms leave an ambiguity in the correspondence between unipotent characters γᵩ and characters ϕ (as well as with parameters): the correspondence is known only up to a translation by 3 (in this case, the same as a global multiplication of all ϕ by -1).

Finally, we should note that if the reflection group or coset W is not defined over the integers, what counts is not cyclotomic polynomials but factors of them over the field of definition of W. In this case, one should not give as argument an integer d representing $ζ_d$ but specify a root of unity. For instance, in the above case we get a different answer with:

julia> cuspidal_data(W,E(3,2))
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = G₄, cuspidal = 2, d = ζ₃²)
 (levi = G₄, cuspidal = 5, d = ζ₃²)
 (levi = G₄, cuspidal = 7, d = ζ₃²)
 (levi = G₄, cuspidal = 10, d = ζ₃²)
 (levi = G₄₍₎=Φ₁Φ″₃, cuspidal = 1, d = ζ₃²)

Chevie.Uch.cuspidal_data — Function

cuspidal_data(W[,d[,ad]];proper=false,all=false)

returns named tuples (levi=LF,cuspidal=λ,d=d) where LF is a d-split Levi (with d-center of dimension ad if ad is given) and λ is a d-cuspidal character of LF. If d=1 this returns ordinary cuspidal characters. The character λ is given as its index in the list of unipotent characters. If d was given as an integer, it is returned as a Root1 representing E(d).

If the keyword proper=true is given, only the data where LF!=W (or equivalently ad>0) are returned.

If d is omitted, data for all d orders of eigenvalues of elements of W is returned. If in addition the keyword argument all=true is given, data for all eigenvalues of elements of W is returned.

julia> cuspidal_data(coxgroup(:F,4),1)
9-element Vector{@NamedTuple{levi::Spets{FiniteCoxeterSubGroup{Perm{Int16},Int64}}, cuspidal::Int64, d::Root1}}:
 (levi = F₄, cuspidal = 31, d = 1)
 (levi = F₄, cuspidal = 32, d = 1)
 (levi = F₄, cuspidal = 33, d = 1)
 (levi = F₄, cuspidal = 34, d = 1)
 (levi = F₄, cuspidal = 35, d = 1)
 (levi = F₄, cuspidal = 36, d = 1)
 (levi = F₄, cuspidal = 37, d = 1)
 (levi = F₄₍₂₃₎=B₂₍₂₁₎Φ₁², cuspidal = 6, d = 1)
 (levi = F₄₍₎=Φ₁⁴, cuspidal = 1, d = 1)

julia> cuspidal_data(complex_reflection_group(4),3)
5-element Vector{@NamedTuple{levi::Spets{PRSG{Cyc{Rational{Int64}}, Int16}}, cuspidal::Int64, d::Root1}}:
 (levi = G₄, cuspidal = 3, d = ζ₃)
 (levi = G₄, cuspidal = 6, d = ζ₃)
 (levi = G₄, cuspidal = 7, d = ζ₃)
 (levi = G₄, cuspidal = 10, d = ζ₃)
 (levi = G₄₍₎=Φ₁Φ′₃, cuspidal = 1, d = ζ₃)

Chevie.dSeries.Series — Type

Series(W, L, cuspidal, d)

If the reflection coset or group W corresponds to the algebraic group 𝐆 and cuspidal is the index of a d-cuspidal unipotent character λ of the d-split Levi 𝐋, constructs the d-series $R_𝐋^𝐆(λ)$. The result s has the following properties and functions:

s.spets: the reflection group or coset W.

s.levi: the subcoset L.

s.cuspidal: the index of λ in UnipotentCharacters(L).

s.d: the value of d (a Root1).

relative_group(s): the group $W_𝐆(𝐋,λ)$.

s.e: the order of $W_𝐆(𝐋,λ)$.

dSeries.RLG(s): the UnipotentCharacter given by $R_𝐋^𝐆(λ)$.

signs(s): for each character φ of relative_group(s) the sign $(-1)^{n_φ}$ in RLG(s) of the corresponding constituent γᵩ of RLG(s).

degree(s): the generic degree of RLG(s), as a CycPol.

charnumbers(s): the indices in UnipotentCharacters(W) of the

constituents of RLG(s).

hecke(s): the hecke algebra $H_𝐆(𝐋,λ)$.

The function Series has another form:

Series(W [,d [,ad]];k...)

where it returns a vector of Series corresponding to the cuspidal_data described by the arguments and the keywords.

julia> W=complex_reflection_group(4)
G₄

julia> Series(W,3;proper=true)
1-element Vector{Series}:
 ζ₃-series R^G₄_{G₄₍₎=Φ₁Φ′₃}(λ==Id)  W_G(L,λ)==Z₆

julia> s=Series(W,3,1)[1]
ζ₃-series R^G₄_{G₄₍₎=Φ₁Φ′₃}(λ==Id)  W_G(L,λ)==Z₆
┌─┬─────────────────────────────────────┐
│ │   γᵩ φ(mod 3)  ε parameter family n⁰│
├─┼─────────────────────────────────────┤
│1│ φ₁‚₀        1  1      ζ₃q²         1│
│5│ φ₂‚₃       ζ₆  1      -ζ₃q         2│
│2│ φ₁‚₄       ζ₃ -1        ζ₃         4│
│8│ Z₃:2       -1 -1     -ζ₃²q         2│
│9│Z₃:11      ζ₃² -1       ζ₃²         4│
│4│ φ₂‚₅      ζ₆⁵ -1       -ζ₃         4│
└─┴─────────────────────────────────────┘

julia> s.spets
G₄

julia> s.levi
G₄₍₎=Φ₁Φ′₃

julia> s.cuspidal
1

julia> s.d
Root1: ζ₃

julia> hecke(s)
hecke(G₆‚₁‚₁,Vector{Mvp{Cyc{Int64}, Int64}}[[ζ₃q², -ζ₃q, ζ₃, -ζ₃²q, ζ₃², -ζ₃]])

julia> degree(s)
ζ₃Φ₁Φ₂²Φ″₃Φ₄Φ₆

julia> dSeries.RLG(s)
[G₄]:<φ₁‚₀>-<φ₁‚₄>-<φ₂‚₅>+<φ₂‚₃>-<Z₃:2>-<Z₃:11>

julia> charnumbers(s)
6-element Vector{Int64}:
 1
 5
 2
 8
 9
 4

julia> signs(s)
6-element Vector{Int64}:
  1
  1
 -1
 -1
 -1
 -1

julia> relative_group(s)
G₆‚₁‚₁

Chevie.Symbols.ennola — Function

ennola(S::CharSymbol)

Ennola of e-symbol S (of content 1 or 0) The order of Ennola (order of center of reflection group) is computed automatically: it is e for content 1 and gcd(e,rank(S)) for content 0.

ennola(W[,z=E(gcd(degrees(W)))])

Let W be an irreducible spetsial reflection group or coset, and z the generator of the center of W, viewed as the root of unity E(gcd(degrees(W)))]). Let 𝔾 be the spets attached to W. A property checked case-by case is that, for a unipotent character γ of 𝔾 with polynomial generic degree deg γ(q) , deg γ(zq) is equal to ±deg γ'(q) for another unipotent character γ'; ±γ' is called the Ennola transform of γ. For W a Weyl group, the spets 𝔾 is a finite reductive group, in which case z=-1 if -1 is in W and z=1 otherwise. The function returns the signed permutation e done by ennola on the unipotent degrees (as an SPerm of 1:length(UnipotentCharacters(W))).

The SPerm e is not uniquely determined by the degrees since two degrees may be equal, but is uniquely determined by some additional axioms that we do not recall here (they include a description of the Ennola-permutation in terms of the Z-based rings attached to each Lusztig family).

If a second argument z is given, it should be a power of the default z and the corresponding power of e is returned.

julia> ennola(rootdatum("3D4"))
SPerm{Int64}: (3,-4)(5,-5)(6,-6)(7,-8)

julia> ennola(complex_reflection_group(14))
SPerm{Int64}: (2,43,-14,16,41,34)(3,35,40,18,-11,42)(4,-37,25,-17,-26,-36)(5,-6,-79)(7,-7)(8,-74)(9,-73)(10,-52,13,31,-50,29)(12,53,15,32,-51,-30)(19,71,70,21,67,68,20,69,72)(22,-39,27,-33,-28,-38)(23,24,-66,-23,-24,66)(44,46,49,-44,-46,-49)(45,48,47,-45,-48,-47)(54,-63,-55,-57,62,-56)(58,-65,-59,-61,64,-60)(75,-77)(76,-76)(78,-78)

The above example shows that it may happen that the order of z-Ennola (here 18) is greater than the order of z (here 6); this is related to the presence of irrationalities q⅓ in the character table of the spetsial Hecke algebra of W.

For a non-irreducible group, z-ennola is defined if z can be considered an element of the the centre of each irreducible component.