Symmetric functions

Chevie.SymFuncs

Chevie.SymFuncs.SymFuncAlgebra

Chevie.Chars.scalarproduct

FinitePosets.:⊗

Chevie.SymFuncs — Module

This module deals with symmetric functions. The main object is the (truncated) algebra of symmetric functions, and the bases p (power sums), h (complete symmetric functions) and s (Schur functions). Each of these basies in degree n is indexed by the partitions of n.

A typical session would begin by defining the symmetric functions and these bases:

julia> A=SymFuncAlgebra(10) # truncated in degree >10
SymFuncAlgebra(10)

julia> p=pbasis(A);h=hbasis(A);s=sbasis(A);

Then making elements in one of these bases. The following forms are equivalent:

julia> p(Partition(3,2,1))
p₃₂₁

julia> p([3,2,1])
p₃₂₁

julia> p(3,2,1)
p₃₂₁

The functions can be used to convert between bases:

julia> s(p(3,2,1))
-s₁₁₁₁₁₁-s₂₂₂+s₃₁₁₁+s₃₃-s₄₁₁+s₆

julia> h(p(3,2,1))
-h₁₁₁₁₁₁+5h₂₁₁₁₁-6h₂₂₁₁-3h₃₁₁₁+6h₃₂₁

The following operations are defined on symmetric functions, in addition to + and -:

julia> s(2,1)*s(2,1) # product
s₂₂₁₁+s₂₂₂+s₃₁₁₁+2s₃₂₁+s₃₃+s₄₁₁+s₄₂

julia> s(2,1)⊗s(2,1) # inner product
s₁₁₁+s₂₁+s₃

julia> SymFuncs.plethysm(p(2,1),s(2,1))
(1//9)p₂₂₂₁₁₁+(-1//9)p₃₂₂₂+(-1//9)p₆₁₁₁+(1//9)p₆₃

julia> scalarproduct(p(1,1,1),s(2,1))
2//1

One can mix bases in these operations. The basis of the left argument wins:

julia> s(2,1)+p(3)
s₁₁₁+s₃

finally one can convert a symmetric function to a symmetric polynomial. The number of variables is the highest degree in the function.

julia> Mvp(p(2)+p(3))
Mvp{Int64}: x₁³+x₁²+x₂³+x₂²+x₃³+x₃²

Chevie.SymFuncs.SymFuncAlgebra — Type

SymFuncAlgebra(n) the algebra of symmetric functions truncated in degree>n

FinitePosets.:⊗ — Method

a⊗b the inner product of the symmetric functions a and b

Chevie.Chars.scalarproduct — Method

scalarproduct(a,b) the scalar product of the symmetric functions a and b