Unipotent classes of reductive groups

Chevie.Ucl

Chevie.Ucl.GreenTable

Chevie.Ucl.ICCTable

Chevie.Ucl.UnipotentClass

Chevie.Ucl.UnipotentClasses

Chevie.Ucl.XTable

Chevie.Ucl.UnipotentValues

Chevie.Ucl.distinguished_parabolics

Chevie.Ucl.induced_linear_form

Chevie.Ucl.special_pieces

Chevie.Ucl вҖ” Module

This module gives information about the unipotent conjugacy classes of a connected reductive group рқҗҶ over an algebraically closed field k of characteristic p, and various associated invariants, such as the generalised Springer correspondence. The unipotent classes depend on p; their classification differs when p is not good (that is, when p divides one of the coefficients of the highest root of one of the irreducible components). In good characteristic, the unipotent classes are in bijection with nilpotent orbits on the Lie algebra рқ”Ө of рқҗҶ.

For each unipotent conjugacy class, we give the following information for a representative u of that class:

the centralizer $C_рқҗҶ (u)$, that we describe by the reductive part of $C_рқҗҶ (u)^0$, by the group of components $A(u):=C_рқҗҶ (u)/C_рқҗҶ (u)^0$, and by the dimension of its unipotent radical.

in good characteristic, the Dynkin-Richardson diagram.

the generalised Springer correspondence, attaching a character of the Weyl group or a relative Weyl group to each character of A(u).

The Dynkin-Richarson diagram is attached to a nilpotent element $e$ of the Lie algebra $рқ”Ө$. By the Jacobson-Morozov theorem there exists an $рқ”°рқ”©вӮӮ$ subalgebra of $рқ”Ө$ containing $e$ as the element $\begin{pmatrix}0&1\\0&0 \end{pmatrix}$. Let $рқҗ’$ be the torus $\begin{pmatrix}h&0\\0&h^{-1} \end{pmatrix}$ of $SLвӮӮ$ and let $рқҗ“$ be a maximal torus containing $рқҗ’$, so that $рқҗ’$ is the image of a one-parameter subgroup $ПғвҲҲ Y(рқҗ“)$. Consider the root decomposition $рқ”Ө=вҲ‘_{ОұвҲҲОҰ}рқ”Ө_Оұ$ given by $рқҗ“$ and the root system ОҰ; then $ОұвҶҰвҹЁПғ,Оұвҹ©$ defines a linear form on $ОҰ$, determined by its value on simple roots О . It is possible to choose a system of simple roots such that $вҹЁПғ,Оұвҹ©вүҘ 0$ for $ОұвҲҲО $, and then $вҹЁПғ,Оұвҹ©вҲҲ{0,1,2}$ for any $ОұвҲҲО $. The Dynkin diagram of $О $ decorated by these values $0,1,2$ is called the Dynkin-Richardson diagram of $e$, and in good characteristic is a complete invariant of the $рқ”Ө$-orbit of $e$. The Dynkin-Richardson diagrams thus classify unipotent classes of рқҗҶ in good characteristic.

Another classification of unipotent classes was given by (Bala and Carter, 1976). A standard parabolic subgroup рқҗҸ of рқҗҶ associated with the subset IвҠӮО of the simple roots is distinguished if the linear form Пғ taking the value 2 on ОұвҲҲ I and 0 on other simple roots satisfies 2nвӮҖ+semisimplerank(рқҗҶ)=nвӮӮ, where nбөў is the number of roots in ОҰ where Пғ takes the value i. Given a distinguished parabolic рқҗҸ, there is a unique unipotent class which is dense in the unipotent radical of рқҗҸ. For this class, the linear form described by the Dynkin-Richardson diagram is equal to Пғ. Such unipotent classes are called distinguished. The Bala-Carter theorem says that every unipotent class is distinguished in the smallest Levi subgroup рқҗӢ which contains it, and that such pairs of рқҗӢ and the distinguished parabolic рқҗҸ of рқҗӢ taken up to рқҗҶ-conjugacy are in bijection with unipotent classes of рқҗҶ. The function induced_linear_form allows to find the Dynkin-Richardson diagram in рқҗҶ of a unipotent class of рқҗӢ and thus recover these diagrams from the Bala-Carter classification.

Let $в„¬$ be the variety of all Borel subgroups and let $в„¬бөӨ$ be the subvariety of Borel subgroups containing the unipotent element u. Then $dim C_рқҗҶ(u)=rank рқҗҶ + 2 dim в„¬_u$ and in good characteristic this dimension can be computed from the linear form Пғ defined by the Dynkin-Richardson diagram: the dimension of the class of u is the number of roots Оұ such that $вҹЁПғ,Оұвҹ©вҲү{0,1}$.

We now describe the Springer correspondence. Indecomposable locally constant $рқҗҶ$-equivariant sheaves on a unipotent class $C$, called local systems, are parameterised by irreducible characters of $A(u)$ for uвҲҲ C. The ordinary Springer correspondence is a bijection between irreducible characters of the Weyl group and a large subset of the local systems, containing all trivial local systems (those parameterised by the trivial character of $A(u)$ for each $u$). More generally, the generalised Springer correspondence associates to each local system a (unique up to $рқҗҶ$-conjugacy) cuspidal datum, a Levi subgroup $рқҗӢ$ of $рқҗҶ$ and a cuspidal local system on an unipotent class of $рқҗӢ$, and a character of the relative Weyl group $W_рқҗҶ (рқҗӢ):=N_рқҗҶ (рқҗӢ)/рқҗӢ$. There are few cuspidal local systems (at most one in each dimension for classical groups). The ordinary Springer correspondence is the special case where $рқҗӢ$ is a maximal torus, the cuspidal local system is the trivial system on the identity element, and $W_рқҗҶ (рқҗӢ)=W$, the Weyl group.

The Springer correspondence gives information on the character values of a finite reductive group as follows: assume that $k$ is the algebraic closure of a finite field $рқ”Ҫ_q$ and that $F$ is the Frobenius attached to an $рқ”Ҫ_q$-structure of $рқҗҶ$. Let $C$ be an $F$-stable unipotent class and let $uвҲҲ C^F$; we call $C$ the geometric class of $u$ and the $рқҗҶ^F$-classes within $C^F$ are parameterised by the $F$-conjugacy classes of $A(u)$, denoted by $HВ№(F,A(u))$ (most of the time we can find $uвҲҲ C$ such that $F$ acts trivially on $A(u)$ and $HВ№(F,A(u))$ parametrises the conjugacy classes). To an $F$-stable character $ПҶ$ of $A(u)$ we associate the characteristic function of the corresponding local system (actually associated with an extension $ПҶМғ$ of $ПҶ$ to $A(u).F$); it is a class function $Y_{u,ПҶ}$ on $рқҗҶ^F$ which can be normalised so that: $Y_{u,ПҶ}(uвӮҒ)=ПҶМғ(cF)$ if $uвӮҒ$ is geometrically conjugate to $u$ and its $рқҗҶ^F$-class is parameterised by the $F$-conjugacy class $cF$ of $A(u)$, otherwise $Y_{u,ПҶ}(uвӮҒ)=0$. If the pair $u,ПҶ$ corresponds via the Springer correspondence to the character $ПҮ$ of $W_рқҗҶ(рқҗӢ)$, then $Y_{u,ПҶ}$ is also denoted $YбөӘ$. There is another important class of functions indexed by local systems: to a local system on class $C$ is attached an intersection cohomology complex, which is a complex of sheaves supported on the closure $CМ„$. Such a complex of sheaves is associated with its characteristic function, a class function of $рқҗҶ^F$ obtained by taking the alternating trace of the Frobenius acting on the stalks of the cohomology sheaves. If $Y_ПҲ$ is the characteristic function of a local system, the characteristic function of the corresponding intersection cohomology complex is denoted by $X_ПҲ$. This function is supported on $CМ„$, and Lusztig has shown that $X_ПҲ=вҲ‘_ПҮ P_{ПҲ,ПҮ} YбөӘ$ where $P_{ПҲ,ПҮ}$ are integer polynomials in $q$ and $YбөӘ$ are attached to local systems on classes lying in $CМ„$.

Lusztig and Shoji have given an algorithm to compute the matrix $P_{ПҲ,ПҮ}$, which is implemented in Chevie. The relation to the characters of $рқҗҶ(рқ”Ҫ_q)$, considering for simplicity the ordinary Springer correspondence, is that the restriction to the unipotent elements of the almost character $R_ПҮ$ is equal to $q^{bбөӘ} XбөӘ$, where $bбөӘ$ is $dim в„¬бөӨ$ for an element u of the class C such that the support of ПҮ is $CМ„$. The restrictions of the Deligne-Lusztig characters $R_w$ for wвҲҲ W on the unipotents are called the Green functions and can also be computed by Chevie. The values of all unipotent characters on unipotent elements can also be computed in principle by applying Lusztig's Fourier transform matrix (see the section on the Fourier matrix) but there is a difficulty in that the $XбөӘ$ must first be multiplied by some roots of unity which are not known in all cases (and when known may depend on the congruence class of $q$ modulo some small prime).

Finally, we describe how unipotent classes of рқҗҶ are parameterised in various quasisimple groups. In classical types, the classes are parametrised by partitions corresponding to the Jordan form in the standard representation. Thus,

for AвӮҷ we have partitions of n+1.

for B_n we have partitions of 2n+1 where even parts occur an even number of times. In characteristic 2, types B and C are isogenous so have the same classification; thus see the next paragraph.

for C_n we have partitions of 2n where odd parts occur an even number of times. In characteristic 2, there are 2бөҸ classes attached to a partition where k is the number of even parts which occur an even number of times.

for D_n we have partitions of 2n where even parts occur an even number of times, excepted there are two classes when all parts are even. In characteristic 2, we have partitions of 2n where odd parts occur an even number of times, excepted there are 2бөҸ+Оҙ classes attached to a partition where k is the number of even parts which occur an even number of times, and Оҙ is 2 when all parts are even and 0 otherwise.

In exceptional groups, the names of the classes are derived from the Bala-Carter classification. The name of a class parametrised by (рқҗӢ,рқҗҸ) is of the form l(p) where l is the name of рқҗӢ and (p) is present if there is more than one distinguished parabolic in рқҗӢ and describes which one it is. Before the Bala-Carter classification was universally adopted, Shoji and Mizuno used a different scheme where a class was sometimes parametrised by a reductive subgroup of maximal rank which was not a Levi. These older labels can be obtained instead by giving the IO property :shoji=>true or :mizuno=>true. In a bad characteristic p, there are extra classes. Each of them is associated with a class c in good characteristic and is called (c)вӮҡ.

We illustrate the above descriptions on some examples:

julia> UnipotentClasses(rootdatum(:sl,4))
UnipotentClasses(slвӮ„)
1111<211<22<31<4
в”Ңв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯu   в”ӮD-R dв„¬ бөӨ B-C   C_рқҗҶ(u) AвӮғ(ОҰвӮҒВі) AвӮҒ(AвӮҒГ—AвӮҒОҰвӮҒ)/-1 .(AвӮғ)/О¶вӮ„ .(AвӮғ)/О¶вӮ„Вів”Ӯ
в”ңв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”Ӯ4   в”Ӯ222    0 222    qВі.ZвӮ„     1:4           -1:2    О¶вӮ„:Id    О¶вӮ„Ві:Idв”Ӯ
в”Ӯ31  в”Ӯ202    1 22.    qвҒҙ.ОҰвӮҒ   Id:31                                  в”Ӯ
в”Ӯ22  в”Ӯ020    2 2.2 qвҒҙ.AвӮҒ.ZвӮӮ    2:22          11:11                   в”Ӯ
в”Ӯ211 в”Ӯ101    3 2..  qвҒө.AвӮҒОҰвӮҒ  Id:211                                  в”Ӯ
в”Ӯ1111в”Ӯ000    6 ...       AвӮғ Id:1111                                  в”Ӯ
в””в”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

The first column of the table gives the name of the unipotent class, here a partition describing the Jordan form. The partial order on unipotent classes given by the Zariski closure is given before the table. The column D-R, which is only shown in good characteristic, gives the Dynkin-Richardson diagram for each class; the column dBu gives the dimension of the variety $в„¬ бөӨ$. The column B-C gives the Bala-Carter classification of $u$, that is in the case of $slвӮ„$ it shows $u$ as a regular unipotent in a Levi subgroup by giving the Dynkin-Richardson diagram of a regular unipotent (all 2's) for the entries corresponding to the Levi and . for the entries which not corresponding to the Levi. The column C(u) describes the group $C_рқҗҶ(u)$: a power $qбөҲ$ describes that the unipotent radical of $C_рқҗҶ(u)$ has dimension $d$ (thus $qбөҲ$ rational points); then follows a description of the reductive part of the neutral component of $C_рқҗҶ(u)$, given by the name of its root datum. Then if $C_рқҗҶ(u)$ is not connected, the description of $A(u)$ is given using a different vocabulary: a cyclic group of order 4 is given as Z4, and a symmetric group on 3 points would be given as S3.

For example, the first class 4 has $C_рқҗҶ(u)^0$ unipotent of dimension 3 and $A(u)$ equal to Z4, the cyclic group of order 4. The class 22 has $C_G(u)$ with unipotent radical of dimension 4, reductive part of type A1 and $A(u)$ is Z2, the cyclic group of order 2. The other classes have $C_рқҗҶ(u)$ connected. For the class 31 the reductive part of $C_G(u)$ is a torus of rank 1.

Then there is a column for each Springer series, giving for each class the pairs 'a:b' where 'a' is the name of the character of $A(u)$ describing the local system involved and 'b' is the name of the character of the (relative) Weyl group corresponding by the Springer correspondence. At the top of the column is the name of the relative Weyl group, and in brackets the name of the Levi affording a cuspidal local system; then, separated by a /, is a description of the central character associated with the Springer series (omitted if this central character is trivial): all local systems in a given Springer series have the same restriction to the centre of $рқҗҶ$. To see what the picture becomes for another algebraic group in the same isogeny class, for example the adjoint group, one simply discards the Springer series whose central character becomes trivial on the centre of $рқҗҶ$; and each group $A(u)$ must be quotiented by the common kernel of the remaining characters. Here is the table for the adjoint group:

julia> UnipotentClasses(coxgroup(:A,3))
UnipotentClasses(AвӮғ)
1111<211<22<31<4
в”Ңв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯu   в”ӮD-R dв„¬ бөӨ B-C  C_рқҗҶ(u) AвӮғ(ОҰвӮҒВі)в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”Ӯ4   в”Ӯ222    0 222      qВі    Id:4в”Ӯ
в”Ӯ31  в”Ӯ202    1 22.   qвҒҙ.ОҰвӮҒ   Id:31в”Ӯ
в”Ӯ22  в”Ӯ020    2 2.2   qвҒҙ.AвӮҒ   Id:22в”Ӯ
в”Ӯ211 в”Ӯ101    3 2.. qвҒө.AвӮҒОҰвӮҒ  Id:211в”Ӯ
в”Ӯ1111в”Ӯ000    6 ...      AвӮғ Id:1111в”Ӯ
в””в”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

Here is another example:

julia> UnipotentClasses(coxgroup(:G,2))
UnipotentClasses(GвӮӮ)
1<AвӮҒ<AМғвӮҒ<GвӮӮ(aвӮҒ)<GвӮӮ
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯu     в”ӮD-R dв„¬ бөӨ B-C C_рқҗҶ(u)         GвӮӮ(ОҰвӮҒВІ)  .(GвӮӮ)в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮGвӮӮ    в”Ӯ 22    0  22     qВІ         Id:ПҶвӮҒвҖҡвӮҖ       в”Ӯ
в”ӮGвӮӮ(aвӮҒ)в”Ӯ 20    1  20  qвҒҙ.SвӮғ 21:ПҶвҖІвӮҒвҖҡвӮғ 3:ПҶвӮӮвҖҡвӮҒ 111:Idв”Ӯ
в”ӮAМғвӮҒ    в”Ӯ 01    2  .2  qВі.AвӮҒ         Id:ПҶвӮӮвҖҡвӮӮ       в”Ӯ
в”ӮAвӮҒ    в”Ӯ 10    3  2.  qвҒө.AвӮҒ        Id:ПҶвҖівӮҒвҖҡвӮғ       в”Ӯ
в”Ӯ1     в”Ӯ 00    6  ..     GвӮӮ         Id:ПҶвӮҒвҖҡвӮҶ       в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

which illustrates that on class GвӮӮ(aвӮҒ) there are two local systems in the principal series of the Springer correspondence, and another cuspidal local system. It also illustrates how we represent the Bala-Carter classification in general. If a class is attached to (рқҗӢ,рқҗҸ) then the simple roots in the complement of рқҗӢ have a .. Those in рқҗӢ have a 0 or a 2, the 2s characterising рқҗҸ. So, from the B-C column, we see that that GвӮӮ(aвӮҒ) is not in a proper Levi, in which case the Bala-Carter diagram coincides with the Dynkin-Richardson diagram.

Characteristics 2 and 3 are not good for G2. To get the unipotent classes and the Springer correspondence in a bad characteristic, give a second argument to the function UnipotentClasses:

julia> UnipotentClasses(coxgroup(:G,2),3)
UnipotentClasses(GвӮӮ,3)
1<AвӮҒ,(AМғвӮҒ)вӮғ<AМғвӮҒ<GвӮӮ(aвӮҒ)<GвӮӮ
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯu     в”Ӯdв„¬ бөӨ B-C C_рқҗҶ(u)  GвӮӮ(ОҰвӮҒВІ) .(GвӮӮ) .(GвӮӮ)  .(GвӮӮ)в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮGвӮӮ    в”Ӯ   0  22  qВІ.ZвӮғ   1:ПҶвӮҒвҖҡвӮҖ       О¶вӮғ:Id О¶вӮғВІ:Idв”Ӯ
в”ӮGвӮӮ(aвӮҒ)в”Ӯ   1  20  qвҒҙ.ZвӮӮ   2:ПҶвӮӮвҖҡвӮҒ 11:Id             в”Ӯ
в”ӮAМғвӮҒ    в”Ӯ   2  .2     qвҒ¶  Id:ПҶвӮӮвҖҡвӮӮ                   в”Ӯ
в”ӮAвӮҒ    в”Ӯ   3  2.  qвҒө.AвӮҒ Id:ПҶвҖівӮҒвҖҡвӮғ                   в”Ӯ
в”Ӯ(AМғвӮҒ)вӮғ в”Ӯ   3  ??  qвҒө.AвӮҒ Id:ПҶвҖІвӮҒвҖҡвӮғ                   в”Ӯ
в”Ӯ1     в”Ӯ   6  ..     GвӮӮ  Id:ПҶвӮҒвҖҡвӮҶ                   в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

The function ICCTable gives the transition matrix between the functions $XбөӘ$ and $Y_ПҲ$.

julia> uc=UnipotentClasses(coxgroup(:G,2));
julia> t=ICCTable(uc;q=Pol(:q))
Coefficients of XбөӘ on Yбө© for series L=GвӮӮвӮҚвӮҺ=ОҰвӮҒВІ W_G(L)=GвӮӮ
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮX\Y   в”ӮGвӮӮ GвӮӮ(aвӮҒ)вҒҪВІВ№вҒҫ GвӮӮ(aвӮҒ) AМғвӮҒ AвӮҒ  1в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮXПҶвӮҒвҖҡвӮҖ в”Ӯ 1          .      1  1  1  1в”Ӯ
в”ӮXПҶвҖІвӮҒвҖҡвӮғв”Ӯ .          1      .  1  . qВІв”Ӯ
в”ӮXПҶвӮӮвҖҡвӮҒ в”Ӯ .          .      1  1  1 ОҰвӮҲв”Ӯ
в”ӮXПҶвӮӮвҖҡвӮӮ в”Ӯ .          .      .  1  1 ОҰвӮ„в”Ӯ
в”ӮXПҶвҖівӮҒвҖҡвӮғв”Ӯ .          .      .  .  1  1в”Ӯ
в”ӮXПҶвӮҒвҖҡвӮҶ в”Ӯ .          .      .  .  .  1в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

An example which illustrates how to get the shoji names of classes

julia> uc=UnipotentClasses(coxgroup(:F,4));

julia> uc.classes[10:end]
7-element Vector{UnipotentClass}:
 UnipotentClass(CвӮғ(aвӮҒ))
 UnipotentClass(FвӮ„(aвӮғ))
 UnipotentClass(CвӮғ)
 UnipotentClass(BвӮғ)
 UnipotentClass(FвӮ„(aвӮӮ))
 UnipotentClass(FвӮ„(aвӮҒ))
 UnipotentClass(FвӮ„)

julia> xdisplay(uc.classes[10:end],shoji=true)
7-element Vector{UnipotentClass}:
 UnipotentClass(AвӮҒ+BвӮӮ)
 UnipotentClass(AвӮғ+AМғвӮҒ)
 UnipotentClass(CвӮғ)
 UnipotentClass(BвӮғ)
 UnipotentClass(CвӮғ+AвӮҒ)
 UnipotentClass(BвӮ„)
 UnipotentClass(FвӮ„)

In the ICCTable the row labels and column labels show the two ways of indexing local systems: the row labels give the character of the relative Weyl group and the column labels give the class and the name of the local system as a character of A(u): for example, G2(a1) is the trivial local system of the class G2(a1), while G2(a1)(21) is the local system on this class corresponding to the 2-dimensional character of $A(u)=AвӮӮ$.

The data on unipotent classes for arbitrary reductive groups are obtained as follows. We may reduce to the case where the radical is trivial, that is рқҗҶ is semisimple. Then рқҗҶ is the quotient of a simply connected group by a finite central group. The data for quasi-simple simply connected groups have been entered by hand for each type and a general simply connected group is a direct product of quasi-simple ones. In such a group, a character of A(Z), the group of components of the centre, is attached to each Springer series. For an isogenous reductive group рқҗҶ' with centre Z' the group A(Z') is a quotient of the group A(Z). The Springer series for рқҗҶ' are those such that the corresponding character of A(Z) factors through A(Z') (to compute A(Z') see algebraic_center). The geometric unipotent classes of рқҗҶ and рқҗҶ' are in bijection. For u a unipotent element of рқҗҶ' (which we can also consider as a unipotent element of рқҗҶ) the group AвӮҒ=A(u) in рқҗҶ' is a quotient of A=A(u) in рқҗҶ which we can compute as follows: the Springer correspondence for рқҗҶ' tells us which characters of A survive in рқҗҶ'. Then A' is the quotient of A by the common kernel of these characters.

Chevie.Ucl.UnipotentClasses вҖ” Type

UnipotentClasses(W[,p])

W should be a FiniteCoxeterGroup record for a Weyl group or rootdatum describing a reductive algebraic group рқҗҶ. The function returns a record containing information about the unipotent classes of рқҗҶ in characteristic p (if omitted, p is assumed to be any good characteristic for рқҗҶ). This contains the following fields:

group: a pointer to W

p: the characteristic of the field for which the unipotent classes were computed. It is 0 for any good characteristic.

orderclasses: a list describing the Hasse diagram of the partial order induced on unipotent classes by the closure relation. That is .orderclasses[i] is the list of j such that $CМ„вұјвҠӢ CМ„бөў$ and there is no class $CвӮ–$ such that $CМ„вұјвҠӢ CМ„вӮ–вҠӢ CМ„бөў$.

classes: a list of records holding information for each unipotent class. See the help for UnipotentClass for a description of these records.

springerseries: a list of records, each of which describes a Springer series of рқҗҶ.

The records describing individual Springer series have the following fields:

levi:the indices of the reflections corresponding to the Levi subgroup рқҗӢ where lives the cuspidal local system О№ from which the Springer series is induced.

relgroup: The relative Weyl group $N_рқҗҶ(рқҗӢ,О№)/рқҗӢ$. The first series is the principal series for which .levi=[] and .relgroup=W.

locsys: a list of length nconjugacy_classes(.relgroup), holding in i-th position a pair describing which local system corresponds to the i-th character of $N_рқҗҶ(рқҗӢ,О№)$. The first element of the pair is the index of the concerned unipotent class u, and the second is the index of the corresponding character of A(u).

Z: the central character associated to the Springer series, specified by its value on the generators of the center.

julia> W=rootdatum(:sl,4)
slвӮ„

julia> uc=UnipotentClasses(W);

julia> uc.classes
5-element Vector{UnipotentClass}:
 UnipotentClass(1111)
 UnipotentClass(211)
 UnipotentClass(22)
 UnipotentClass(31)
 UnipotentClass(4)

The show function for unipotent classes accepts all the options of showtable and of charnames. Giving the option mizuno (resp. shoji) uses the names given by (Mizuno, 1977; Mizuno, 1980) (resp. (Shoji, 1982)) for unipotent classes. Moreover, there is also an option fourier which gives the Springer correspondence tensored with the sign character of each relative Weyl group, which is the correspondence obtained via a Fourier-Deligne transform (here we assume that p is very good, so that there is a nondegenerate invariant bilinear form on the Lie algebra, and also one can identify nilpotent orbits with unipotent classes).

Here is how to display the non-cuspidal part of the Springer correspondence of the unipotent classes of EвӮҶ using the notations of (Mizuno, 1977) for the classes and those of (Frame, 1951) for the characters of the Weyl group and of (Spaltenstein, 1985) for the characters of GвӮӮ (this is convenient for checking our data with (Spaltenstein, 1985) which used that same conventions):

julia> uc=UnipotentClasses(rootdatum(:E6sc));

julia> xdisplay(uc;cols=[5,6,7],spaltenstein=true,frame=true,mizuno=true,
      order=false)
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯu     в”Ӯ             EвӮҶ(ОҰвӮҒвҒ¶) GвӮӮ(AвӮӮГ—AвӮӮОҰвӮҒВІ)/О¶вӮғ GвӮӮ(AвӮӮГ—AвӮӮОҰвӮҒВІ)/О¶вӮғВІв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮEвӮҶ    в”Ӯ                1:1вӮҡ            О¶вӮғ:1            О¶вӮғВІ:1в”Ӯ
в”ӮEвӮҶ(aвӮҒ)в”Ӯ                1:6вӮҡ           О¶вӮғ:ОөвӮ—           О¶вӮғВІ:ОөвӮ—в”Ӯ
в”ӮDвӮ…    в”Ӯ              Id:20вӮҡ                                 в”Ӯ
в”ӮAвӮ…+AвӮҒ в”Ӯ        -1:15вӮҡ 1:30вӮҡ           О¶вӮғ:ОёвҖІ           О¶вӮғВІ:ОёвҖІв”Ӯ
в”ӮAвӮ…    в”Ӯ              1:15_q           О¶вӮғ:ОёвҖі           О¶вӮғВІ:ОёвҖів”Ӯ
в”ӮDвӮ…(aвӮҒ)в”Ӯ              Id:64вӮҡ                                 в”Ӯ
в”ӮAвӮ„+AвӮҒ в”Ӯ              Id:60вӮҡ                                 в”Ӯ
в”ӮDвӮ„    в”Ӯ              Id:24вӮҡ                                 в”Ӯ
в”ӮAвӮ„    в”Ӯ              Id:81вӮҡ                                 в”Ӯ
в”ӮDвӮ„(aвӮҒ)в”Ӯ111:20вӮӣ 3:80вӮӣ 21:90вӮӣ                                 в”Ӯ
в”ӮAвӮғ+AвӮҒ в”Ӯ              Id:60вӮӣ                                 в”Ӯ
в”Ӯ2AвӮӮ+AвӮҒв”Ӯ               1:10вӮӣ          О¶вӮғ:Оө_c          О¶вӮғВІ:Оө_cв”Ӯ
в”ӮAвӮғ    в”Ӯ             Id:81вӮҡвҖІ                                 в”Ӯ
в”ӮAвӮӮ+2AвӮҒв”Ӯ             Id:60вӮҡвҖІ                                 в”Ӯ
в”Ӯ2AвӮӮ   в”Ӯ              1:24вӮҡвҖІ            О¶вӮғ:Оө            О¶вӮғВІ:Оөв”Ӯ
в”ӮAвӮӮ+AвӮҒ в”Ӯ             Id:64вӮҡвҖІ                                 в”Ӯ
в”ӮAвӮӮ    в”Ӯ      11:15вӮҡвҖІ 2:30вӮҡвҖІ                                 в”Ӯ
в”Ӯ3AвӮҒ   в”Ӯ            Id:15_qвҖІ                                 в”Ӯ
в”Ӯ2AвӮҒ   в”Ӯ             Id:20вӮҡвҖІ                                 в”Ӯ
в”ӮAвӮҒ    в”Ӯ              Id:6вӮҡвҖІ                                 в”Ӯ
в”Ӯ1     в”Ӯ              Id:1вӮҡвҖІ                                 в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

Chevie.Ucl.UnipotentClass вҖ” Type

A struct UnipotentClass representing the class C of a unipotent element u of the reductive group рқҗҶ with Weyl group W, contains always the following information

.name The name of C

.parameter A parameter describing C. Sometimes the same as .name; a partition describing the Jordan form, for classical groups.

.dimBu The dimension of the variety of Borel subgroups containing u.

For some classes in types EвӮҶ, EвӮҮ, EвӮҲ there is a field .mizuno containing the names given by Mizuno for some classes, and for some classes in type FвӮ„ a field .shoji containing the names given by Shoji.

A UnipotentClass contains also some of the following information (all of it for some types and some characteristics but sometimes much less)

.dynkin the Dynkin-Richardson diagram of C (a vector giving a weight 0, 1 or 2 to the simple roots).

.dimred the dimension of the reductive part of C_G(u).

.red a CoxeterCoset recording the type of the reductive part of C_G(u), with the twisting induced by the Frobenius if any.

.Au the group A_G(u)=C_G(u)/C^0_G(u).

.balacarter encodes the Bala-Carter classification of C, which says that u is distinguished in a Levi L (the Richardson class in a parabolic P_L) as a vector listing the indices of the simple roots in L, with those not in P_L negated.

.rep a list of indices for roots such that if U=UnipotentGroup(W) then prod(U,u.rep) is a representative of C (which can be obtained also by representative(W,u)).

.dimunip the dimension of the unipotent part of C_G(u).

.AuAction an ExtendedCoxeterGroup recording the action of A_G(u) on red.

Chevie.Ucl.ICCTable вҖ” Type

ICCTable(uc,seriesNo=1;q=Pol())

This function gives the table of decompositions of the functions $X_О№$ in terms of the functions $Y_О№$. Here О№ is a рқҗҶ-equivariant local system on the class C of a unipotent element u. Such a local system is parametrised by the pair (u,П•) of u and a character of the group of components A(u) of $C_рқҗҶ (u)$. The function $Y_О№$ is the characteristic function of this local system and $X_О№$ is the characteristic function of the corresponding intersection cohomology complex on CМ„. The Springer correspondence says that the local systems can also be indexed by characters of a relative Weyl group. Since the coefficient of XбөӘ on Yбө© is 0 if ПҮ and ПҶ are not characters of the same relative Weyl group (are not in the same Springer series), the table is for one Springer series, specified by the argument 'seriesNo' (this defaults to 'seriesNo=1' which is the principal series). The decomposition multiplicities are graded, and are given as polynomials in one variable (specified by the argument q; if not given Pol() is assumed).

julia> uc=UnipotentClasses(coxgroup(:A,3));t=ICCTable(uc)
Coefficients of XбөӘ on Yбө© for series L=AвӮғвӮҚвӮҺ=ОҰвӮҒВі W_G(L)=AвӮғ
в”Ңв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮX\Y  в”Ӯ4 31 22 211 1111в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮX4   в”Ӯ1  1  1   1    1в”Ӯ
в”ӮX31  в”Ӯ.  1  1  ОҰвӮӮ   ОҰвӮғв”Ӯ
в”ӮX22  в”Ӯ.  .  1   1   ОҰвӮ„в”Ӯ
в”ӮX211 в”Ӯ.  .  .   1   ОҰвӮғв”Ӯ
в”ӮX1111в”Ӯ.  .  .   .    1в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

In the above the multiplicities are given as products of cyclotomic polynomials to display them more compactly. However the format of such a table can be controlled more precisely.

For instance, one can ask to not display the entries as products of cyclotomic polynomials and not display the zeroes as '.'

julia> xdisplay(t;cycpol=false,dotzero=false)
Coefficients of XбөӘ on Yбө© for A3
в”Ңв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮX\Y  в”Ӯ4 31 22 211   1111в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮX4   в”Ӯ1  1  1   1      1в”Ӯ
в”ӮX31  в”Ӯ0  1  1 q+1 qВІ+q+1в”Ӯ
в”ӮX22  в”Ӯ0  0  1   1   qВІ+1в”Ӯ
в”ӮX211 в”Ӯ0  0  0   1 qВІ+q+1в”Ӯ
в”ӮX1111в”Ӯ0  0  0   0      1в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

Since show uses the function showtable, all the options of this function are also available. We can use this to restrict the entries displayed to a given sublist of the rows and columns (here the indices correspond to the number in Chevie of the corresponding character of the relative Weyl group of the given Springer series):

julia> uc=UnipotentClasses(coxgroup(:F,4));
julia> t=ICCTable(uc);
julia> sh=[13,24,22,18,14,9,11,19];
julia> xdisplay(t,rows=sh,cols=sh)
Coefficients of XбөӘ on Yбө© for series L=FвӮ„вӮҚвӮҺ=ОҰвӮҒвҒҙ W_G(L)=FвӮ„
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮX\Y    в”ӮAвӮҒ+AМғвӮҒ AвӮӮ AМғвӮӮ AвӮӮ+AМғвӮҒ AМғвӮӮ+AвӮҒ BвӮӮвҒҪВ№В№вҒҫ BвӮӮ CвӮғ(aвӮҒ)вҒҪВ№В№вҒҫв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮXПҶвӮүвҖҡвӮҒвӮҖ в”Ӯ    1  .  .     .     .      .  .          .в”Ӯ
в”ӮXПҶвҖівӮҲвҖҡвӮү в”Ӯ    1  1  .     .     .      .  .          .в”Ӯ
в”ӮXПҶвҖІвӮҲвҖҡвӮү в”Ӯ    1  .  1     .     .      .  .          .в”Ӯ
в”ӮXПҶвҖівӮ„вҖҡвӮҮ в”Ӯ    1  1  .     1     .      .  .          .в”Ӯ
в”ӮXПҶвҖІвӮҶвҖҡвӮҶ в”Ӯ   ОҰвӮ„  1  1     1     1      .  .          .в”Ӯ
в”ӮXПҶвӮ„вҖҡвӮҲ  в”Ӯ   qВІ  .  .     .     .      1  .          .в”Ӯ
в”ӮXПҶвҖівӮүвҖҡвӮҶ в”Ӯ   ОҰвӮ„ ОҰвӮ„  .     1     .      .  1          .в”Ӯ
в”ӮXПҶвҖІвӮ„вҖҡвӮҮ в”Ӯ   qВІ  . ОҰвӮ„     .     1      .  .          1в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

The Г¬o option rowlocsys=true will display local systems also for the row labels.

The function 'ICCTable' returns an object with various pieces of information which can help further computations.

.scalar: this contains the table of multiplicities PбөӘбө© of the XбөӘ on the Yбө©. One should pay attention that by default, the table is not displayed in the same order as the stored |.scalar|, which is in order in Chevie of the characters in the relative Weyl group; the table is transposed, then lines and rows are sorted by dimBu,class no,index of character in A(u) while displayed.

.group: The group W.

.relgroup: The relative Weyl group for the Springer series.

.series: The index of the Springer series given for W.

.dimBu: The list of $dimв„¬бөӨ$ for each local system (u,ПҶ) in the series.

:L: The matrix of (unnormalised) scalar products of the functions $Yбө©$ with themselves, that is the $(ПҶ,ПҮ)$ entry is $вҲ‘_{gвҲҲрқҗҶ(рқ”Ҫ_q)} Yбө©(g) YбөӘ(g)$. This is thus a symmetric, block-diagonal matrix where the diagonal blocks correspond to geometric unipotent conjugacy classes. This matrix is obtained as a by-product of Lusztig's algorithm to compute $Pбө©бөӘ$.

Chevie.Ucl.XTable вҖ” Type

XTable(uc;classes=false)

This function presents in a different way the information obtained from ICCTable. Let $XМғ_{u,П•}=q^{1/2(codim C-dim Z(рқҗӢ ))}X_{u,П•}$ where C is the class of u and Z(рқҗӢ) is the center of Levi subgroup on which lives the cuspidal local system attached to the local system (u,П•).

Then XTable(uc) gives the decomposition of the functions $XМғ_{u,П•}$ on local systems. t=XTable(uc,classes==true) gives the values of the functions $XМғ_{u,П•}$ on unipotent classes. A side effect of classes=true is to compute the cardinal of the unipotent conjugacy classes, available in t.cardClass; in this case displaying t will show the cardinal of the centralizers of unipotent elements, available in t.centClass.

julia> W=coxgroup(:G,2)
GвӮӮ

julia> XTable(UnipotentClasses(W))
Values of character sheaves XМғбөӘ of GвӮӮ on local systems ПҶ
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮXМғбөӘ|ПҶ      в”Ӯ   1 AвӮҒ AМғвӮҒ GвӮӮ(aвӮҒ)вҒҪВ№В№В№вҒҫ GвӮӮ(aвӮҒ)вҒҪВІВ№вҒҫ GвӮӮ(aвӮҒ) GвӮӮв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮX_ПҶвӮҒвҖҡвӮҖ^GвӮӮ в”Ӯ   1  1  1           .          .      1  1в”Ӯ
в”ӮX_ПҶвӮҒвҖҡвӮҶ^GвӮӮ в”Ӯ  qвҒ¶  .  .           .          .      .  .в”Ӯ
в”ӮX_ПҶвҖІвӮҒвҖҡвӮғ^GвӮӮв”Ӯ  qВі  .  q           .          q      .  .в”Ӯ
в”ӮX_ПҶвҖівӮҒвҖҡвӮғ^GвӮӮв”Ӯ  qВі qВі  .           .          .      .  .в”Ӯ
в”ӮX_ПҶвӮӮвҖҡвӮҒ^GвӮӮ в”Ӯ qОҰвӮҲ  q  q           .          .      q  .в”Ӯ
в”ӮX_ПҶвӮӮвҖҡвӮӮ^GвӮӮ в”ӮqВІОҰвӮ„ qВІ qВІ           .          .      .  .в”Ӯ
в”ӮX_Id^.    в”Ӯ   .  .  .          qВІ          .      .  .в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

The functions XМғ in the first column are decorated by putting as an exponent the relative groups $W_рқҗҶ (рқҗӢ)$.

julia> t=XTable(UnipotentClasses(W);classes=true)
Values of character sheaves XМғбөӘ of GвӮӮ on unipotent classes
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮXМғбөӘ|class  в”Ӯ           1     AвӮҒ     AМғвӮҒ GвӮӮ(aвӮҒ) GвӮӮ(aвӮҒ)вӮҚвӮӮвӮҒвӮҺ GвӮӮ(aвӮҒ)вӮҚвӮғвӮҺ GвӮӮв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮX_ПҶвӮҒвҖҡвӮҖ^GвӮӮ в”Ӯ           1      1      1      1          1         1  1в”Ӯ
в”ӮX_ПҶвӮҒвҖҡвӮҶ^GвӮӮ в”Ӯ          qвҒ¶      .      .      .          .         .  .в”Ӯ
в”ӮX_ПҶвҖІвӮҒвҖҡвӮғ^GвӮӮв”Ӯ          qВі      .      q     2q          .        -q  .в”Ӯ
в”ӮX_ПҶвҖівӮҒвҖҡвӮғ^GвӮӮв”Ӯ          qВі     qВі      .      .          .         .  .в”Ӯ
в”ӮX_ПҶвӮӮвҖҡвӮҒ^GвӮӮ в”Ӯ         qОҰвӮҲ      q      q      q          q         q  .в”Ӯ
в”ӮX_ПҶвӮӮвҖҡвӮӮ^GвӮӮ в”Ӯ        qВІОҰвӮ„     qВІ     qВІ      .          .         .  .в”Ӯ
в”ӮX_Id^.    в”Ӯ           .      .      .     qВІ        -qВІ        qВІ  .в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”Ӯ|C_рқҗҶ(u)|  в”ӮqвҒ¶ОҰвӮҒВІОҰвӮӮВІОҰвӮғОҰвӮҶ qвҒ¶ОҰвӮҒОҰвӮӮ qвҒҙОҰвӮҒОҰвӮӮ    6qвҒҙ        2qвҒҙ       3qвҒҙ qВІв”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

julia> XTable(UnipotentClasses(W,2))
Values of character sheaves XМғбөӘ of GвӮӮ on local systems ПҶ
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮXМғбөӘ|ПҶ      в”Ӯ   1 AвӮҒ AМғвӮҒ GвӮӮ(aвӮҒ)вҒҪВ№В№В№вҒҫ GвӮӮ(aвӮҒ)вҒҪВІВ№вҒҫ GвӮӮ(aвӮҒ) GвӮӮвҒҪВ№В№вҒҫ GвӮӮв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮX_ПҶвӮҒвҖҡвӮҖ^GвӮӮ в”Ӯ   1  1  1           .          .      1      .  1в”Ӯ
в”ӮX_ПҶвӮҒвҖҡвӮҶ^GвӮӮ в”Ӯ  qвҒ¶  .  .           .          .      .      .  .в”Ӯ
в”ӮX_ПҶвҖІвӮҒвҖҡвӮғ^GвӮӮв”Ӯ  qВі  .  q           .          q      .      .  .в”Ӯ
в”ӮX_ПҶвҖівӮҒвҖҡвӮғ^GвӮӮв”Ӯ  qВі qВі  .           .          .      .      .  .в”Ӯ
в”ӮX_ПҶвӮӮвҖҡвӮҒ^GвӮӮ в”Ӯ qОҰвӮҲ  q  q           .          .      q      .  .в”Ӯ
в”ӮX_ПҶвӮӮвҖҡвӮӮ^GвӮӮ в”ӮqВІОҰвӮ„ qВІ qВІ           .          .      .      .  .в”Ӯ
в”ӮX_Id^.    в”Ӯ   .  .  .          qВІ          .      .      .  .в”Ӯ
в”ӮX_Id^.    в”Ӯ   .  .  .           .          .      .      q  .в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

julia> XTable(UnipotentClasses(rootdatum(:sl,4)))
Values of character sheaves XМғбөӘ of slвӮ„ on local systems ПҶ
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮXМғбөӘ|ПҶ    в”Ӯ1111 211 22вҒҪВ№В№вҒҫ 22 31 4 4^(О¶вӮ„) 4вҒҪвҒ»В№вҒҫ 4^(О¶вӮ„Ві)в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮXвӮҒвӮҒвӮҒвӮҒ^AвӮғв”Ӯ  qвҒ¶   .      .  .  . .      .     .       .в”Ӯ
в”ӮXвӮӮвӮҒвӮҒ^AвӮғ в”ӮqВіОҰвӮғ  qВі      .  .  . .      .     .       .в”Ӯ
в”ӮXвӮӮвӮӮ^AвӮғ  в”ӮqВІОҰвӮ„  qВІ      . qВІ  . .      .     .       .в”Ӯ
в”ӮXвӮғвӮҒ^AвӮғ  в”Ӯ qОҰвӮғ qОҰвӮӮ      .  q  q .      .     .       .в”Ӯ
в”ӮXвӮ„^AвӮғ   в”Ӯ   1   1      .  1  1 1      .     .       .в”Ӯ
в”ӮXвӮҒвӮҒ^AвӮҒ  в”Ӯ   .   .     qВі  .  . .      .     .       .в”Ӯ
в”ӮXвӮӮ^AвӮҒ   в”Ӯ   .   .     qВІ  .  . .      .     q       .в”Ӯ
в”ӮX_Id^.  в”Ӯ   .   .      .  .  . .   qВівҒ„вӮӮ     .       .в”Ӯ
в”ӮX_Id^.  в”Ӯ   .   .      .  .  . .      .     .    qВівҒ„вӮӮв”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

Chevie.Ucl.GreenTable вҖ” Type

GreenTable(uc;classes=false)

Keeping the same notations as in the description of XTable, this function returns a table of the functions $Q_{wF}$, attached to elements $wFвҲҲ W_рқҗҶ (рқҗӢ)вӢ…F$ where $W_рқҗҶ (рқҗӢ)$ are the relative weyl groups attached to cuspidal local systems. These functions are defined by $Q_{wF}=вҲ‘_{u,П•} П•Мғ(wF) XМғ_{u,П•}$. Note that for the principal Springer series, when рқҗӢ=рқҗ“ is a maximal torus, the function $Q_{wF}$ coincides with the unipotent values of the Deligne-Lusztig character $R^рқҗҶ_{рқҗ“_W}(1)$. As for XTable, by default values are given in terms of local systems. If classes=true is given, then the values of the $Q_{wF}$ on conjugacy classes are given.

julia> W=coxgroup(:G,2)
GвӮӮ

julia> GreenTable(UnipotentClasses(W))
Values of Green functions Q_wF of GвӮӮ on local systems ПҶ
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮQбҙө_wF|ПҶ   в”Ӯ        1     AвӮҒ       AМғвӮҒ GвӮӮ(aвӮҒ)вҒҪВ№В№В№вҒҫ GвӮӮ(aвӮҒ)вҒҪВІВ№вҒҫ GвӮӮ(aвӮҒ) GвӮӮв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮQ_AвӮҖ^GвӮӮ   в”Ӯ  ОҰвӮӮВІОҰвӮғОҰвӮҶ   ОҰвӮӮОҰвӮғ (2q+1)ОҰвӮӮ           .          q   2q+1  1в”Ӯ
в”ӮQ_AМғвӮҒ^GвӮӮ   в”Ӯ-ОҰвӮҒОҰвӮӮОҰвӮғОҰвӮҶ  -ОҰвӮҒОҰвӮғ       ОҰвӮӮ           .          q      1  1в”Ӯ
в”ӮQ_AвӮҒ^GвӮӮ   в”Ӯ-ОҰвӮҒОҰвӮӮОҰвӮғОҰвӮҶ   ОҰвӮӮОҰвӮҶ      -ОҰвӮҒ           .         -q      1  1в”Ӯ
в”ӮQ_GвӮӮ^GвӮӮ   в”Ӯ ОҰвӮҒВІОҰвӮӮВІОҰвӮғ -ОҰвӮҒОҰвӮӮВІ    -ОҰвӮҒОҰвӮӮ           .         -q     ОҰвӮӮ  1в”Ӯ
в”ӮQ_AвӮӮ^GвӮӮ   в”Ӯ ОҰвӮҒВІОҰвӮӮВІОҰвӮҶ  ОҰвӮҒВІОҰвӮӮ    -ОҰвӮҒОҰвӮӮ           .          q    -ОҰвӮҒ  1в”Ӯ
в”ӮQ_AвӮҒ+AМғвӮҒ^GвӮӮв”Ӯ  ОҰвӮҒВІОҰвӮғОҰвӮҶ  -ОҰвӮҒОҰвӮҶ (2q-1)ОҰвӮҒ           .         -q  -2q+1  1в”Ӯ
в”ӮQ^.       в”Ӯ        .      .        .          qВІ          .      .  .в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

The functions $Q_{wF}$ depend only on the conjugacy class of wF, so in the first column the indices of 'Q' are the names of the conjugacy classes of $W_рқҗҶ(рқҗӢ)$. The exponents are the names of the groups $W_рқҗҶ(рқҗӢ)$.

julia> GreenTable(UnipotentClasses(W);classes=true)
Values of Green functions Q_wF of GвӮӮ on unipotent classes
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮQбҙө_wF|classв”Ӯ        1     AвӮҒ       AМғвӮҒ GвӮӮ(aвӮҒ) GвӮӮ(aвӮҒ)вӮҚвӮӮвӮҒвӮҺ GвӮӮ(aвӮҒ)вӮҚвӮғвӮҺ GвӮӮв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮQ_AвӮҖ^GвӮӮ    в”Ӯ  ОҰвӮӮВІОҰвӮғОҰвӮҶ   ОҰвӮӮОҰвӮғ (2q+1)ОҰвӮӮ   4q+1       2q+1        ОҰвӮӮ  1в”Ӯ
в”ӮQ_AМғвӮҒ^GвӮӮ    в”Ӯ-ОҰвӮҒОҰвӮӮОҰвӮғОҰвӮҶ  -ОҰвӮҒОҰвӮғ       ОҰвӮӮ   2q+1          1       -ОҰвӮҒ  1в”Ӯ
в”ӮQ_AвӮҒ^GвӮӮ    в”Ӯ-ОҰвӮҒОҰвӮӮОҰвӮғОҰвӮҶ   ОҰвӮӮОҰвӮҶ      -ОҰвӮҒ  -2q+1          1        ОҰвӮӮ  1в”Ӯ
в”ӮQ_GвӮӮ^GвӮӮ    в”Ӯ ОҰвӮҒВІОҰвӮӮВІОҰвӮғ -ОҰвӮҒОҰвӮӮВІ    -ОҰвӮҒОҰвӮӮ    -ОҰвӮҒ         ОҰвӮӮ      2q+1  1в”Ӯ
в”ӮQ_AвӮӮ^GвӮӮ    в”Ӯ ОҰвӮҒВІОҰвӮӮВІОҰвӮҶ  ОҰвӮҒВІОҰвӮӮ    -ОҰвӮҒОҰвӮӮ     ОҰвӮӮ        -ОҰвӮҒ     -2q+1  1в”Ӯ
в”ӮQ_AвӮҒ+AМғвӮҒ^GвӮӮ в”Ӯ  ОҰвӮҒВІОҰвӮғОҰвӮҶ  -ОҰвӮҒОҰвӮҶ (2q-1)ОҰвӮҒ  -4q+1      -2q+1       -ОҰвӮҒ  1в”Ӯ
в”ӮQ^.        в”Ӯ        .      .        .     qВІ        -qВІ        qВІ  .в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

julia> GreenTable(UnipotentClasses(rootdatum(:sl,4)))
Values of Green functions Q_wF of slвӮ„ on local systems ПҶ
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”ӮQбҙө_wF|ПҶ в”Ӯ     1111          211 22вҒҪВ№В№вҒҫ       22   31 4 4^(О¶вӮ„) 4вҒҪвҒ»В№вҒҫ 4^(О¶вӮ„Ві)в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮQвӮҒвӮҒвӮҒвӮҒ^AвӮғв”Ӯ  ОҰвӮӮВІОҰвӮғОҰвӮ„ (3qВІ+2q+1)ОҰвӮӮ      . (2q+1)ОҰвӮӮ 3q+1 1      .     .       .в”Ӯ
в”ӮQвӮӮвӮҒвӮҒ^AвӮғ в”Ӯ-ОҰвӮҒОҰвӮӮОҰвӮғОҰвӮ„   -qВі+qВІ+q+1      .       ОҰвӮӮ   ОҰвӮӮ 1      .     .       .в”Ӯ
в”ӮQвӮӮвӮӮ^AвӮғ  в”Ӯ  ОҰвӮҒВІОҰвӮғОҰвӮ„        -ОҰвӮҒОҰвӮ„      .  2qВІ-q+1  -ОҰвӮҒ 1      .     .       .в”Ӯ
в”ӮQвӮғвӮҒ^AвӮғ  в”Ӯ ОҰвӮҒВІОҰвӮӮВІОҰвӮ„        -ОҰвӮҒОҰвӮӮ      .    -ОҰвӮҒОҰвӮӮ    1 1      .     .       .в”Ӯ
в”ӮQвӮ„^AвӮғ   в”Ӯ -ОҰвӮҒВіОҰвӮӮОҰвӮғ        ОҰвӮҒВІОҰвӮӮ      .      -ОҰвӮҒ  -ОҰвӮҒ 1      .     .       .в”Ӯ
в”ӮQвӮҒвӮҒ^AвӮҒ  в”Ӯ        .            .   qВІОҰвӮӮ        .    . .      .     q       .в”Ӯ
в”ӮQвӮӮ^AвӮҒ   в”Ӯ        .            .  -qВІОҰвӮҒ        .    . .      .     q       .в”Ӯ
в”ӮQ^.     в”Ӯ        .            .      .        .    . .   qВівҒ„вӮӮ     .       .в”Ӯ
в”ӮQ^.     в”Ӯ        .            .      .        .    . .      .     .    qВівҒ„вӮӮв”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

Chevie.Ucl.UnipotentValues вҖ” Function

UnipotentValues(uc,classes=false)

This function returns a table of the values of unipotent characters on local systems (by default) or on unipotent classes (if classes=true).

julia> W=coxgroup(:G,2)
GвӮӮ

julia> UnipotentValues(UnipotentClasses(W);classes=true)
Values of unipotent characters for GвӮӮ on unipotent classes
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯ       в”Ӯ        1          AвӮҒ     AМғвӮҒ   GвӮӮ(aвӮҒ) GвӮӮ(aвӮҒ)вӮҚвӮӮвӮҒвӮҺ GвӮӮ(aвӮҒ)вӮҚвӮғвӮҺ GвӮӮв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮПҶвӮҒвҖҡвӮҖ   в”Ӯ        1           1      1        1          1         1  1в”Ӯ
в”ӮПҶвӮҒвҖҡвӮҶ   в”Ӯ       qвҒ¶           .      .        .          .         .  .в”Ӯ
в”ӮПҶвҖІвӮҒвҖҡвӮғ  в”Ӯ  qОҰвӮғОҰвӮҶ/3    -qОҰвӮҒОҰвӮӮ/3      q (q+5)q/3     -qОҰвӮҒ/3     qОҰвӮҒ/3  .в”Ӯ
в”ӮПҶвҖівӮҒвҖҡвӮғ  в”Ӯ  qОҰвӮғОҰвӮҶ/3  (2qВІ+1)q/3      .    qОҰвӮҒ/3     -qОҰвӮҒ/3  (q+2)q/3  .в”Ӯ
в”ӮПҶвӮӮвҖҡвӮҒ   в”Ӯ qОҰвӮӮВІОҰвӮғ/6 (2q+1)qОҰвӮӮ/6  qОҰвӮӮ/2 (q+5)q/6     -qОҰвӮҒ/6     qОҰвӮҒ/6  .в”Ӯ
в”ӮПҶвӮӮвҖҡвӮӮ   в”Ӯ qОҰвӮӮВІОҰвӮҶ/2       qОҰвӮӮ/2  qОҰвӮӮ/2   -qОҰвӮҒ/2      qОҰвӮӮ/2    -qОҰвӮҒ/2  .в”Ӯ
в”ӮGвӮӮ[-1] в”Ӯ qОҰвӮҒВІОҰвӮғ/2      -qОҰвӮҒ/2 -qОҰвӮҒ/2   -qОҰвӮҒ/2      qОҰвӮӮ/2    -qОҰвӮҒ/2  .в”Ӯ
в”ӮGвӮӮ[1]  в”Ӯ qОҰвӮҒВІОҰвӮҶ/6 (2q-1)qОҰвӮҒ/6 -qОҰвӮҒ/2 (q+5)q/6     -qОҰвӮҒ/6     qОҰвӮҒ/6  .в”Ӯ
в”ӮGвӮӮ[О¶вӮғ] в”ӮqОҰвӮҒВІОҰвӮӮВІ/3    -qОҰвӮҒОҰвӮӮ/3      .    qОҰвӮҒ/3     -qОҰвӮҒ/3  (q+2)q/3  .в”Ӯ
в”ӮGвӮӮ[О¶вӮғВІ]в”ӮqОҰвӮҒВІОҰвӮӮВІ/3    -qОҰвӮҒОҰвӮӮ/3      .    qОҰвӮҒ/3     -qОҰвӮҒ/3  (q+2)q/3  .в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

julia> UnipotentValues(UnipotentClasses(W,3);classes=true)
Values of unipotent characters for GвӮӮ on unipotent classes
в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯ       в”Ӯ        1          AвӮҒ         AМғвӮҒ GвӮӮ(aвӮҒ) GвӮӮ(aвӮҒ)вӮҚвӮӮвӮҺ    GвӮӮ       GвӮӮ_(О¶вӮғ)в”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮПҶвӮҒвҖҡвӮҖ   в”Ӯ        1           1          1      1         1     1             1в”Ӯ
в”ӮПҶвӮҒвҖҡвӮҶ   в”Ӯ       qвҒ¶           .          .      .         .     .             .в”Ӯ
в”ӮПҶвҖІвӮҒвҖҡвӮғ  в”Ӯ  qОҰвӮғОҰвӮҶ/3    -qОҰвӮҒОҰвӮӮ/3        q/3  qОҰвӮӮ/3    -qОҰвӮҒ/3 -2q/3           q/3в”Ӯ
в”ӮПҶвҖівӮҒвҖҡвӮғ  в”Ӯ  qОҰвӮғОҰвӮҶ/3  (2qВІ+1)q/3        q/3  qОҰвӮӮ/3    -qОҰвӮҒ/3 -2q/3           q/3в”Ӯ
в”ӮПҶвӮӮвҖҡвӮҒ   в”Ӯ qОҰвӮӮВІОҰвӮғ/6 (2q+1)qОҰвӮӮ/6  (3q+1)q/6  qОҰвӮӮ/6    -qОҰвӮҒ/6  2q/3          -q/3в”Ӯ
в”ӮПҶвӮӮвҖҡвӮӮ   в”Ӯ qОҰвӮӮВІОҰвӮҶ/2       qОҰвӮӮ/2      qОҰвӮӮ/2 -qОҰвӮҒ/2     qОҰвӮӮ/2     .             .в”Ӯ
в”ӮGвӮӮ[-1] в”Ӯ qОҰвӮҒВІОҰвӮғ/2      -qОҰвӮҒ/2     -qОҰвӮҒ/2 -qОҰвӮҒ/2     qОҰвӮӮ/2     .             .в”Ӯ
в”ӮGвӮӮ[1]  в”Ӯ qОҰвӮҒВІОҰвӮҶ/6 (2q-1)qОҰвӮҒ/6 (-3q+1)q/6  qОҰвӮӮ/6    -qОҰвӮҒ/6  2q/3          -q/3в”Ӯ
в”ӮGвӮӮ[О¶вӮғ] в”ӮqОҰвӮҒВІОҰвӮӮВІ/3    -qОҰвӮҒОҰвӮӮ/3        q/3  qОҰвӮӮ/3    -qОҰвӮҒ/3   q/3 (-О¶вӮғ+2О¶вӮғВІ)q/3в”Ӯ
в”ӮGвӮӮ[О¶вӮғВІ]в”ӮqОҰвӮҒВІОҰвӮӮВІ/3    -qОҰвӮҒОҰвӮӮ/3        q/3  qОҰвӮӮ/3    -qОҰвӮҒ/3   q/3  (2О¶вӮғ-О¶вӮғВІ)q/3в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

в”Ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”¬в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”җ
в”Ӯ       в”Ӯ     GвӮӮ_(О¶вӮғВІ)       (AМғвӮҒ)вӮғв”Ӯ
в”ңв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”јв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Ө
в”ӮПҶвӮҒвҖҡвӮҖ   в”Ӯ            1           1в”Ӯ
в”ӮПҶвӮҒвҖҡвӮҶ   в”Ӯ            .           .в”Ӯ
в”ӮПҶвҖІвӮҒвҖҡвӮғ  в”Ӯ          q/3  (2qВІ+1)q/3в”Ӯ
в”ӮПҶвҖівӮҒвҖҡвӮғ  в”Ӯ          q/3    -qОҰвӮҒОҰвӮӮ/3в”Ӯ
в”ӮПҶвӮӮвҖҡвӮҒ   в”Ӯ         -q/3 (2q+1)qОҰвӮӮ/6в”Ӯ
в”ӮПҶвӮӮвҖҡвӮӮ   в”Ӯ            .       qОҰвӮӮ/2в”Ӯ
в”ӮGвӮӮ[-1] в”Ӯ            .      -qОҰвӮҒ/2в”Ӯ
в”ӮGвӮӮ[1]  в”Ӯ         -q/3 (2q-1)qОҰвӮҒ/6в”Ӯ
в”ӮGвӮӮ[О¶вӮғ] в”Ӯ (2О¶вӮғ-О¶вӮғВІ)q/3    -qОҰвӮҒОҰвӮӮ/3в”Ӯ
в”ӮGвӮӮ[О¶вӮғВІ]в”Ӯ(-О¶вӮғ+2О¶вӮғВІ)q/3    -qОҰвӮҒОҰвӮӮ/3в”Ӯ
в””в”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҙв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”Җв”ҳ

Chevie.Ucl.induced_linear_form вҖ” Function

induced_linear_form(W, H, h)

This routine can be used to find the unipotent class in the reductive group with Weyl group W which contains a given unipotent class of a reductive subgroup of maximum rank represented by the reflection subgroup H of W.

The argument h is a linear form on the roots of H, given by its value on the simple roots; this linear form is extended to the roots of W by 0 on the orthogonal of the roots of H; and finally the resulting form is conjugated by an element of W so that it takes positive values on the simple roots. If the initial form describes a Dynkin-Richardson diagram for H, the result will describe a Dynkin-Richardson diagram for W.

julia> W=coxgroup(:F,4)
FвӮ„

julia> H=reflection_subgroup(W,[1,3])
FвӮ„вӮҚвӮҒвӮғвӮҺ=AвӮҒГ—AМғвӮҒОҰвӮҒВІ

julia> induced_linear_form(W,H,[2,2])
4-element Vector{Int64}:
 0
 1
 0
 0

julia> uc=UnipotentClasses(W);

julia> uc.classes[4].dynkin
4-element Vector{Int64}:
 0
 1
 0
 0

julia> uc.classes[4]
UnipotentClass(AвӮҒ+AМғвӮҒ)

The example above shows that the class containing the regular class of the Levi subgroup of type AвӮҒГ—AМғвӮҒ is the class AвӮҒ+AМғвӮҒ.

Chevie.Ucl.special_pieces вҖ” Function

'special_pieces(<uc>)'

The special pieces form a partition of the unipotent variety of a reductive group рқҗҶ which was first defined in (Spaltenstein, 1982; chap. III) as the fibres of d^2, where d is a "duality map". Another definition is as the set of classes in the Zariski closure of a special class and not in the Zariski closure of any smaller special class, where a special class is a class such that the trivial local system on it corresponds to a special character by the Springer correspondence.

Since each Lusztig family of characters contains a unique speical character, special pieces are in bijection with lusztig families.

Each piece is a union of unipotent conjugacy classes so is represented in Chevie as a list of class numbers. Thus the list of special pieces is returned as a list of lists of class numbers. The list is sorted by increasing piece dimension, while each piece is sorted by decreasing class dimension, so that the special class is listed first.

julia> W=coxgroup(:G,2)
GвӮӮ

julia> special_pieces(UnipotentClasses(W))
3-element Vector{Vector{Int64}}:
 [1]
 [4, 3, 2]
 [5]

julia> special_pieces(UnipotentClasses(W,3))
3-element Vector{Vector{Int64}}:
 [1]
 [4, 3, 2, 6]
 [5]

The example above shows that the special pieces may be different in bad characteristic.

Chevie.Ucl.distinguished_parabolics вҖ” Function

distinguished_parabolics(W)

the list of distinguished (in the sense of Richardson) standard parabolic subgroups W_I of W, each given by the list I of indices in the simple reflections. The distinguished unipotent conjugacy classes of the reductive group рқҗҶ with Weyl group W consist of the dense unipotent orbit in the unipotent radical of a parabolic subgroup рқҗҸ_I of рқҗҶ associated with a distinguished W_I. Their Dynkin-Richardson diagram contains a 0 at indices I and a 2 in other entries. If рқҗҸ_I=рқҗӢрқҗ” is a Levi decomposition, рқҗҸ_I is distinguished iff dim рқҗӢвҲ©рқҗҶ'=dim рқҗ”/рқҗ”'.

julia> W=coxgroup(:F,4)
FвӮ„

julia> distinguished_parabolics(W)
4-element Vector{Vector{Int64}}:
 []
 [3]
 [1, 3]
 [1, 3, 4]